### 2026-03-13 #10178 @danielskatz
👋 @AlessandroPomponio - thanks for your submission. It looks mostly ready for JOSS review, with the exception that the paper is a bit long. JOSS recommends papers be between 750 and 1750 words, and yours is about 2400. Please try to remove some of the text, perhaps by pointing to content in the repo or docs instead. You can then use the command `@editorialbot generate pdf` to make a new PDF. You can also use the command `@editorialbot check repository` to check a bunch of things in the repository, including the paper's word count.

In addition, you could work on the possibly missing DOI that editorialbot suggests, but note that it may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again.

An editorialbot command needs to be the first entry in a new comment.
