### 2026-03-13 77ddc6a
docs(paper): update Kwon et al. reference
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
