+++ paper/paper.bib
@@ -123,11 +123,21 @@
   doi        = {10.1109/ICRA55743.2025.11128578}
 }
 
-@inproceedings{kwon2023efficient,
+@inproceedings{10.1145/3600006.3613165,
+  author    = {Kwon, Woosuk and Li, Zhuohan and Zhuang, Siyuan and Sheng, Ying and Zheng, Lianmin and Yu, Cody Hao and Gonzalez, Joseph and Zhang, Hao and Stoica, Ion},
   title     = {Efficient Memory Management for Large Language Model Serving with PagedAttention},
-  author    = {Woosuk Kwon and Zhuohan Li and Siyuan Zhuang and Ying Sheng and Lianmin Zheng and Cody Hao Yu and Joseph E. Gonzalez and Hao Zhang and Ion Stoica},
-  booktitle = {Proceedings of the ACM SIGOPS 29th Symposium on Operating Systems Principles},
-  year      = {2023}
+  year      = {2023},
+  isbn      = {9798400702297},
+  publisher = {Association for Computing Machinery},
+  address   = {New York, NY, USA},
+  url       = {https://doi.org/10.1145/3600006.3613165},
+  doi       = {10.1145/3600006.3613165},
+  abstract  = {High throughput serving of large language models (LLMs) requires batching sufficiently many requests at a time. However, existing systems struggle because the key-value cache (KV cache) memory for each request is huge and grows and shrinks dynamically. When managed inefficiently, this memory can be significantly wasted by fragmentation and redundant duplication, limiting the batch size. To address this problem, we propose PagedAttention, an attention algorithm inspired by the classical virtual memory and paging techniques in operating systems. On top of it, we build vLLM, an LLM serving system that achieves (1) near-zero waste in KV cache memory and (2) flexible sharing of KV cache within and across requests to further reduce memory usage. Our evaluations show that vLLM improves the throughput of popular LLMs by 2--4\texttimes{} with the same level of latency compared to the state-of-the-art systems, such as FasterTransformer and Orca. The improvement is more pronounced with longer sequences, larger models, and more complex decoding algorithms. vLLM's source code is publicly available at https://github.com/vllm-project/vllm.},
+  booktitle = {Proceedings of the 29th Symposium on Operating Systems Principles},
+  pages     = {611–626},
+  numpages  = {16},
+  location  = {Koblenz, Germany},
+  series    = {SOSP '23}
 }
 
 @software{pydantic,
