### 2026-03-13 f162f3a
docs(paper): compress
