+++ paper/paper.md
@@ -58,7 +58,7 @@ ado is open source and available at <https://github.com/ibm/ado>
 # Statement of need
 
 While the domains of computational science are diverse, spanning machine
-learning, physics simulation, and hardware design, the process of systematic
+learning, physics simulation, and hardware design, the process of
 experimentation is remarkably uniform. Whether tuning hyperparameters,
 benchmarking foundation models, or sweeping simulation parameters, researchers
 consistently follow a structured pattern: define a configuration space; select
@@ -71,69 +71,52 @@ Scientific and ML workflow systems like Galaxy, AiiDA, and Kubeflow excel at
 executing general directed acyclic graphs (DAGs). However, they are
 fundamentally context-free, treating each step as a black box
 [@10.1093/nar/gkae410; @Huber2020; @George2022EndtoendML]. When it comes to
-implementing experiment campaigns, this forces researchers to implement
-mechanisms for trial submission, parameter handling, logging, and result
-collation. This imperative approach leads to duplicated engineering effort,
-inconsistent practices, and slower scientific progress.
+experiment campaigns, this forces researchers to implement mechanisms for trial
+submission, parameter handling, and result collation.
 
 ado directly addresses this gap. Instead of orchestrating arbitrary DAGs, ado
-provides a semantic experimentation model centered on experiment campaigns.
-Users define configuration spaces and operations on them (e.g., sampling or
-analysis) declaratively. ado then applies the required orchestration using its
-own protocols. For example, in sampling workflows, it handles reuse of prior
-measurements, trial execution and monitoring, and time-resolved measurement
-recording, maintaining consistency over a shared sample store. This approach
-mirrors the advantages of declarative systems like SQL or Terraform: reduced
-boilerplate, fewer errors, and greater clarity. This structured, declarative
-representation of experimental campaigns also aids code generation tools in
-automatically producing experiment definitions and formulating design spaces.
+focuses on the generic experiment campaign process. Users define configuration
+spaces and operations on them declaratively. ado then orchestrates the
+operations using its own protocols. For example, in sampling workflows, it
+handles reuse of prior measurements, trial execution and monitoring, and
+time-resolved measurement recording, maintaining consistency over a shared
+sample store. This approach mirrors the advantages of declarative systems like
+SQL or Terraform: reduced boilerplate, fewer errors, and greater clarity. It
+also aids code generators in creating experiment definitions and design spaces.
 
-![A schematic overview of the ado architecture, illustrating its human-centered
-declarative interface, a scalable and extensible core, and a central database
-that enables collaboration, reuse, and
-data provenance.\label{fig:ado}](ADOSchematic.png)
+![A schematic overview ado's architecture. \label{fig:ado}](ADOSchematic.png)
 
 ado extends its core model with valuable support capabilities. Specifications
 (configuration spaces, operations) and measurements are stored in a database
-with flexible deployment options. A local database is available by default for
-standalone work, while a connection to a shared, central database can be
-configured to support team collaboration and transparent result reuse. Built on
-the Ray execution engine, ado seamlessly scales from a researcher's laptop to a
-large remote cluster [@Moritz2018], with all functionality accessible via a
-human-centric CLI and Python API. Researchers can contribute custom experiments
-or operators through a simple plugin interface. The result is a system that is
-context-specific yet domain-agnostic (see \autoref{fig:ado}).
+with flexible deployment options (local or remote shared). It leverages Ray to
+seamlessly scale from a researcher's laptop to a large remote cluster
+[@Moritz2018], with all functionality accessible via a human-centric CLI and
[…]
