### 2026-03-13 4feb994
docs(paper): further compression
