+++ paper/paper.md
@@ -46,11 +46,10 @@ model, where integrating new components can be as simple as decorating a Python
 function. This is enabled by ado's core abstraction: the _Discovery Space_.
 
 Out-of-the-box, ado includes state-of-the-art optimization algorithms and
-predictive modeling tools, alongside concrete experiments targeting
+predictive modeling tools, alongside experiments targeting
 foundation-model performance. Our aim is for ado to become a focal point for
 developing and consuming advanced capabilities for defining and executing
-experiment campaigns that accelerate computational research, from initial design
-to final analysis.
+experiment campaigns.
 
 ado is open source and available at <https://github.com/ibm/ado>
 [@johnston_2026_18957620].
@@ -62,8 +61,8 @@ learning, physics simulation, and hardware design, the process of
 experimentation is remarkably uniform. Whether tuning hyperparameters,
 benchmarking foundation models, or sweeping simulation parameters, researchers
 consistently follow a structured pattern: define a configuration space; select
-points within it; execute experiments at those points; record results; and
-analyze the outcomes to guide the next iteration. This recurring workflow, the
+points; execute experiments; record results; and
+analyze the outcomes to guide the next iteration. This workflow, the
 _experiment campaign_, is central to modern research and development, yet it is
 often managed with tools that fail to capture its essential structure.
 
@@ -126,7 +125,7 @@ fragmented approach. For automated lab systems, their focus is on managing
 operational complexity, not providing domain-agnostic, declarative campaign
 semantics above the lab layer.
 
-ado synergizes with, rather than replaces, these existing tools. It can use
+ado synergizes with, rather than replaces, these tools. It can use
 workflow managers as experiment executors, integrate optimization frameworks,
 and orchestrate physical experiments by coupling with robotic lab systems. At
 the same time, individual experiment implementations within ado's plugin
@@ -187,7 +186,7 @@ configuration probability space and action space definitions it is
 **encapsulated** and **actionable**; it is **time-resolved** as it contains the
 sample set.
 
-![A view of data sharing between Discovery Spaces. Each space reads/writes
+![Data sharing between Discovery Spaces. Each space reads/writes
 samples and time series
 membership details from a shared sample store (_common context_).
 In this case both spaces have the same action space and contain point X.
@@ -198,56 +197,44 @@ width=80% }
 
 We obtain a **common context** by storing the sample time series in a shared
 sample store with a common schema. Finally, it is **reconcilable** as it
-enforces that samples in the common context are only part of the sample set of a
-Discovery Space if they are part of the sample time series of an operation on
-that space (see \autoref{fig:ds_interaction}). Hence, it displays the TRACE
+enforces a strict membership rule for its sample set: only samples in the
+common-context that are associated with an operation conducted on that space are
+included. (see \autoref{fig:ds_interaction}). Hence, it displays the TRACE
 characteristics.
 
 The Discovery Space abstraction effectively decouples workload-specific
 experiments from the search and optimization algorithms, enabling the kind of
 versatile, workload-agnostic capabilities that are a key goal of the program.
 
-## Data Modeling with Python and Pydantic
+## Pydantic-based Core Architecture
 
-The core of our architecture is implemented in Python, chosen for its ubiquity
-in scientific and research domains and its extensive ecosystem. For all data
-modeling, we leverage the Pydantic framework [@pydantic]. Pydantic's
-declarative, type-hinted models provide automatic validation for all core
-abstractions ensuring data integrity, providing self-documenting schemas, and
-creating a clear target for AI code generation.
+ado is implemented in Python, chosen for its ubiquity in scientific domains.
+Central to ado is the Pydantic framework, which we use for all data modeling.
+These Pydantic models serve as the primary user-facing components, providing
[…]
