### 2026-03-13 8e99062
docs(paper): remove ray section
Its mentioned already.
