### 2026-03-13 1c03fb0
docs(paper): add missing doi
