+++ paper/paper.bib
@@ -6,6 +6,7 @@
   archiveprefix = {arXiv},
   primaryclass  = {cs.DC},
   url           = {https://arxiv.org/abs/2506.21467}
+  doi = {10.48550/arXiv.2506.21467}
 }
 
 @article{Liaw2018,
@@ -22,6 +23,8 @@
   booktitle = {13th USENIX Symposium on Operating Systems Design and Implementation (OSDI 18)},
   year      = {2018},
   pages     = {561--577}
+  doi       = {10.48550/arXiv.1712.05889}
+
 }
 
 @inproceedings{Akiba2019,
@@ -164,6 +167,7 @@
   journal = {Proceedings of the 5th Joint International Conference on Data Science \& Management of Data (9th ACM IKDD CODS and 27th COMAD)},
   year    = {2022},
   url     = {https://api.semanticscholar.org/CorpusID:245803553}
+  doi     = {10.1145/3493700.3493768}
 }
 
 @software{johnston_2026_18957620,
@@ -191,4 +195,5 @@
   title     = {The FAIR Guiding Principles for scientific data management and stewardship},
   volume    = 3,
   year      = 2016
+  doi       = {10.1038/sdata.2016.18}
 }
