### 2026-03-13 54c7aad
docs(paper): fix bib
missing commas
