+++ paper/paper.bib
@@ -5,8 +5,8 @@
   eprint        = {2506.21467},
   archiveprefix = {arXiv},
   primaryclass  = {cs.DC},
-  url           = {https://arxiv.org/abs/2506.21467}
-  doi = {10.48550/arXiv.2506.21467}
+  url           = {https://arxiv.org/abs/2506.21467},
+  doi           = {10.48550/arXiv.2506.21467}
 }
 
 @article{Liaw2018,
@@ -22,7 +22,7 @@
   author    = {Moritz, Philipp and Nishihara, Robert and Wang, Stephanie and Tumanov, Alexey and Liaw, Richard and Liang, Eric and Elibol, Melih and Yang, Zongheng and Paul, William and Jordan, Michael I. and Stoica, Ion},
   booktitle = {13th USENIX Symposium on Operating Systems Design and Implementation (OSDI 18)},
   year      = {2018},
-  pages     = {561--577}
+  pages     = {561--577},
   doi       = {10.48550/arXiv.1712.05889}
 
 }
@@ -166,7 +166,7 @@
   author  = {Johnu George and Amit Saha},
   journal = {Proceedings of the 5th Joint International Conference on Data Science \& Management of Data (9th ACM IKDD CODS and 27th COMAD)},
   year    = {2022},
-  url     = {https://api.semanticscholar.org/CorpusID:245803553}
+  url     = {https://api.semanticscholar.org/CorpusID:245803553},
   doi     = {10.1145/3493700.3493768}
 }
 
@@ -194,6 +194,6 @@
   timestamp = {2018-11-11T17:01:17.000+0100},
   title     = {The FAIR Guiding Principles for scientific data management and stewardship},
   volume    = 3,
-  year      = 2016
+  year      = 2016,
   doi       = {10.1038/sdata.2016.18}
 }
