### 2026-03-13 #10178 @michael-johnston
@danielskatz We've reduced the word count to 1815. If you need us to go further let us know.

I've checked the refs with missing DOIs.  Two are papers which don't have DOIs and the other is to software docs.
