### 2026-03-16 d9dd158
docs(paper): minor changes
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
