### 2026-03-16 c01d246
feat(store): get space id when getting operations
