+++ orchestrator/metastore/sqlstore.py
@@ -555,7 +555,7 @@ class SQLResourceStore(ResourceStore):
         else:
             select_statement = f"{select_statement} {select_name} {select_age} "
 
-        # Add the status to the resources that have it
+        # Add the status and space to the resources that have it
         if kind == orchestrator.core.resources.CoreResourceKinds.OPERATION.value:
             select_status = (
                 orchestrator.metastore.sql.statements.resource_select_data_field(
@@ -564,7 +564,12 @@ class SQLResourceStore(ResourceStore):
                     dialect=self.engine.dialect.name,
                 )
             )
-            select_statement = f"{select_statement} {select_status} "
+            select_space = orchestrator.metastore.sql.statements.resource_select_config_spaces_first(
+                output_column="space",
+                needs_select=False,
+                dialect=self.engine.dialect.name,
+            )
+            select_statement = f"{select_statement} {select_status} {select_space}"
 
         # FROM
         from_statement = "FROM resources "
@@ -630,6 +635,8 @@ class SQLResourceStore(ResourceStore):
         if kind == orchestrator.core.resources.CoreResourceKinds.OPERATION.value:
             columns.insert(-1, "STATUS")
             output_df["STATUS"] = table["status"]
+            columns.insert(-1, "SPACE")
+            output_df["SPACE"] = table["space"]
 
         return output_df[columns]
 
