### 2026-03-16 fbc5bd4
feat(store):  add space id to get operation output
