### 2026-03-16 453161d
chore(deps): update dependency uvicorn to v0.42.0
