### 2026-03-16 6a0f91e
docs(code): doc string
