### 2026-03-16 653dffc
test(cli): add assertion on space id
