### 2026-03-16 72e0f49
fix(ray_tune): Handle if cause attribute not present
