### 2026-03-16 6d5d9cc
feat(core): enable aggregation of multi-valued metrics in ado show entities operation (#681)
* fix(core): Missing cache addition in last commit

* feat(core): enable aggregation in show operations

This is across multi-values metrics returned by a single experiment.

---------

Co-authored-by: Alessandro Pomponio <10339005+AlessandroPomponio@users.noreply.github.com>
