+++ orchestrator/cli/utils/remote/dispatch.py
@@ -26,6 +26,7 @@ from orchestrator.cli.models.remote_submission import (
 from orchestrator.cli.utils.output.prints import (
     ADO_SPINNER_REMOTE_PORT_FORWARD,
     ADO_SPINNER_REMOTE_PREPARING_FILES,
+    console_print,
 )
 from orchestrator.cli.utils.remote.arg_parser import (
     rewrite_flag_values,
@@ -353,6 +354,16 @@ def _symlink_additional_files(
         log.debug("Symlinked additional path %s -> %s", working_dir / path.name, path)
 
 
+def _newest_source_mtime(plugin_path: Path) -> float:
+    """Return the mtime of the most recently modified source file in *plugin_path*."""
+    newest = 0.0
+    for ext in ("*.py", "*.toml", "*.cfg"):
+        for f in plugin_path.rglob(ext):
+            if f.is_file():
+                newest = max(newest, f.stat().st_mtime)
+    return newest
+
+
 def _build_source_wheels(
     from_source: list[str],
     working_dir: Path,
@@ -362,6 +373,8 @@ def _build_source_wheels(
 
     Each entry in *from_source* is a path relative to *repo_root* (or absolute).
     Runs ``uv build --wheel -o dist --clear`` in each plugin directory.
+    Verifies each wheel was freshly built (newer than source files) so that
+    source changes between remote executions produce a new wheel.
 
     Args:
         from_source: Plugin directory paths to build.
@@ -382,8 +395,15 @@ def _build_source_wheels(
                 f"fromSource plugin path does not exist: {plugin_path}"
             )
 
+        try:
+            display_path = str(plugin_path.relative_to(repo_root))
+        except ValueError:
+            display_path = plugin_path_str
+        console_print(f"Building wheel for source package: {display_path}")
+
         dist_dir = plugin_path / "dist"
-        log.info("Building wheel for plugin at %s", plugin_path)
+        source_max_mtime = _newest_source_mtime(plugin_path)
+        build_start = time.time()
         result = subprocess.run(  # noqa: S603
             ["uv", "build", "--wheel", "-o", str(dist_dir), "--clear"],  # noqa: S607
             cwd=str(plugin_path),
@@ -397,19 +417,86 @@ def _build_source_wheels(
         if not wheels:
             raise RuntimeError(f"No wheel produced by uv build in {dist_dir}")
 
+        # Confirm wheels were freshly built (newer than source) so source changes
+        # between executions produce a new wheel.
+        for whl in wheels:
+            whl_mtime = whl.stat().st_mtime
+            if whl_mtime < build_start - 2:
+                raise RuntimeError(
+                    f"Wheel {whl.name} appears to be reused (mtime predates build). "
+                    "Expected a fresh build."
+                )
+            if source_max_mtime > 0 and whl_mtime < source_max_mtime - 2:
+                raise RuntimeError(
+                    f"Wheel {whl.name} is older than source (mtime {whl_mtime} < "
+                    f"newest source {source_max_mtime}). Source changes must produce "
+                    "a new wheel."
+                )
+
         for whl in wheels:
             dest = working_dir / whl.name
             shutil.copy2(whl, dest)
             wheel_names.append(whl.name)
-            log.info("Copied wheel %s to %s", whl.name, working_dir)
+            console_print(f"Built wheel: {whl.name}")
+            log.debug("Copied wheel %s to %s", whl.name, working_dir)
 
     return wheel_names
+
+
+def identify_and_copy_local_wheels(
+    from_pypi: list[str],
+    cwd: Path,
+    working_dir: Path,
+    seen_basenames: set[str],
+) -> tuple[list[str], list[str]]:
[…]
