### 2026-03-16 554f02e
docs(developing): improve prettier-related instructions
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
