### 2026-03-16 9e2ab87
refactor(core): remove orchestrator/plugins module (#695)
* fix(core): Missing cache addition in last commit

* refactor(core): deprecated plugins/samplestore

These classes were just thin wrappers around csvsamplestore and we're only used in a few places.

This change removes them, replacing them with  CSVSampleStore.

Later we can add a proper plugin arch for samplestores.

---------

Co-authored-by: Alessandro Pomponio <10339005+AlessandroPomponio@users.noreply.github.com>
