+++ orchestrator/plugins/samplestores/__init__.py
@@ -1,2 +1,0 @@
-# Copyright IBM Corporation 2025, 2026
-# SPDX-License-Identifier: MIT
