### 2026-03-16 678163e
docs(prettier): Update examples/no-priors-characterization/README.md
Co-authored-by: Alessandro Pomponio <10339005+AlessandroPomponio@users.noreply.github.com>
Signed-off-by: Daniele Lotito <99284466+danielelotito@users.noreply.github.com>
