### 2026-03-16 779e094
docs: address https://github.com/IBM/ado/pull/686#discussion_r2940770044 , all algos are fast
