### 2026-03-16 f7dc9b5
refactor: update structure according to https://github.com/IBM/ado/pull/686/changes/BASE..6c85e73df3a0511640ac254be9a953977abceeee#r2918667637
