+++ plugins/operators/no-priors-characterization/pyproject.toml
@@ -19,6 +19,7 @@ requires = ["setuptools", "setuptools_scm"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
+package-dir = { "" = "src" }
 packages = ["no_priors_characterization"]
 
 [tool.setuptools_scm]
