### 2026-03-16 6f49630
style(logs): address comment
"the error level is the same one that's used for exceptions: either this is an exception worthy occurrence or this should be only a warning"
