### 2026-03-16 ffdde9a
refactor(utils): remove unnecessary if block
Moved resources["STATUS"].apply(lambda x: x.event.value) directly into the operation block removing unnecessary if block.
