### 2026-03-16 bf52147
refactor(metastore): Use resource_select_data_field
- remove resource_select_config_spaces_first
- update resource_select_data_field to have a parameter allowing setting what the extracted field should be called

e.g. get the value config.spaces[0] but return it as "space"
