+++ orchestrator/metastore/sql/statements.py
@@ -236,54 +236,36 @@ def resource_select_data_field(
     field_name: str,
     needs_select: bool = False,
     dialect: Literal["mysql", "sqlite"] = "mysql",
+    output_field_name: str | None = None,
 ) -> str:
-
-    #
-    statement_preamble = "SELECT" if needs_select else ","
-
-    #
-    data_path = f"$.{field_name}"
-    statement = (
-        "{statement_preamble} data ->> '{data_path}' as {field_name}"
-        if dialect == "sqlite"
-        else "{statement_preamble} data->>'{data_path}' as {field_name}"
-    )
-
-    return statement.format(
-        statement_preamble=statement_preamble,
-        data_path=data_path,
-        field_name=field_name,
-    )
-
-
-def resource_select_config_spaces_first(
-    output_column: str = "space",
-    needs_select: bool = False,
-    dialect: Literal["mysql", "sqlite"] = "mysql",
-) -> str:
-    """Extract the first discovery space identifier from config.spaces array.
+    """Extract a field from the JSON ``data`` column.
 
     Args:
-        output_column: The SQL alias for the extracted column.
+        field_name: Dot-notation path within the JSON ``data`` object
+            (e.g. ``"status"`` or ``"config.spaces[0]"``).
         needs_select: If True, prefix with SELECT; otherwise prefix with a
             comma (for appending to an existing SELECT clause).
         dialect: The SQL dialect to use. Determines the JSON extraction syntax.
+        output_field_name: SQL alias for the extracted column. Defaults to
+            ``field_name`` when not provided.
 
     Returns:
-        A SQL fragment that extracts ``config.spaces[0]`` from the ``data``
-        column and aliases it as ``output_column``.
+        A SQL fragment that extracts ``$.{field_name}`` from the ``data``
+        column and aliases it as ``output_field_name`` (or ``field_name``).
     """
     statement_preamble = "SELECT" if needs_select else ","
-    data_path = "$.config.spaces[0]"
+    data_path = f"$.{field_name}"
+    alias = output_field_name or field_name
     statement = (
-        "{statement_preamble} data ->> '{data_path}' as {output_column}"
+        "{statement_preamble} data ->> '{data_path}' as {alias}"
         if dialect == "sqlite"
-        else "{statement_preamble} data->>'{data_path}' as {output_column}"
+        else "{statement_preamble} data->>'{data_path}' as {alias}"
     )
+
     return statement.format(
         statement_preamble=statement_preamble,
         data_path=data_path,
-        output_column=output_column,
+        alias=alias,
     )
 
 
