### 2026-03-16 4266871
test(cli): remove space assertion
The output formatter truncates the space id (and generally the columns). There is no guarantess what part of the id will be in the string to test.
