+++ website/docs/examples/no-priors-characterization.md
@@ -1,0 +1,6 @@
+<!-- markdownlint-disable-next-line first-line-h1 -->
+{%
+
+  include-markdown "../../../examples/no-priors-characterization/README.md"
+  
+%}
