### 2026-03-16 a795673
test(cli): Fix _write_runtime_env tests; add local wheel coverage
Update the four existing test_write_runtime_env_* tests to pass the three new
required arguments (cwd, working_dir, seen_basenames) introduced when
_write_runtime_env was extended to handle local wheel paths in fromPyPI.

Add test_write_runtime_env_local_wheel_in_pypi to verify that a local .whl
path in fromPyPI is copied to the working directory and its entry rewritten to
use RAY_RUNTIME_ENV_CREATE_WORKING_DIR, while a plain PyPI package name is
left unchanged.
