### 2026-03-16 de75cca
docs: remove text about legacy behavior
