### 2026-03-16 1adbac2
chore(schema): using str instead of format string
