### 2026-03-16 1c0ea7f
refactor(core): remove ray.remote decorator from actuators (#698)
* fix(core): Missing cache addition in last commit

* refactor(core): Remove ray.remote decorator from actuators

The decoration was being dynamically removed for any use via the registry since  #185

---------

Co-authored-by: Alessandro Pomponio <10339005+AlessandroPomponio@users.noreply.github.com>
