### 2026-03-16 4930fbc
build: add missing pyproject.toml sections for custom experiments
- Add ado-core dependency
- Add build-system configuration with setuptools
- Add tool.setuptools_scm with root path
- Addresses review comment to follow plugin-development.mdc guidelines
