### 2026-03-16 #10178 @matt-graham
Hi @michael-johnston. Thanks for editing the submission and cutting the word count. I think 1815 should be okay to proceed. From a quick scan through the paper though I would suggest removing the

> ado is open source and available at https://github.com/ibm/ado (Johnston & Pomponio, 2026)

sentence at the end of the summary as this is is somewhat redundant as the repository is linked to in the margin notes on the paper template and a DOI will also be linked to there corresponding to an archive of the reviewed version of repository if accepted, so the Zenodo DOI linked to the (Johnston & Pomponio, 2026) citation will also be redundant / potentially confusing.
