### 2026-03-16 5f2ec0b
docs(paper): remove final sentence in summary.
Its redundant as the margin notes will contain a link to the repo and to a DOI for the reviewed version of the code.
