### 2026-03-18 533e421
chore(operators): simplification
The value should only be a string.
