### 2026-03-18 10e7d2b
fix(schema): ensure experiment_id is string in series representation (#703)
* fix(schema): ensure experiment_id is string

The experiment_id field in ValidMeasurementResult.series_representation and Entity.experiment_series was being set to the ExperimentReference instance.

This rendered as the id when outputting as CSV. However, when outputting as json it was the pydantic JSON dump.

The field properly should just be the string id c.f. entity "identifier" field. This commit makes this the case.

* chore(schema): using str instead of format string

* test(schema): update tests

* fix(operator
[…]
