### 2026-03-18 d1204a5
refactor(cli): address code review comments.
