### 2026-03-18 48ecf35
refactor(cli): address code review comments.
