+++ examples/no-priors-characterization/README.md
@@ -143,6 +143,8 @@ The configuration for a basic sampling operation using CLHS is in
 {%
   include-markdown "./example_yamls/op_basic_sampling.yaml"
 %}
+```
+<!-- prettier-ignore-end -->
 
 To run the operation, execute:
 
