### 2026-03-18 fef58fb
docs: replace markdown preprocessor with symbolic links for no-priors-characterization
