+++ website/docs/examples/no-priors-characterization.md
@@ -1,6 +1,1 @@
-<!-- markdownlint-disable-next-line first-line-h1 -->
-{%
-
-  include-markdown "../../../examples/no-priors-characterization/README.md"
-  
-%}
+../../../examples/no-priors-characterization/README.md
