+++ website/docs/operators/no-priors-characterization.md
@@ -1,6 +1,1 @@
-<!-- markdownlint-disable-next-line first-line-h1 -->
-{%
-
-  include-markdown "../../../plugins/operators/no-priors-characterization/README.md"
-  
-%}
+../../../plugins/operators/no-priors-characterization/README.md
