### 2026-03-19 654988d
fix(core): add binary variable type validation with value normalization
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
