+++ tests/schema/test_domain.py
@@ -936,3 +936,152 @@ def test_open_categorical_variable_type_property_domain() -> None:
     dser = PropertyDomain.model_validate(dump)
     assert dser.variableType == VariableTypeEnum.OPEN_CATEGORICAL_VARIABLE_TYPE
     assert dser.values == ["a", "b", "c"]
+
+
+def test_binary_variable_type_with_invalid_single_true_value() -> None:
+    """Test that BINARY_VARIABLE_TYPE rejects single True value"""
+    with pytest.raises(
+        ValueError,
+        match="BINARY_VARIABLE_TYPE must have exactly 2 values representing",
+    ):
+        PropertyDomain(
+            variableType=VariableTypeEnum.BINARY_VARIABLE_TYPE, values=[True]
+        )
+
+
+def test_binary_variable_type_with_invalid_single_false_value() -> None:
+    """Test that BINARY_VARIABLE_TYPE rejects single False value"""
+    with pytest.raises(
+        ValueError,
+        match="BINARY_VARIABLE_TYPE must have exactly 2 values representing",
+    ):
+        PropertyDomain(
+            variableType=VariableTypeEnum.BINARY_VARIABLE_TYPE, values=[False]
+        )
+
+
+def test_binary_variable_type_with_invalid_single_numeric_one() -> None:
+    """Test that BINARY_VARIABLE_TYPE rejects single numeric value 1"""
+    with pytest.raises(
+        ValueError,
+        match="BINARY_VARIABLE_TYPE must have exactly 2 values representing",
+    ):
+        PropertyDomain(variableType=VariableTypeEnum.BINARY_VARIABLE_TYPE, values=[1])
+
+
+def test_binary_variable_type_with_invalid_single_numeric_zero() -> None:
+    """Test that BINARY_VARIABLE_TYPE rejects single numeric value 0"""
+    with pytest.raises(
+        ValueError,
+        match="BINARY_VARIABLE_TYPE must have exactly 2 values representing",
+    ):
+        PropertyDomain(variableType=VariableTypeEnum.BINARY_VARIABLE_TYPE, values=[0])
+
+
+def test_binary_variable_type_with_invalid_numeric_values() -> None:
+    """Test that BINARY_VARIABLE_TYPE rejects invalid numeric values like [2, 3]"""
+    with pytest.raises(
+        ValueError,
+        match="BINARY_VARIABLE_TYPE must have values equivalent to",
+    ):
+        PropertyDomain(
+            variableType=VariableTypeEnum.BINARY_VARIABLE_TYPE, values=[2, 3]
+        )
+
+
+def test_binary_variable_type_with_invalid_string_values() -> None:
+    """Test that BINARY_VARIABLE_TYPE rejects string values"""
+    with pytest.raises(
+        ValueError,
+        match="BINARY_VARIABLE_TYPE must have values equivalent to",
+    ):
+        PropertyDomain(
+            variableType=VariableTypeEnum.BINARY_VARIABLE_TYPE, values=["yes", "no"]
+        )
+
+
+def test_binary_variable_type_with_valid_boolean_values() -> None:
+    """Test that BINARY_VARIABLE_TYPE accepts [False, True]"""
+    domain = PropertyDomain(
+        variableType=VariableTypeEnum.BINARY_VARIABLE_TYPE, values=[False, True]
+    )
+    assert domain.values == [False, True]
+    assert domain.domain_values == [False, True]
+
+
+def test_binary_variable_type_with_valid_reversed_boolean_values() -> None:
+    """Test that BINARY_VARIABLE_TYPE accepts [True, False] and normalizes it"""
+    domain = PropertyDomain(
[…]
