### 2026-03-19 #10178 @matt-graham
@simoninireland @Torigh @Daniel-Dodd can you review this submission for JOSS?
