### 2026-03-20 5af3ad1
fix(vllm_performance): Replace binary type with discrete type for properties in geospatial experiments (#720)
feat(vllm_performance): Replaces binary type properties with discrete ones in the geospatial experiments

Signed-off-by: Christian Pinto <christian.pinto@ibm.com>
