### 2026-03-20 47fb44b
fix: use generatorid parameter in space_df_connector
Update get_list_of_entities_from_df_and_space() to pass
'no_priors_characterization' as generatorid when creating entities.
This ensures entities are properly tagged with their generator instead
of defaulting to 'unk'.
