### 2026-03-20 d045d9f
docs: update README to reflect correct generatorid
Replace 'unk' with 'no_priors_characterization' in example output
to reflect the actual generatorid now being set correctly.
