+++ examples/no-priors-characterization/README.md
@@ -211,7 +211,7 @@ submitted and completed:
 (RandomWalk pid=82843) Continuous batching: SUBMIT EXPERIMENT. Submitted experiment custom_experiments.calculate_reaction_yield for temperature.353-concentration.4.1-catalyst_amount.4.5. Request identifier: c72090
 (RandomWalk pid=82843)
 (RandomWalk pid=82843) Continuous batching: SUMMARY. Entities sampled and submitted: 2. Experiments completed: 1 Waiting on 1 active requests. There are 0 dependent experiments
-(RandomWalk pid=82843) Continuous Batching: EXPERIMENT COMPLETION. Received finished notification for experiment in measurement request in group 1: request-c72090-experiment-calculate_reaction_yield-entities-temperature.353-concentration.4.1-catalyst_amount.4.5 (unk)-requester-randomwalk-1.6.1.dev9+03a65e7b.dirty-9a277d-time-2026-03-10 11:43:11.066810+00:00
+(RandomWalk pid=82843) Continuous Batching: EXPERIMENT COMPLETION. Received finished notification for experiment in measurement request in group 1: request-c72090-experiment-calculate_reaction_yield-entities-temperature.353-concentration.4.1-catalyst_amount.4.5 (no_priors_characterization)-requester-randomwalk-1.6.1.dev9+03a65e7b.dirty-9a277d-time-2026-03-10 11:43:11.066810+00:00
 ```
 
 <!-- markdownlint-enable line-length -->
@@ -259,9 +259,9 @@ yield values.
 ┌───────┬──────────────────────────────────────────────────────────┬─────────────┬──────────────────────────────────────────────────┬─────────────┬───────────────┬─────────────────┬──────────┐
 │ INDEX │ identifier                                               │ generatorid │ experiment_id                                    │ temperature │ concentration │ catalyst_amount │ yield    │
 ├───────┼──────────────────────────────────────────────────────────┼─────────────┼──────────────────────────────────────────────────┼─────────────┼───────────────┼─────────────────┼──────────┤
-│ 0     │ temperature.300-concentration.1.0-catalyst_amount.2.0    │ unk         │ custom_experiments.calculate_reaction_yield      │ 300         │ 1.0           │ 2.0             │ 45.23    │
-│ 1     │ temperature.350-concentration.2.5-catalyst_amount.5.0    │ unk         │ custom_experiments.calculate_reaction_yield      │ 350         │ 2.5           │ 5.0             │ 78.91    │
-│ 2     │ temperature.400-concentration.0.5-catalyst_amount.1.0    │ unk         │ custom_experiments.calculate_reaction_yield      │ 400         │ 0.5           │ 1.0             │ 92.15    │
+│ 0     │ temperature.300-concentration.1.0-catalyst_amount.2.0    │ no_priors_characterization │ custom_experiments.calculate_reaction_yield      │ 300         │ 1.0           │ 2.0             │ 45.23    │
+│ 1     │ temperature.350-concentration.2.5-catalyst_amount.5.0    │ no_priors_characterization │ custom_experiments.calculate_reaction_yield      │ 350         │ 2.5           │ 5.0             │ 78.91    │
+│ 2     │ temperature.400-concentration.0.5-catalyst_amount.1.0    │ no_priors_characterization │ custom_experiments.calculate_reaction_yield      │ 400         │ 0.5           │ 1.0             │ 92.15    │
 │ ...   │ ...                                                      │ ...         │ ...                                              │ ...         │ ...           │ ...             │ ...      │
 └───────┴──────────────────────────────────────────────────────────┴─────────────┴──────────────────────────────────────────────────┴─────────────┴───────────────┴─────────────────┴──────────┘
 ```
