### 2026-03-21 e6d9329
docs(concepts): refine core concepts docs
Introduce core-concepts/properties-and-domains.md.
Reframe concepts.md around what can be measured, what you want to measure, and the shared sample store.
Slim actuators.md by moving property/domain material to the new page and
dropping the long experiment listing example. Update and refine discovery-spaces,
entity-spaces, data-sharing, and working-with-actuators for consistency.
