+++ website/docs/core-concepts/data-sharing.md
@@ -1,102 +1,80 @@
 # Shared Sample Stores
 
-In `ado` Entities and measurement results are stored in a database called a
-**Sample Store**. This document describes how Sample Stores enable sharing of
-data. For more general information about these databases see
+In `ado`, Entities and measurement results are stored in a database called a
+**Sample Store**. For more on how Sample Stores are configured and managed see
 [their dedicated page](../resources/sample-stores.md).
 
-There are two key points that underpin data reuse in `ado`:
+Two principles underpin data reuse in `ado`:
 
-- You can **share** a Sample Store between multiple Discovery Spaces
-    - This allows a Discovery Space to (re)use relevant Entities and Measurements
-       stored in the Sample Store by operations on other Discovery Spaces
-- **Entities are always shared**. There is only one entry in a Sample Store for
-  an Entity
+- **A Sample Store can be shared across multiple Discovery Spaces.** This allows
+  any Discovery Space to access Entities and measurements recorded by operations
+  on other Discovery Spaces that use the same store.
+- **Each Entity has exactly one record in a Sample Store.** If two Discovery
+  Spaces both include the same Entity, they reference the same record — there is
+  no duplication.
 
 > [!NOTE]
 >
-> To maximize the chance of data-reuse, similar Discovery Spaces should use the
-> same Sample Store. However, Discovery Spaces do not have to be similar to use
-> the same Sample Store.
+> To maximise the chance of data reuse, similar Discovery Spaces should use the
+> same Sample Store. However, any Discovery Spaces can share a store regardless
+> of how similar they are.
 
-## When data can be shared in `ado`
-
-There are two situations where data can be shared between Discovery Spaces in
-`ado`:
-
-- **Data Retrieval**: retrieving data about entities and measurements from the
-  Discovery Space e.g. `ado show entities space`
-- **Data Generation**: When performing an explore operation on a Discovery
-  Space - this type of data reuse is called `memoization`
-
-## How `ado` determines what data can be shared
-
-As a quick recap, a Discovery Space is composed of:
-
-- an [Entity Space](entity-spaces.md) which describes a set of Entities (points)
-  to be measured
-- a [Measurement Space](actuators.md#measurement-space) which describes a set of
-  Experiments to apply to the points
+## How `ado` matches shared data
 
 ### Entities
 
-Each Entity in the Entity Space has a unique identifier, usually determined by
-its set of constitutive property values. For example, if an Entity has two
-constitutive properties `X` an `Y` with values 4 and 10, its id will be
-'X:4-Y:10'. Since the identifiers of all the Entities in the Entity Space are
-known, the Sample Store can be searched to see if it contains a record for any
-of the Entities.
+Each Entity has a unique identifier derived from its
+[constitutive property](properties-and-domains.md#property-types) values.
+For example, an Entity with properties `X=4` and `Y=10` gets the id
+`X:4-Y:10`. `ado` uses these identifiers to look up Entities in the Sample
+Store, regardless of which Discovery Space originally recorded them.
 
 ### Measurements
 
-Each experiment in a Measurement Space has a unique identifier, determined from
-its base name plus any optional properties that have been explicitly set. When
-an Entity is retrieved from the Sample Store, it contains results of all the
-experiments that have been applied to it. If the identifier of a result matches
-the identifier of an Experiment in the Measurement Space, `ado` determines it
-can be reused.
+Each Experiment also has a unique identifier (its name plus any explicitly set
+optional inputs). When an Entity is retrieved from the Sample Store, it carries
+the results of all Experiments that have been applied to it. `ado` checks
+whether any of those result identifiers match an Experiment in the current
+Measurement Space — if so, the result can be reused.
 
[…]
