### 2026-03-23 67c0693
docs(concepts): tidying
