### 2026-03-23 911a8c1
docs(concepts): add sub-domains
