### 2026-03-23 005589c
chore(deps): update dependency python-discovery to v1.2.0
