### 2026-03-23 611ef52
chore(deps): update dependency attrs to v26
