### 2026-03-23 0a0a010
refactor(core): prefer not having anything specified for BINARY_VARIABLE_TYPEs
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
