### 2026-03-23 1ec8e56
Apply suggestion from @michael-johnston
Signed-off-by: Michael Johnston <66301584+michael-johnston@users.noreply.github.com>
