### 2026-03-23 669e42f
refactor(cli): remove useless code
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
