### 2026-03-23 23874bf
docs: fix table layout
