### 2026-03-23 b1f8076
refactor: reduce duplication
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
