+++ orchestrator/cli/utils/legacy/common.py
@@ -1,0 +1,175 @@
+# Copyright IBM Corporation 2025, 2026
+# SPDX-License-Identifier: MIT
+
+"""Common utilities for legacy validator handling"""
+
+from typing import TYPE_CHECKING
+
+import pydantic
+from rich.console import Console
+
+if TYPE_CHECKING:
+    from orchestrator.core.legacy.metadata import LegacyValidatorMetadata
+    from orchestrator.core.resources import CoreResourceKinds
+
+
+def import_legacy_validators() -> None:
+    """Import all legacy validator modules to ensure they're registered"""
+    # Import validator modules to trigger decorator registration
+    try:
+        # Discovery Space validators
+        import orchestrator.core.legacy.validators.discoveryspace.entitysource_to_samplestore  # noqa: F401
+        import orchestrator.core.legacy.validators.discoveryspace.properties_field_removal  # noqa: F401
+
+        # Operation validators
+        import orchestrator.core.legacy.validators.operation.actuators_field_removal  # noqa: F401
+        import orchestrator.core.legacy.validators.operation.randomwalk_mode_to_sampler_config  # noqa: F401
+
+        # Sample Store validators
+        import orchestrator.core.legacy.validators.resource.entitysource_to_samplestore  # noqa: F401
+        import orchestrator.core.legacy.validators.samplestore.entitysource_migrations  # noqa: F401
+        import orchestrator.core.legacy.validators.samplestore.v1_to_v2_csv_migration  # noqa: F401
+    except ImportError:
+        pass  # Validators may not be available in all installations
+
+
+def print_validator_suggestions(
+    validators: list["LegacyValidatorMetadata"],
+    resource_type: "CoreResourceKinds",
+    console: Console,
+    show_all_validators: bool = False,
+) -> None:
+    """Print legacy validator suggestions to the console
+
+    Args:
+        validators: List of applicable validators
+        resource_type: The resource type
+        console: Rich console to print to
+        show_all_validators: If True, show all validators in the command example
+    """
+    # Resources can be referenced by their CoreResourceKinds value or by shorthands
+    # from cli_shorthands_to_cli_names in orchestrator/cli/utils/resources/mappings.py
+    resource_cli_name = resource_type.value
+
+    console.print("\n[bold cyan]Available legacy validators:[/bold cyan]\n")
+
+    for validator in validators:
+        console.print(f"  • [green]{validator.identifier}[/green]")
+        console.print(f"    {validator.description}")
+        console.print(f"    Handles: {', '.join(validator.deprecated_fields)}")
+        console.print(f"    Deprecated: v{validator.deprecated_from_version}")
+        console.print()
+
+    console.print("[bold magenta]To upgrade using legacy validators:[/bold magenta]")
+    if show_all_validators:
+        validator_args = " ".join(
+            f"--apply-legacy-validator {v.identifier}" for v in validators
+        )
+    else:
+        validator_args = f"--apply-legacy-validator {validators[0].identifier}"
+    console.print(f"  ado upgrade {resource_cli_name} {validator_args}")
+    console.print()
+    console.print("[bold magenta]To list all legacy validators:[/bold magenta]")
+    console.print(f"  ado upgrade {resource_cli_name} --list-legacy-validators")
+
+
+# Made with Bob
+
+
+def extract_deprecated_fields_from_validation_error(
+    error: pydantic.ValidationError,
+) -> tuple[set[str], dict[str, list[str]], set[str]]:
+    """Extract field names and error details from pydantic validation errors
+
+    Args:
+        error: The pydantic validation error
+
+    Returns:
+        Tuple of (full field paths, field error details mapping, leaf field names)
+        - full field paths: Set of full dotted paths like 'config.specification.module.moduleType'
[…]
