### 2026-03-23 22aac8e
refactor(cli): do not hardcode field patterns
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
