### 2026-03-23 cd1b0fe
feat(no-priors): create standalone no-priors characterization operator (#686)
* refactor(operators): extract no-priors characterization from TRIM into separate plugin

- Create new no-priors-characterization operator plugin per issue [#683](https://github.com/IBM/ado/issues/683)
- Move sampling logic to new plugin
- Update TRIM to depend on ado-no-priors-characterization
- Add examples and documentation
- Fix small issue about edge case in high dim sampling

Related to [#683](https://github.com/IBM/ado/issues/683)

* docs: Update examples/no-priors-characterization/README.md

Co-authored-by: A
[…]
