+++ plugins/operators/trim/trim/trim_pydantic.py
@@ -5,9 +5,8 @@ import logging
 from typing import Annotated
 
 import pydantic
+from no_priors_characterization.no_priors_pydantic import NoPriorsParameters
 from pydantic import BaseModel, ConfigDict, Field, model_validator
-
-from trim.no_priors_pydantic import NoPriorsParameters
 
 
 class SamplingBudget(pydantic.BaseModel):
