+++ examples/no-priors-characterization/README.md
@@ -1,0 +1,281 @@
+# Exploring Parameter Spaces with No-Priors Characterization
+
+<!-- markdownlint-disable no-blanks-blockquote -->
+
+> [!NOTE] The scenario
+>
+> You have an experiment with multiple parameters,
+> and you want to understand how these parameters influence the outcome.
+> **In this example, `ado`'s no-priors characterization operator is used to
+> systematically sample and measure the target property across the parameter
+> space using various sampling strategies aimed at covering uniformly the
+> parameter space.** Using the no-priors characterization
+> operator involves:
+>
+> 1. Defining the parameter space to explore.
+> 2. Creating an `operation` that uses no-priors characterization to sample
+>    points using a chosen strategy.
+> 3. Observing the sampling process as it measures the target output property with
+>    the selected strategy.
+
+> [!IMPORTANT] Prerequisites
+>
+> Get the example files and install dependencies:
+>
+> ```commandline
+> git clone https://github.com/IBM/ado.git
+> cd ado
+> pip install plugins/operators/no-priors-characterization/
+> pip install examples/no-priors-characterization/custom_experiments/
+> ```
+
+> [!CAUTION]
+>
+> All commands below assume you are running them from the
+> **top-level of the `ado` repository**.
+
+> [!TIP] TL;DR
+>
+> To create a `discoveryspace` and explore it with the no-priors
+> characterization operator, execute the following from the root of the `ado`
+> repository:
+>
+> ```bash
+> : # Create the space to explore based on a custom experiment
+> ado create space -f \
+>   examples/no-priors-characterization/example_yamls/space_reaction.yaml \
+>   --new-sample-store
+> : # Explore it with no-priors characterization!
+> ado create operation -f \
+>   examples/no-priors-characterization/example_yamls/op_basic_sampling.yaml \
+>     --use-latest space
+> ```
+
+<!-- markdownlint-enable no-blanks-blockquote -->
+
+## What is No-Priors Characterization?
+
+**No-Priors Characterization** is a sampling operator designed to explore a
+parameter space systematically without requiring any prior knowledge or
+existing data. It's perfect for initial exploration of a system where you want
+to gather representative samples across the entire parameter space.
+
+**Handling Existing Measurements**: If the discovery space already contains
+measured entities for the target property, the operator automatically:
+
+- Identifies which entities have already been measured
+- Excludes them from sampling, so that the operator will measure the
+  desired amount of entities
+
+The operator supports three sampling strategies:
+
+1. **Random Sampling (`random`)**: Uniformly random sampling across the
+   parameter space. Fast and simple, but may not provide optimal coverage.
+
+2. **Concatenated Latin Hypercube Sampling (`clhs`)**: An adaptation of Latin
+   Hypercube Sampling for discrete spaces. Good coverage in each dimension is
+   obtained by avoiding measuring parameters combinations with many common
+   values. Particularly effective for high-dimensional spaces.
+
+3. **Sobol Sampling (`sobol`)**: A quasi-random low-discrepancy sampling
+   method that provides better space-filling properties than pure random
+   sampling. It has been adapted for discrete parameter spaces. It falls back
+   to Concatenated Latin Hypercube Sampling when collisions are detected
+   during the discretization process.
+
+<!-- markdownlint-disable no-blanks-blockquote -->
+> [!CAUTION]
+>
+> In the current version of no-priors characterization, if not all
+> measurements produce the observed target output property specified in the
+> `operation.parameters.targetOutput` field, the operation may fail or produce
+> incomplete results. Ensure all experiments return the expected target property.
+
[…]
