+++ website/docs/examples/example_yamls/op_basic_sampling.yaml
@@ -1,0 +1,1 @@
+../../../../examples/no-priors-characterization/example_yamls/op_basic_sampling.yaml
