+++ website/mkdocs.yml
@@ -155,6 +155,7 @@ nav:
       - Testing the throughput of an inference endpoint: examples/vllm-performance-endpoint.md
       - Exploring vLLM deployment performance: examples/vllm-performance-full.md
       - Efficiently Exploring Parameter Spaces with TRIM: examples/trim.md
+      - Characterizing Spaces Without Prior Knowledge: examples/no-priors-characterization.md
   - Core Concepts:
       - core-concepts/concepts.md
       - Entities and Entity Spaces: core-concepts/entity-spaces.md
@@ -187,3 +188,4 @@ nav:
       - The Random Walk Operator: operators/random-walk.md
       - The Ray Tune Operator: operators/optimisation-with-ray-tune.md
       - The TRIM Operator: operators/trim.md
+      - The No-Priors Characterization Operator: operators/no-priors-characterization.md
