### 2026-03-24 34a3bb9
build: refactor to achieve proper automatic package discovery
