### 2026-03-24 7cf2266
refactor(cli): remove debug code
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
