### 2026-03-24 0bbac99
fix(legacy): avoid modifying wrong fields in legacy validators
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
