### 2026-03-24 1ed3874
feat(cli): add transaction safety to ado upgrade of legacy resources
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
