### 2026-03-24 e7bc134
test: add tests for new functionalities
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
