+++ tests/core/test_upgrade_transaction_safety.py
@@ -1,0 +1,277 @@
+# Copyright IBM Corporation 2025, 2026
+# SPDX-License-Identifier: MIT
+
+"""Tests for Phase 1 transaction safety in upgrade handler"""
+
+from unittest.mock import MagicMock, patch
+
+import pytest
+import typer
+
+from orchestrator.core.legacy.registry import LegacyValidatorRegistry, legacy_validator
+from orchestrator.core.resources import CoreResourceKinds
+
+
+class TestUpgradeTransactionSafety:
+    """Test transaction safety in upgrade handler - validate-all-before-save pattern"""
+
+    def setup_method(self) -> None:
+        """Clear the registry before each test"""
+        LegacyValidatorRegistry._validators = {}
+
+    def test_all_resources_validated_before_any_saved(self) -> None:
+        """Test that all resources are validated before any are saved"""
+
+        # Register a test validator
+        @legacy_validator(
+            identifier="test_transaction_validator",
+            resource_type=CoreResourceKinds.SAMPLESTORE,
+            deprecated_fields=["old_field"],
+            deprecated_from_version="1.0.0",
+            removed_from_version="2.0.0",
+            description="Test transaction validator",
+        )
+        def test_validator(data: dict) -> dict:
+            if "config" in data and "old_field" in data["config"]:
+                data["config"]["new_field"] = data["config"].pop("old_field")
+            return data
+
+        # Create mock resources
+        mock_resource1 = MagicMock()
+        mock_resource1.model_dump.return_value = {
+            "kind": "samplestore",
+            "identifier": "res1",
+            "config": {"old_field": "value1"},
+        }
+
+        mock_resource2 = MagicMock()
+        mock_resource2.model_dump.return_value = {
+            "kind": "samplestore",
+            "identifier": "res2",
+            "config": {"old_field": "value2"},
+        }
+
+        # Mock resource class
+        mock_resource_class = MagicMock()
+        validated_resources = []
+
+        def mock_validate(data: dict) -> MagicMock:
+            validated = MagicMock()
+            validated.model_dump.return_value = data
+            validated_resources.append(data["identifier"])
+            return validated
+
+        mock_resource_class.model_validate.side_effect = mock_validate
+
+        # Mock SQL store
+        mock_sql_store = MagicMock()
+        mock_sql_store.getResourcesOfKind.return_value = {
+            "res1": mock_resource1,
+            "res2": mock_resource2,
+        }
+        mock_sql_store.getResourceIdentifiersOfKind.return_value = {
+            "IDENTIFIER": ["res1", "res2"]
+        }
+        mock_sql_store.getResourceRaw.side_effect = lambda id: (
+            mock_resource1.model_dump() if id == "res1" else mock_resource2.model_dump()
+        )
+
+        update_calls = []
+
+        def track_update(resource: MagicMock) -> None:
+            update_calls.append(resource.model_dump()["identifier"])
+
+        mock_sql_store.updateResource.side_effect = track_update
+
+        # Mock parameters
+        mock_params = MagicMock()
+        mock_params.apply_legacy_validator = ["test_transaction_validator"]
+        mock_params.list_legacy_validators = False
+        mock_params.ado_configuration.project_context = "test_context"
+
+        # Patch dependencies
+        with (
+            patch(
+                "orchestrator.cli.utils.resources.handlers.get_sql_store",
+                return_value=mock_sql_store,
+            ),
+            patch(
+                "orchestrator.core.kindmap",
+                {"samplestore": mock_resource_class},
+            ),
+            patch("orchestrator.cli.utils.legacy.common.import_legacy_validators"),
+            patch("orchestrator.cli.utils.resources.handlers.Status"),
+            patch("orchestrator.cli.utils.resources.handlers.console_print"),
+        ):
[…]
