+++ tests/core/test_validator_scope_fixes.py
@@ -1,0 +1,283 @@
+# Copyright IBM Corporation 2025, 2026
+# SPDX-License-Identifier: MIT
+
+"""Tests for Phase 1 validator scope fixes - verifying validators only operate on config level"""
+
+from orchestrator.core.legacy.registry import LegacyValidatorRegistry
+
+
+class TestValidatorScopeFixes:
+    """Test that validators correctly operate only on config level after Phase 1 fixes"""
+
+    @classmethod
+    def setup_class(cls) -> None:
+        """Import validators once for all tests in this class"""
+        # Import all validators to register them
+        from orchestrator.cli.utils.legacy.common import import_legacy_validators
+
+        import_legacy_validators()
+
+    def test_discoveryspace_properties_removal_scope(self) -> None:
+        """Verify properties_field_removal only modifies config, not resource level"""
+
+        # Test data with 'properties' at both levels
+        resource_data = {
+            "kind": "discoveryspace",
+            "identifier": "test-space",
+            "properties": "SHOULD_NOT_BE_REMOVED",  # Resource level
+            "config": {
+                "properties": ["prop1", "prop2"],  # Config level (should be removed)
+                "sampleStoreIdentifier": "store-1",
+            },
+        }
+
+        # Get validator
+        validator = LegacyValidatorRegistry.get_validator(
+            "discoveryspace_properties_field_removal"
+        )
+        assert validator is not None
+
+        # Apply validator
+        result = validator.validator_function(resource_data.copy())
+
+        # Verify: resource-level field unchanged, config-level field removed
+        assert "properties" in result  # Resource level preserved
+        assert result["properties"] == "SHOULD_NOT_BE_REMOVED"
+        assert "properties" not in result["config"]  # Config level removed
+        assert result["config"]["sampleStoreIdentifier"] == "store-1"
+
+    def test_discoveryspace_entitysource_migration_scope(self) -> None:
+        """Verify entitysource_to_samplestore only modifies config, not resource level"""
+
+        # Test data with entitySourceIdentifier at both levels
+        resource_data = {
+            "kind": "discoveryspace",
+            "identifier": "test-space",
+            "entitySourceIdentifier": "SHOULD_NOT_BE_REMOVED",  # Resource level
+            "config": {
+                "entitySourceIdentifier": "old-source",  # Config level (should migrate)
+            },
+        }
+
+        # Get validator
+        validator = LegacyValidatorRegistry.get_validator(
+            "discoveryspace_entitysource_to_samplestore"
+        )
+        assert validator is not None
+
+        # Apply validator
+        result = validator.validator_function(resource_data.copy())
+
+        # Verify: resource-level field unchanged, config-level field migrated
+        assert "entitySourceIdentifier" in result  # Resource level preserved
+        assert result["entitySourceIdentifier"] == "SHOULD_NOT_BE_REMOVED"
+        assert "entitySourceIdentifier" not in result["config"]  # Config level removed
+        assert result["config"]["sampleStoreIdentifier"] == "old-source"  # Migrated
+
+    def test_operation_actuators_removal_scope(self) -> None:
+        """Verify actuators_field_removal only modifies config, not resource level"""
+
+        # Test data with actuators at both levels
+        resource_data = {
+            "kind": "operation",
+            "identifier": "test-op",
+            "actuators": "SHOULD_NOT_BE_REMOVED",  # Resource level
+            "config": {
+                "actuators": ["act1", "act2"],  # Config level (should be removed)
+                "operatorIdentifier": "op1",
+            },
+        }
+
+        # Get validator
+        validator = LegacyValidatorRegistry.get_validator(
+            "operation_actuators_field_removal"
+        )
+        assert validator is not None
+
+        # Apply validator
+        result = validator.validator_function(resource_data.copy())
+
[…]
