+++ orchestrator/core/legacy/validators/resource/__init__.py
@@ -3,4 +3,8 @@
 
 """Legacy validators for generic resource migrations"""
 
+from orchestrator.core.legacy.validators.resource import entitysource_to_samplestore
+
+__all__ = ["entitysource_to_samplestore"]
+
 # Made with Bob
