+++ orchestrator/core/legacy/validators/discoveryspace/__init__.py
@@ -3,4 +3,11 @@
 
 """Legacy validators for discovery space migrations"""
 
+from orchestrator.core.legacy.validators.discoveryspace import (
+    entitysource_to_samplestore,
+    properties_field_removal,
+)
+
+__all__ = ["entitysource_to_samplestore", "properties_field_removal"]
+
 # Made with Bob
