+++ orchestrator/core/legacy/validators/operation/__init__.py
@@ -3,4 +3,11 @@
 
 """Legacy validators for operation migrations"""
 
+from orchestrator.core.legacy.validators.operation import (
+    actuators_field_removal,
+    randomwalk_mode_to_sampler_config,
+)
+
+__all__ = ["actuators_field_removal", "randomwalk_mode_to_sampler_config"]
+
 # Made with Bob
