+++ orchestrator/core/legacy/validators/__init__.py
@@ -3,4 +3,14 @@
 
 """Legacy validators for deprecated resource formats"""
 
+# Import all validator subpackages to trigger registration
+from orchestrator.core.legacy.validators import (
+    discoveryspace,
+    operation,
+    resource,
+    samplestore,
+)
+
+__all__ = ["discoveryspace", "operation", "resource", "samplestore"]
+
 # Made with Bob
