### 2026-03-24 f5599d0
feat(core): add utility getters/setters for dictionaries
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
