+++ tests/core/test_legacy_utils.py
@@ -1,0 +1,192 @@
+# Copyright IBM Corporation 2025, 2026
+# SPDX-License-Identifier: MIT
+
+"""Tests for legacy validator utility functions"""
+
+from orchestrator.core.legacy.utils import (
+    get_nested_value,
+    has_nested_field,
+    remove_nested_field,
+    set_nested_value,
+)
+
+
+class TestGetNestedValue:
+    """Tests for get_nested_value function"""
+
+    def test_simple_path(self) -> None:
+        """Test getting a simple top-level field"""
+        data = {"config": {"properties": ["a", "b"]}}
+        parent, field = get_nested_value(data, "config")
+        assert parent == data
+        assert field == "config"
+
+    def test_nested_path(self) -> None:
+        """Test getting a nested field"""
+        data = {"config": {"specification": {"module": {"moduleType": "test"}}}}
+        parent, field = get_nested_value(data, "config.specification.module.moduleType")
+        assert parent == {"moduleType": "test"}
+        assert field == "moduleType"
+
+    def test_nonexistent_path(self) -> None:
+        """Test getting a path that doesn't exist"""
+        data = {"config": {}}
+        parent, field = get_nested_value(data, "config.nonexistent.field")
+        assert parent is None
+        assert field is None
+
+    def test_path_through_non_dict(self) -> None:
+        """Test path that goes through a non-dict value"""
+        data = {"config": "string_value"}
+        parent, field = get_nested_value(data, "config.field")
+        assert parent is None
+        assert field is None
+
+
+class TestSetNestedValue:
+    """Tests for set_nested_value function"""
+
+    def test_set_simple_value(self) -> None:
+        """Test setting a simple nested value"""
+        data = {"config": {}}
+        result = set_nested_value(data, "config.test", "value")
+        assert result is True
+        assert data["config"]["test"] == "value"
+
+    def test_set_deeply_nested_value(self) -> None:
+        """Test setting a deeply nested value"""
+        data = {"config": {"specification": {"module": {}}}}
+        result = set_nested_value(data, "config.specification.module.type", "new_type")
+        assert result is True
+        assert data["config"]["specification"]["module"]["type"] == "new_type"
+
+    def test_set_nonexistent_path(self) -> None:
+        """Test setting a value on a nonexistent path"""
+        data = {"config": {}}
+        result = set_nested_value(data, "config.nonexistent.field", "value")
+        assert result is False
+        assert "nonexistent" not in data["config"]
+
+    def test_overwrite_existing_value(self) -> None:
+        """Test overwriting an existing value"""
+        data = {"config": {"test": "old_value"}}
+        result = set_nested_value(data, "config.test", "new_value")
+        assert result is True
+        assert data["config"]["test"] == "new_value"
+
+
+class TestRemoveNestedField:
+    """Tests for remove_nested_field function"""
+
+    def test_remove_simple_field(self) -> None:
+        """Test removing a simple field"""
+        data = {"config": {"properties": ["a", "b"], "other": "value"}}
+        result = remove_nested_field(data, "config.properties")
+        assert result is True
+        assert "properties" not in data["config"]
+        assert data["config"]["other"] == "value"
+
+    def test_remove_deeply_nested_field(self) -> None:
+        """Test removing a deeply nested field"""
+        data = {
+            "config": {
+                "specification": {"module": {"moduleType": "old", "other": "value"}}
+            }
+        }
+        result = remove_nested_field(data, "config.specification.module.moduleType")
+        assert result is True
+        assert "moduleType" not in data["config"]["specification"]["module"]
+        assert data["config"]["specification"]["module"]["other"] == "value"
+
+    def test_remove_nonexistent_field(self) -> None:
+        """Test removing a field that doesn't exist"""
+        data = {"config": {}}
[…]
