### 2026-03-24 3b81216
feat(core): add field_paths to validator metadata
For safer updates

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
