### 2026-03-24 f1a2503
fix(docs): update description with table
