### 2026-03-24 707e1fe
fix(docs): admonition
