### 2026-03-24 be8a4ce
refactor(legacy): simplify validators
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
