+++ orchestrator/core/legacy/validators/samplestore/entitysource_migrations.py
@@ -4,6 +4,11 @@
 """Legacy validators for migrating entitysource to samplestore naming"""
 
 from orchestrator.core.legacy.registry import legacy_validator
+from orchestrator.core.legacy.utils import (
+    get_nested_value,
+    has_nested_field,
+    set_nested_value,
+)
 from orchestrator.core.resources import CoreResourceKinds
 
 
@@ -14,13 +19,13 @@ from orchestrator.core.resources import CoreResourceKinds
     deprecated_from_version="0.9.6",
     removed_from_version="1.0.0",
     description="Converts moduleType value from 'entity_source' to 'sample_store'",
+    field_paths=["config.moduleType"],
 )
 def migrate_module_type(data: dict) -> dict:
     """Convert moduleType from entity_source to sample_store
 
-    This validator recursively searches for moduleType fields within the config
-    and converts them from 'entity_source' to 'sample_store'. It operates only
-    within the config level, matching the original pydantic validator behavior.
+    This validator checks for moduleType field within the config
+    and converts it from 'entity_source' to 'sample_store'.
 
     Old format:
         config:
@@ -40,26 +45,12 @@ def migrate_module_type(data: dict) -> dict:
     if not isinstance(data, dict):
         return data
 
-    # Only operate within config
-    if "config" not in data or not isinstance(data["config"], dict):
-        return data
+    # Check and update config.moduleType
+    if has_nested_field(data, "config.moduleType"):
+        parent, field = get_nested_value(data, "config.moduleType")
+        if parent is not None and parent[field] == "entity_source":
+            set_nested_value(data, "config.moduleType", "sample_store")
 
-    def convert_module_type_in_dict(d: dict) -> None:
-        """Recursively convert moduleType in nested structures"""
-        if "moduleType" in d and d["moduleType"] == "entity_source":
-            d["moduleType"] = "sample_store"
-
-        # Check in nested structures
-        for value in d.values():
-            if isinstance(value, dict):
-                convert_module_type_in_dict(value)
-            elif isinstance(value, list):
-                for item in value:
-                    if isinstance(item, dict):
-                        convert_module_type_in_dict(item)
-
-    # Start recursion from config level
-    convert_module_type_in_dict(data["config"])
     return data
 
 
@@ -70,13 +61,13 @@ def migrate_module_type(data: dict) -> dict:
     deprecated_from_version="0.9.6",
     removed_from_version="1.0.0",
     description="Converts moduleClass values from EntitySource to SampleStore naming (CSVEntitySource -> CSVSampleStore, SQLEntitySource -> SQLSampleStore)",
+    field_paths=["config.moduleClass"],
 )
 def migrate_module_class(data: dict) -> dict:
     """Convert moduleClass from EntitySource to SampleStore naming
 
-    This validator recursively searches for moduleClass fields within the config
-    and converts them from EntitySource to SampleStore naming. It operates only
-    within the config level, matching the original pydantic validator behavior.
+    This validator checks for moduleClass field within the config
+    and converts it from EntitySource to SampleStore naming.
 
     Old format:
         config:
@@ -96,31 +87,17 @@ def migrate_module_class(data: dict) -> dict:
     if not isinstance(data, dict):
         return data
 
-    # Only operate within config
-    if "config" not in data or not isinstance(data["config"], dict):
-        return data
-
     value_mappings = {
         "CSVEntitySource": "CSVSampleStore",
         "SQLEntitySource": "SQLSampleStore",
     }
 
-    def convert_module_class_in_dict(d: dict) -> None:
-        """Recursively convert moduleClass in nested structures"""
-        if "moduleClass" in d and d["moduleClass"] in value_mappings:
-            d["moduleClass"] = value_mappings[d["moduleClass"]]
+    # Check and update config.moduleClass
[…]
