+++ orchestrator/core/legacy/validators/discoveryspace/properties_field_removal.py
@@ -4,6 +4,7 @@
 """Legacy validator for removing deprecated properties field from discovery spaces"""
 
 from orchestrator.core.legacy.registry import legacy_validator
+from orchestrator.core.legacy.utils import remove_nested_field
 from orchestrator.core.resources import CoreResourceKinds
 
 
@@ -11,6 +12,7 @@ from orchestrator.core.resources import CoreResourceKinds
     identifier="discoveryspace_properties_field_removal",
     resource_type=CoreResourceKinds.DISCOVERYSPACE,
     deprecated_fields=["properties"],
+    field_paths=["config.properties"],
     deprecated_from_version="0.10.1",
     removed_from_version="1.0.0",
     description="Removes the deprecated 'properties' field from discovery space configurations",
@@ -36,13 +38,8 @@ def remove_properties_field(data: dict) -> dict:
     Returns:
         The migrated resource data dictionary
     """
-
-    if not isinstance(data, dict):
-        return data
-
-    # Only check in config (where the field actually exists)
-    if "config" in data and isinstance(data["config"], dict):
-        data["config"].pop("properties", None)
+    if isinstance(data, dict):
+        remove_nested_field(data, "config.properties")
 
     return data
 
