+++ orchestrator/core/legacy/validators/samplestore/v1_to_v2_csv_migration.py
@@ -4,6 +4,7 @@
 """Legacy validator for migrating CSV sample stores from v1 to v2 format"""
 
 from orchestrator.core.legacy.registry import legacy_validator
+from orchestrator.core.legacy.utils import has_nested_field
 from orchestrator.core.resources import CoreResourceKinds
 
 
@@ -13,27 +14,26 @@ from orchestrator.core.resources import CoreResourceKinds
     deprecated_fields=["constitutivePropertyColumns", "propertyMap"],
     deprecated_from_version="1.3.5",
     removed_from_version="1.6.0",
-    description="Migrates CSV sample stores from v1 format (constitutivePropertyColumns at top level) to v2 format (per-experiment constitutivePropertyMap)",
+    description="Migrates CSV sample stores from v1 format (constitutivePropertyColumns in config) to v2 format (per-experiment constitutivePropertyMap)",
+    field_paths=["config.constitutivePropertyColumns", "config.experiments"],
 )
 def migrate_csv_v1_to_v2(data: dict) -> dict:
     """Migrate old CSVSampleStoreDescription format to new format
 
-    This validator operates on the config level, migrating CSV sample store
-    configurations from v1 to v2 format. It matches the original pydantic
+    This validator operates on the config section of the CSV sample store,
+    migrating from v1 to v2 format. It matches the original pydantic
     validator behavior.
 
-    Old format:
-        config:
-            constitutivePropertyColumns: [...]  # at config level
-            experiments:
-              - propertyMap: {...}
+    Old format (in config):
+        constitutivePropertyColumns: [...]
+        experiments:
+          - propertyMap: {...}
 
-    New format:
-        config:
-            # No constitutivePropertyColumns at config level
-            experiments:
-              - observedPropertyMap: {...}
-                constitutivePropertyMap: [...]
+    New format (in config):
+        # No constitutivePropertyColumns
+        experiments:
+          - observedPropertyMap: {...}
+            constitutivePropertyMap: [...]
 
     Args:
         data: The resource data dictionary
@@ -45,14 +45,14 @@ def migrate_csv_v1_to_v2(data: dict) -> dict:
     if not isinstance(data, dict):
         return data
 
-    # Only operate within config
+    # Check if config exists
     if "config" not in data or not isinstance(data["config"], dict):
         return data
 
     config = data["config"]
 
     # Check if this is old format (has constitutivePropertyColumns in config)
-    if "constitutivePropertyColumns" not in config:
+    if not has_nested_field(data, "config.constitutivePropertyColumns"):
         return data
 
     # Extract and remove the constitutivePropertyColumns from config
