### 2026-03-24 a2d8707
feat(legacy): add support for validator dependencies
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
