+++ orchestrator/core/legacy/registry.py
@@ -79,6 +79,99 @@ class LegacyValidatorRegistry:
         """
         return list(cls._validators.values())
 
+    @classmethod
+    def resolve_dependencies(
+        cls, validator_ids: list[str]
+    ) -> tuple[list[str], list[str]]:
+        """Resolve validator dependencies and return ordered list
+
+        Uses topological sort to order validators based on their dependencies.
+        Detects circular dependencies. Automatically includes all transitive
+        dependencies.
+
+        Args:
+            validator_ids: List of validator identifiers to order
+
+        Returns:
+            Tuple of (ordered_validator_ids, missing_dependencies)
+            - ordered_validator_ids: Validators in dependency order (includes all dependencies)
+            - missing_dependencies: List of dependency IDs that don't exist
+
+        Raises:
+            ValueError: If circular dependencies are detected
+        """
+        # Build dependency graph - recursively add all dependencies
+        graph: dict[str, list[str]] = {}
+        in_degree: dict[str, int] = {}
+        missing_deps: set[str] = set()
+        to_process = list(validator_ids)
+        processed = set()
+
+        while to_process:
+            vid = to_process.pop(0)
+            if vid in processed:
+                continue
+            processed.add(vid)
+
+            validator = cls.get_validator(vid)
+            if validator is None:
+                continue
+
+            # Initialize this validator in the graph
+            if vid not in graph:
+                graph[vid] = []
+                in_degree[vid] = 0
+
+            # Process dependencies
+            for dep_id in validator.dependencies:
+                if dep_id not in cls._validators:
+                    missing_deps.add(dep_id)
+                    continue
+
+                # Add dependency to graph if not already there
+                if dep_id not in graph:
+                    graph[dep_id] = []
+                    in_degree[dep_id] = 0
+                    # Add to processing queue to handle transitive dependencies
+                    to_process.append(dep_id)
+
+                # Add edge from dependency to dependent
+                if vid not in graph[dep_id]:
+                    graph[dep_id].append(vid)
+
+        # Calculate in-degrees
+        for vid in graph:
+            validator = cls.get_validator(vid)
+            if validator:
+                for dep_id in validator.dependencies:
+                    if dep_id in graph:
+                        in_degree[vid] += 1
+
+        # Topological sort using Kahn's algorithm
+        queue = [vid for vid in graph if in_degree[vid] == 0]
+        ordered = []
+
+        while queue:
+            # Sort queue for deterministic ordering
+            queue.sort()
+            current = queue.pop(0)
+            ordered.append(current)
+
+            # Reduce in-degree for dependents
+            for dependent in graph[current]:
+                in_degree[dependent] -= 1
+                if in_degree[dependent] == 0:
+                    queue.append(dependent)
+
+        # Check for circular dependencies
+        if len(ordered) != len(graph):
+            remaining = [vid for vid in graph if vid not in ordered]
+            raise ValueError(
+                f"Circular dependency detected among validators: {', '.join(remaining)}"
+            )
+
+        return ordered, list(missing_deps)
+
 
 def legacy_validator(
     identifier: str,
@@ -88,6 +181,7 @@ def legacy_validator(
     removed_from_version: str,
     description: str,
     field_paths: list[str] | None = None,
+    dependencies: list[str] | None = None,
 ) -> Callable[[Callable[[dict], dict]], Callable[[dict], dict]]:
     """Decorator to register a legacy validator function
 
@@ -99,6 +193,7 @@ def legacy_validator(
         removed_from_version: ADO version when automatic upgrade was removed
         description: Human-readable description of what this validator does
         field_paths: Optional explicit paths to fields (e.g., 'config.properties')
[…]
