+++ tests/core/test_validator_dependencies.py
@@ -1,0 +1,378 @@
+# Copyright IBM Corporation 2025, 2026
+# SPDX-License-Identifier: MIT
+
+"""Tests for validator dependency resolution and ordering"""
+
+from collections.abc import Generator
+
+import pytest
+
+from orchestrator.core.legacy.metadata import LegacyValidatorMetadata
+from orchestrator.core.legacy.registry import LegacyValidatorRegistry
+from orchestrator.core.resources import CoreResourceKinds
+
+
+@pytest.fixture(autouse=True)
+def clear_registry() -> Generator[None, None, None]:
+    """Clear the registry before and after each test"""
+    LegacyValidatorRegistry._validators.clear()
+    yield
+    LegacyValidatorRegistry._validators.clear()
+
+
+def test_resolve_dependencies_no_dependencies() -> None:
+    """Test resolving validators with no dependencies"""
+
+    def validator_a(data: dict) -> dict:
+        return data
+
+    def validator_b(data: dict) -> dict:
+        return data
+
+    # Register validators without dependencies
+    LegacyValidatorRegistry.register(
+        LegacyValidatorMetadata(
+            identifier="validator_a",
+            resource_type=CoreResourceKinds.DISCOVERYSPACE,
+            deprecated_fields=["field_a"],
+            deprecated_from_version="1.0.0",
+            removed_from_version="2.0.0",
+            description="Validator A",
+            validator_function=validator_a,
+            field_paths=["config.field_a"],
+            dependencies=[],
+        )
+    )
+
+    LegacyValidatorRegistry.register(
+        LegacyValidatorMetadata(
+            identifier="validator_b",
+            resource_type=CoreResourceKinds.DISCOVERYSPACE,
+            deprecated_fields=["field_b"],
+            deprecated_from_version="1.0.0",
+            removed_from_version="2.0.0",
+            description="Validator B",
+            validator_function=validator_b,
+            field_paths=["config.field_b"],
+            dependencies=[],
+        )
+    )
+
+    # Resolve dependencies
+    ordered, missing = LegacyValidatorRegistry.resolve_dependencies(
+        ["validator_a", "validator_b"]
+    )
+
+    # Should return both validators in alphabetical order (no dependencies)
+    assert len(ordered) == 2
+    assert "validator_a" in ordered
+    assert "validator_b" in ordered
+    assert len(missing) == 0
+
+
+def test_resolve_dependencies_simple_chain() -> None:
+    """Test resolving validators with simple dependency chain"""
+
+    def validator_a(data: dict) -> dict:
+        return data
+
+    def validator_b(data: dict) -> dict:
+        return data
+
+    def validator_c(data: dict) -> dict:
+        return data
+
+    # Register validators: C depends on B, B depends on A
+    LegacyValidatorRegistry.register(
+        LegacyValidatorMetadata(
+            identifier="validator_a",
+            resource_type=CoreResourceKinds.DISCOVERYSPACE,
+            deprecated_fields=["field_a"],
+            deprecated_from_version="1.0.0",
+            removed_from_version="2.0.0",
+            description="Validator A",
+            validator_function=validator_a,
+            field_paths=["config.field_a"],
+            dependencies=[],
+        )
+    )
+
+    LegacyValidatorRegistry.register(
+        LegacyValidatorMetadata(
+            identifier="validator_b",
+            resource_type=CoreResourceKinds.DISCOVERYSPACE,
+            deprecated_fields=["field_b"],
+            deprecated_from_version="1.0.0",
+            removed_from_version="2.0.0",
+            description="Validator B",
+            validator_function=validator_b,
+            field_paths=["config.field_b"],
+            dependencies=["validator_a"],
+        )
+    )
+
+    LegacyValidatorRegistry.register(
[…]
