### 2026-03-24 a870c1f
docs(website): refine core concepts docs (#727)
* docs(concepts): refine core concepts docs

Introduce core-concepts/properties-and-domains.md.
Reframe concepts.md around what can be measured, what you want to measure, and the shared sample store.
Slim actuators.md by moving property/domain material to the new page and
dropping the long experiment listing example. Update and refine discovery-spaces,
entity-spaces, data-sharing, and working-with-actuators for consistency.

* docs(concepts): tidying

* docs(concepts): add sub-domains

* fix(docs): update link

* Apply suggestion from @michael-jo
[…]
