+++ website/docs/core-concepts/entity-spaces.md
@@ -1,27 +1,27 @@
 <!-- markdownlint-disable-next-line first-line-h1 -->
 ## Entities
 
-Entities represent things that can be measured. Examples are molecules or points
-in an application configuration space.
+Entities represent the things you want to measure — for example, a molecule,
+a fine-tuning deployment configuration, or a robotic experiment setup.
 
-Entities all have a set of constitutive properties which define them. A
-molecule's constitutive properties might be a SMILES or INCHI string. The
-constitutive properties of a fine-tuning deployment configuration might be GPU
-model, number of GPUs and batch size.
+Every Entity is described by a set of
+[**constitutive properties**](properties-and-domains.md#property-types), and
+corresponding values, that uniquely identify it. For a fine-tuning deployment
+configuration these
+might be GPU model, number of GPUs, and batch size. For a molecule they might
+be a SMILES string.
 
-An entity will also have observed properties. These are properties measured by
-an experiment (or experiment protocol). For example, a molecule might have an
-an observed property for its `band-gap` while a fine-tuning deployment
-configuration might have an an observed property related to `tokens throughput`.
+Once an Experiment has been run on an Entity, it also gains
+[**observed properties**](actuators.md#target-and-observed-properties) — the
+measured outputs produced by that Experiment.
 
-### Example: FM Fine-tuning Deployment Configuration
+### Example
 
-Here is an example of an entity that represents a FM fine-tuning deployment
+Here is an Entity representing a fine-tuning deployment configuration:
 
 <!-- markdownlint-disable line-length -->
 ```terminaloutput
 Identifier: dataset_id.news-tokens-16384plus-entries-4096-model_name.llama3-8b-number_gpus.4.0-model_max_length.2048.0-torch_dtype.bfloat16-batch_size.16.0-gpu_model.NVIDIA-A100-80GB-PCIe
-Generator: explicit_grid_sample_generator
 
 Constitutive properties:
                  name                               value
@@ -32,63 +32,39 @@ Constitutive properties:
   4       torch_dtype                            bfloat16
   5        batch_size                                16.0
   6         gpu_model               NVIDIA-A100-80GB-PCIe
-
-Observed properties:
-                                                   name                                         experiment                    target-property                values
-  0   finetune-lora-fsdp-r-4-a-16-tm-default-v1.2.0-...  SFTTrainer.finetune-lora-fsdp-r-4-a-16-tm-defa...        gpu_compute_utilization_min   [98.14772727272727]
-  1   finetune-lora-fsdp-r-4-a-16-tm-default-v1.2.0-...  SFTTrainer.finetune-lora-fsdp-r-4-a-16-tm-defa...        gpu_compute_utilization_avg   [98.26988636363636]
-  2   finetune-lora-fsdp-r-4-a-16-tm-default-v1.2.0-...  SFTTrainer.finetune-lora-fsdp-r-4-a-16-tm-defa...        gpu_compute_utilization_max   [98.38636363636364]
-  3   finetune-lora-fsdp-r-4-a-16-tm-default-v1.2.0-...  SFTTrainer.finetune-lora-fsdp-r-4-a-16-tm-defa...         gpu_memory_utilization_min  [33.709723284090906]
-  4   finetune-lora-fsdp-r-4-a-16-tm-default-v1.2.0-...  SFTTrainer.finetune-lora-fsdp-r-4-a-16-tm-defa...         gpu_memory_utilization_avg  [33.709723284090906]
[…]
