+++ website/docs/core-concepts/discovery-spaces.md
@@ -1,19 +1,22 @@
 <!-- markdownlint-disable-next-line first-line-h1 -->
-A Discovery Space is made up of an [`Entity Space`](entity-spaces.md) and a
-[`Measurement Space`](actuators.md#measurement-space). The `Entity Space`
-defines the things you want to measure and the `Measurement Space` how you want
-to measure them.
+A Discovery Space combines an [Entity Space](entity-spaces.md) and a
+[Measurement Space](actuators.md#measurement-space). The Entity Space defines
+the Entities you want to measure; the Measurement Space defines how they are
+measured. Results are stored in a [Sample Store](data-sharing.md).
 
-A Discovery Space is also associated with a [Sample Store](data-sharing.md)
-where measurement results and entities are recorded.
+A Discovery Space is a **view** rather than a container — data is fetched from
+the Sample Store on demand. This means multiple Discovery Spaces can share
+measurement results transparently, and any measurement made by anyone using the
+same Sample Store becomes immediately available.
 
-## Example: Fine-Tuning Deployment Configuration Discovery Space
+## Example: Fine-Tuning Deployment Configuration
 
 <!-- markdownlint-disable descriptive-link-text -->
-We can combine the Entity Space example for fine-tuning deployment configuration
-[here](entity-spaces.md#example-fine-tuning-deployment-configuration) with one
-of the experiments from the `SFTTrainer` actuator to create the following
-Discovery Space:
+We can combine the
+[Entity Space example](entity-spaces.md#example-fine-tuning-deployment-configuration)
+with one of the Experiments from the [`SFTTrainer` Actuator](../actuators/sft-trainer.md)
+to create the
+following Discovery Space:
 <!-- markdownlint-enable descriptive-link-text -->
 
 <!-- markdownlint-disable line-length -->
@@ -66,98 +69,78 @@ Sample Store identifier: '2351e8'
 ```
 <!-- markdownlint-enable line-length -->
 
-Here we can see:
+The output shows the unique Discovery Space identifier, the Entity Space (80
+Entities across 7 dimensions), and the Measurement Space (one Experiment with
+17 target properties). Together these define exactly what can be measured and
+what the resulting data will look like.
 
-- A unique id for the discovery space
-- The entity space
-- For each experiment in the measurement space (in this case just one) the
-  target properties it measures.
+## Measurement Space and Entity Space Compatibility
+
+Since an [Experiment](actuators.md#experiments) declares the inputs it needs,
+an Entity can only be measured by that Experiment if its
+[constitutive property](properties-and-domains.md#property-types) values
+satisfy those input requirements.
+
+Since a [Measurement Space](actuators.md#measurement-space) is a set of
+Experiments, it defines a set of required constitutive properties. An Entity
+Space must therefore contain all those properties, and each Entity Space
+[Property Domain](properties-and-domains.md#property-domain-types) must be a
+**subdomain** of the corresponding Experiment's input domain.
+
+In practice this means the Experiment's declared input domains define the
+**maximum possible extent** of any Entity Space used with that Measurement
+Space. Your Entity Space is always a focused subset within those bounds. For
+example, if an Experiment accepts `batch_size` values from 1 to 4096, your
+Entity Space can restrict that to `[1, 2, 4, 8, 16]` — but it cannot extend
+beyond `[1, 4096]`.
+
+| | Full Experiment extent | Focused Entity Space subset |
+| --- | --- | --- |
+| `batch_size` | `[1, 4097]` interval 1 | `[1, 2, 4, 8, 16, 32, 64, 128]` |
+| `model_name` | 40 model names | `[granite-3-8b, llama3-8b]` |
+| `number_gpus` | `[0, 33]` interval 1 | `[2, 4]` |
+
[…]
