+++ website/mkdocs.yml
@@ -158,8 +158,9 @@ nav:
       - Characterizing Spaces Without Prior Knowledge: examples/no-priors-characterization.md
   - Core Concepts:
       - core-concepts/concepts.md
+      - Properties and Domains: core-concepts/properties-and-domains.md
+      - Experiments & Actuators: core-concepts/actuators.md
       - Entities and Entity Spaces: core-concepts/entity-spaces.md
-      - Actuators, Experiments & Measurement Spaces: core-concepts/actuators.md
       - Discovery Spaces: core-concepts/discovery-spaces.md
       - Shared Sample Stores: core-concepts/data-sharing.md
   - Resources:
