### 2026-03-24 8b9aa5f
feat(legacy): support suggesting validator dependencies
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
