### 2026-03-24 193c1a6
feat(cli): support auto-handling validator dependencies
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
