### 2026-03-24 c98c936
docs(website): add docs for legacy validators
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
