+++ website/docs/cli/legacy-validators.md
@@ -1,0 +1,393 @@
+# Legacy Validators
+
+Legacy validators help you upgrade old resource files that use deprecated fields
+or formats. This guide covers the legacy validator system and its advanced
+features.
+
+## Overview
+
+The legacy validator system provides:
+
+- **Automatic dependency resolution**: Validators can depend on other validators
+- **Smart error messages**: Get helpful suggestions when validation fails
+- **Progress tracking**: Visual feedback during upgrade operations
+- **Automatic ordering**: Validators run in the correct order automatically
+- **Dry-run mode**: Preview changes before applying them
+
+## Quick Start
+
+### List Available Validators
+
+View all legacy validators:
+
+```bash
+ado legacy list
+```
+
+View validators for a specific resource type:
+
+```bash
+ado legacy list samplestore
+```
+
+### Get Validator Information
+
+Get detailed information about a specific validator:
+
+```bash
+ado legacy info discoveryspace_properties_field_removal
+```
+
+This shows:
+
+- Description of what the validator does
+- Which deprecated fields it handles
+- Version information
+- Usage examples
+
+### Upgrade Resources
+
+#### Upgrade Resources in Metastore
+
+Apply a legacy validator to resources in your metastore:
+
+```bash
+ado upgrade discoveryspace --apply-legacy-validator discoveryspace_properties_field_removal
+```
+
+Apply multiple validators (they will be automatically ordered):
+
+```bash
+ado upgrade samplestore \
+  --apply-legacy-validator samplestore_kind_entitysource_to_samplestore \
+  --apply-legacy-validator samplestore_module_type_entitysource_to_samplestore
+```
+
+#### Upgrade Local YAML Files
+
+Upgrade local YAML files without loading them into the metastore:
+
+```bash
+ado legacy upgrade --file my-resource.yaml \
+  --apply-legacy-validator discoveryspace_properties_field_removal
+```
+
+Upgrade multiple files:
+
+```bash
+ado legacy upgrade \
+  --file resource1.yaml \
+  --file resource2.yaml \
+  --apply-legacy-validator validator_name
+```
+
+## Advanced Features
+
+### Automatic Dependency Resolution
+
+Validators can depend on other validators. When you specify a validator, the
+system automatically:
+
+1. Includes all required dependencies
+2. Orders validators correctly using topological sort
+3. Notifies you when dependencies are auto-included
+
+**Example:**
+
+If validator B depends on validator A, you only need to specify B:
+
+```bash
+ado upgrade samplestore --apply-legacy-validator validator_b
+```
+
+The system will automatically:
+
+- Include validator A
+- Run A before B
+- Show you: "Auto-included dependencies: validator_a"
+
+### Circular Dependency Detection
+
+The system detects circular dependencies and provides clear error messages:
+
+```text
+Error: Circular dependency detected among validators: validator_a, validator_b
+```
+
+### Enhanced Error Messages
+
+When validation fails, you get:
+
+1. **Detailed field errors**: Exact fields that failed and why
+2. **Applicable validators**: List of validators that can fix the issues
+3. **Dependency information**: Which validators depend on others
+4. **Ready-to-use commands**: Complete commands you can copy-paste
+
+**Example error output:**
+
+```text
+Validation Error in discoveryspace 'my-space'
+
+Fields with validation errors: 2 field(s)
+
+Error details:
+  • config.properties:
[…]
