### 2026-03-24 72e11f8
fix(vllm_performance): set correct experiment identifiers and property configuration to test existing vllm geospatial endpoint
Signed-off-by: Michele Gazzetti <michele.gazzetti1@ibm.com>
