### 2026-03-24 457a2a9
fix(vllm_performance): update default values to benchmark geospatial models
Signed-off-by: Michele Gazzetti <michele.gazzetti1@ibm.com>
