### 2026-03-24 dc64909
test: fix tests
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
