+++ plugins/operators/trim/trim/tests/test_high_dimensional_sampling.py
@@ -16,7 +16,7 @@ import pytest
 from no_priors_characterization.utils.high_dimensional_sampling import (
     concatenated_latin_hypercube_sampling,
 )
-from trim.test.test_data_documentation import TEST_DATAFRAMES
+from trim.tests.test_data_documentation import TEST_DATAFRAMES
 
 
 class TestConcatenatedLatinHypercubeSampling:
