+++ tox.ini
@@ -89,7 +89,7 @@ commands=
     pytest -n auto --dist worksteal -rx -vv --log-level=INFO --color=yes plugins/custom_experiments/autoconf/autoconf/test/
 
     # TRIM operator tests
-    pytest -n auto --dist worksteal -rx -vv --log-level=INFO --color=yes plugins/operators/trim/trim/test
+    pytest -n auto --dist worksteal -rx -vv --log-level=INFO --color=yes plugins/operators/trim/trim/tests
 
     # Ray Tune operator functionality tests
     ado get operator ray_tune
