### 2026-03-25 5796ca8
build(trim): use src layout (#741)
* build(trim): use src layout

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* fix(trim): update pyproject with autodiscovery

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* fix(test): update trim tests location

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

---------

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
