+++ plugins/operators/trim/pyproject.toml
@@ -20,8 +20,9 @@ trim = "trim.operator"
 requires = ["setuptools", "setuptools_scm"]
 build-backend = "setuptools.build_meta"
 
-[tool.setuptools]
-packages = ["trim"]
+[tool.setuptools.packages.find]
+where = ["src"]
+include = ["trim*"]
 
 [tool.setuptools_scm]
 root = "../../../"
