### 2026-03-25 bc22fb1
refactor(cli): use full paths for matching validators
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
