### 2026-03-25 7cdc214
refactor(core): only support fully qualified field names in legacy validator registry
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
