### 2026-03-25 419f411
refactor(core): rename field
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
