+++ orchestrator/core/legacy/validators/discoveryspace/entitysource_to_samplestore.py
@@ -15,7 +15,7 @@ from orchestrator.core.resources import CoreResourceKinds
 @legacy_validator(
     identifier="discoveryspace_entitysource_to_samplestore",
     resource_type=CoreResourceKinds.DISCOVERYSPACE,
-    fully_qualified_deprecated_field_paths=["config.entitySourceIdentifier"],
+    deprecated_field_paths=["config.entitySourceIdentifier"],
     deprecated_from_version="0.9.6",
     removed_from_version="1.0.0",
     description="Renames 'entitySourceIdentifier' to 'sampleStoreIdentifier' in discovery space configurations",
