### 2026-03-25 2709eb6
refactor(cli): remove unused function
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
