### 2026-03-25 19fa9a9
refactor(cli): remove print_validator_suggestions
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
