### 2026-03-25 189fee2
refactor(cli): consolidate duplicate functions into one
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
