### 2026-03-25 1c44854
refactor: style improvements
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
