+++ orchestrator/cli/utils/legacy/common.py
@@ -8,104 +8,18 @@ from typing import TYPE_CHECKING
 import pydantic
 from rich.console import Console
 
+from orchestrator.cli.utils.output.prints import (
+    ERROR,
+    HINT,
+    INFO,
+    WARN,
+    console_print,
+    cyan,
+)
+
 if TYPE_CHECKING:
     from orchestrator.core.legacy.metadata import LegacyValidatorMetadata
     from orchestrator.core.resources import CoreResourceKinds
-
-
-def print_validator_suggestions_with_dependencies(
-    validators: list["LegacyValidatorMetadata"],
-    resource_type: "CoreResourceKinds",
-    console: Console,
-) -> None:
-    """Print legacy validator suggestions with dependency information
-
-    This enhanced version resolves dependencies and shows validators in the
-    correct execution order, along with dependency information.
-
-    Args:
-        validators: List of applicable validators
-        resource_type: The resource type
-        console: Rich console to print to
-    """
-    from orchestrator.core.legacy.registry import LegacyValidatorRegistry
-
-    # Resources can be referenced by their CoreResourceKinds value or by shorthands
-    resource_cli_name = resource_type.value
-
-    # Get validator identifiers
-    validator_ids = [v.identifier for v in validators]
-
-    # Resolve dependencies to get correct order
-    try:
-        ordered_ids, missing_deps = LegacyValidatorRegistry.resolve_dependencies(
-            validator_ids
-        )
-    except ValueError as e:
-        # Circular dependency detected
-        console.print(
-            f"\n[bold red]Warning:[/bold red] {e}",
-            style="red",
-        )
-        # Fall back to original order
-        ordered_ids = validator_ids
-        missing_deps = []
-
-    # Get ordered validators (filter out None values)
-    ordered_validators: list[LegacyValidatorMetadata] = []
-    for vid in ordered_ids:
-        validator = LegacyValidatorRegistry.get_validator(vid)
-        if validator is not None:
-            ordered_validators.append(validator)
-
-    console.print("\n[bold cyan]Available legacy validators:[/bold cyan]\n")
-
-    for i, validator in enumerate(ordered_validators, 1):
-        # Show execution order
-        console.print(f"  {i}. [green]{validator.identifier}[/green]")
-        console.print(f"     {validator.description}")
-        console.print(f"     Handles: {', '.join(validator.deprecated_field_paths)}")
-        console.print(f"     Deprecated: v{validator.deprecated_from_version}")
-
-        # Show dependencies if any
-        if validator.dependencies:
-            dep_names = []
-            for dep_id in validator.dependencies:
-                dep_validator = LegacyValidatorRegistry.get_validator(dep_id)
-                if dep_validator:
-                    dep_names.append(dep_validator.identifier)
-                else:
-                    dep_names.append(f"{dep_id} [red](missing)[/red]")
-            console.print(f"     Dependencies: {', '.join(dep_names)}")
-
-        console.print()
-
-    # Warn about missing dependencies
-    if missing_deps:
-        console.print(
-            f"[bold yellow]Warning:[/bold yellow] Some dependencies are missing: {', '.join(missing_deps)}\n"
-        )
-
-    console.print("[bold magenta]To upgrade using legacy validators:[/bold magenta]")
-
-    # Build command with all validators in correct order
-    validator_args = " ".join(
-        f"--apply-legacy-validator {v.identifier}" for v in ordered_validators
-    )
-    console.print(f"  ado upgrade {resource_cli_name} {validator_args}")
-    console.print()
-
-    # Show note about automatic dependency resolution
-    if len(ordered_validators) > len(validators):
-        console.print(
-            "[dim]Note: Additional validators were included to satisfy dependencies[/dim]\n"
-        )
-
[…]
