### 2026-03-25 7be7396
refactor(cli): consolidate prints
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
