### 2026-03-25 fbede11
refactor: remove useless parameter
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
