### 2026-03-25 a4fc4c0
refactor(cli): reorder upgrade handler
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
