### 2026-03-25 1aac034
refactor: improve readability
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
