### 2026-03-25 e65336c
refactor: update function name
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
