### 2026-03-25 5778cb4
test: avoid mocks
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
