+++ tests/core/test_legacy_validators.py
@@ -3,13 +3,14 @@
 
 """Integration tests for legacy validators with pydantic models and upgrade process"""
 
-from unittest.mock import MagicMock, patch
+from collections.abc import Callable
+from pathlib import Path
 
 import pydantic
-import pytest
 
 from orchestrator.core.legacy.registry import LegacyValidatorRegistry, legacy_validator
 from orchestrator.core.resources import CoreResourceKinds
+from orchestrator.metastore.project import ProjectContext
 
 
 class TestLegacyValidatorWithPydantic:
@@ -195,16 +196,46 @@ class TestLegacyValidatorWithPydantic:
 
 
 class TestUpgradeHandlerIntegration:
-    """Test the upgrade handler with legacy validators"""
+    """Integration tests for ado upgrade with legacy validators via CLI"""
 
     def setup_method(self) -> None:
-        """Clear the registry before each test"""
-        LegacyValidatorRegistry._validators = {}
+        """Setup - ensure validators are registered"""
+        import orchestrator.core.legacy.validators  # noqa: F401
 
-    def test_upgrade_handler_applies_legacy_validator(self) -> None:
-        """Test that handle_ado_upgrade applies legacy validators correctly"""
+        if not hasattr(self.__class__, "_initial_validators"):
+            self.__class__._initial_validators = (
+                LegacyValidatorRegistry._validators.copy()
+            )
+        elif not LegacyValidatorRegistry._validators:
+            LegacyValidatorRegistry._validators = (
+                self.__class__._initial_validators.copy()
+            )
 
-        # Register a test validator
+    def test_upgrade_applies_legacy_validator_via_cli(
+        self,
+        tmp_path: Path,
+        valid_ado_project_context: ProjectContext,
+        create_active_ado_context: Callable,
+    ) -> None:
+        """Test that ado upgrade applies legacy validators correctly via CLI"""
+
+        from typer.testing import CliRunner
+
+        runner = CliRunner()
+
+        from orchestrator.cli.core.cli import app as ado
+        from orchestrator.cli.utils.generic.wrappers import get_sql_store
+        from orchestrator.core.samplestore.config import (
+            SampleStoreConfiguration,
+            SampleStoreModuleConf,
+            SampleStoreSpecification,
+        )
+        from orchestrator.core.samplestore.resource import SampleStoreResource
+
+        # Step 1: Setup active context
+        create_active_ado_context(runner, tmp_path, valid_ado_project_context)
+
+        # Step 2: Register a test validator
         @legacy_validator(
             identifier="test_upgrade_validator",
             resource_type=CoreResourceKinds.SAMPLESTORE,
@@ -214,156 +245,226 @@ class TestUpgradeHandlerIntegration:
             description="Test upgrade validator",
         )
         def test_validator(data: dict) -> dict:
-            if "old_field" in data:
-                data["new_field"] = data.pop("old_field")
+            """Migrate old_field to new_field"""
+            if "config" in data and "old_field" in data["config"]:
+                data["config"]["new_field"] = data["config"].pop("old_field")
             return data
 
-        # Create a mock resource class with model_validate
-        mock_resource_class = MagicMock()
-        mock_validated_resource = MagicMock()
-        mock_resource_class.model_validate.return_value = mock_validated_resource
-
-        mock_sql_store = MagicMock()
-        # Mock getResourceIdentifiersOfKind to return identifiers
-        mock_sql_store.getResourceIdentifiersOfKind.return_value = {
-            "IDENTIFIER": ["res1"]
-        }
-        # Mock getResourceRaw to return raw dict data
-        mock_sql_store.getResourceRaw.return_value = {"old_field": "test_value"}
-
-        # Mock parameters
[…]
