### 2026-03-25 dd28ea0
refactor(tests): use fixture for common metadata pattern
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
