+++ tests/core/test_legacy_registry.py
@@ -3,6 +3,10 @@
 
 """Unit tests for the legacy validator registry"""
 
+from collections.abc import Callable
+
+import pytest
+
 from orchestrator.core.legacy.metadata import LegacyValidatorMetadata
 from orchestrator.core.legacy.registry import (
     LegacyValidatorRegistry,
@@ -11,23 +15,65 @@ from orchestrator.core.legacy.registry import (
 from orchestrator.core.resources import CoreResourceKinds
 
 
+@pytest.fixture
+def dummy_validator() -> Callable[[dict], dict]:
+    """Fixture providing a simple dummy validator function"""
+
+    def validator(data: dict) -> dict:
+        return data
+
+    return validator
+
+
+@pytest.fixture
+def create_validator_metadata(
+    dummy_validator: Callable[[dict], dict],
+) -> Callable[..., LegacyValidatorMetadata]:
+    """Fixture factory for creating LegacyValidatorMetadata instances"""
+
+    def _create_metadata(
+        identifier: str = "test_validator",
+        resource_type: CoreResourceKinds = CoreResourceKinds.SAMPLESTORE,
+        deprecated_field_paths: list[str] | None = None,
+        deprecated_from_version: str = "1.0.0",
+        removed_from_version: str = "2.0.0",
+        description: str = "Test validator",
+        validator_function: Callable[[dict], dict] | None = None,
+        dependencies: list[str] | None = None,
+    ) -> LegacyValidatorMetadata:
+        if deprecated_field_paths is None:
+            deprecated_field_paths = ["config.field1"]
+        if validator_function is None:
+            validator_function = dummy_validator
+        if dependencies is None:
+            dependencies = []
+
+        return LegacyValidatorMetadata(
+            identifier=identifier,
+            resource_type=resource_type,
+            deprecated_field_paths=deprecated_field_paths,
+            deprecated_from_version=deprecated_from_version,
+            removed_from_version=removed_from_version,
+            description=description,
+            validator_function=validator_function,
+            dependencies=dependencies,
+        )
+
+    return _create_metadata
+
+
 class TestLegacyValidatorMetadata:
     """Test the LegacyValidatorMetadata model"""
 
-    def test_create_metadata(self) -> None:
+    def test_create_metadata(
+        self,
+        create_validator_metadata: Callable[..., LegacyValidatorMetadata],
+        dummy_validator: Callable[[dict], dict],
+    ) -> None:
         """Test creating validator metadata"""
-
-        def dummy_validator(data: dict) -> dict:
-            return data
-
-        metadata = LegacyValidatorMetadata(
+        metadata = create_validator_metadata(
             identifier="test_validator",
-            resource_type=CoreResourceKinds.SAMPLESTORE,
             deprecated_field_paths=["config.field1", "config.field2"],
-            deprecated_from_version="1.0.0",
-            removed_from_version="2.0.0",
-            description="Test validator",
-            validator_function=dummy_validator,
         )
 
         assert metadata.identifier == "test_validator"
@@ -41,21 +87,11 @@ class TestLegacyValidatorMetadata:
         assert metadata.description == "Test validator"
         assert metadata.validator_function == dummy_validator
 
-    def test_metadata_serialization(self) -> None:
+    def test_metadata_serialization(
+        self, create_validator_metadata: Callable[..., LegacyValidatorMetadata]
+    ) -> None:
         """Test that validator function is excluded from serialization"""
-
-        def dummy_validator(data: dict) -> dict:
-            return data
-
-        metadata = LegacyValidatorMetadata(
-            identifier="test_validator",
-            resource_type=CoreResourceKinds.SAMPLESTORE,
-            deprecated_field_paths=["config.field1"],
-            deprecated_from_version="1.0.0",
-            removed_from_version="2.0.0",
[…]
