### 2026-03-25 c8cdb92
refactor(tests): remove tests that aren't needed anymore
We don't match on leaf nodes anymore

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
