+++ tests/core/test_validator_scope_fixes.py
@@ -1,367 +1,0 @@
-# Copyright IBM Corporation 2025, 2026
-# SPDX-License-Identifier: MIT
-
-"""Tests for Phase 1 validator scope fixes - verifying validators only operate on config level"""
-
-import importlib
-
-from orchestrator.core.legacy.registry import LegacyValidatorRegistry
-
-
-class TestValidatorScopeFixes:
-    """Test that validators correctly operate only on config level after Phase 1 fixes"""
-
-    def setup_method(self) -> None:
-        """Clear registry and reload validators before each test for isolation"""
-        # Clear the registry to ensure clean state
-        LegacyValidatorRegistry._validators = {}
-        # Reload all validator module files to re-register them
-        import orchestrator.core.legacy.validators.discoveryspace.entitysource_to_samplestore
-        import orchestrator.core.legacy.validators.discoveryspace.properties_field_removal
-        import orchestrator.core.legacy.validators.operation.actuators_field_removal
-        import orchestrator.core.legacy.validators.resource.entitysource_to_samplestore
-        import orchestrator.core.legacy.validators.samplestore.entitysource_migrations
-        import orchestrator.core.legacy.validators.samplestore.v1_to_v2_csv_migration
-
-        importlib.reload(
-            orchestrator.core.legacy.validators.discoveryspace.properties_field_removal
-        )
-        importlib.reload(
-            orchestrator.core.legacy.validators.discoveryspace.entitysource_to_samplestore
-        )
-        importlib.reload(
-            orchestrator.core.legacy.validators.operation.actuators_field_removal
-        )
-        importlib.reload(
-            orchestrator.core.legacy.validators.resource.entitysource_to_samplestore
-        )
-        importlib.reload(
-            orchestrator.core.legacy.validators.samplestore.v1_to_v2_csv_migration
-        )
-        importlib.reload(
-            orchestrator.core.legacy.validators.samplestore.entitysource_migrations
-        )
-
-    def test_discoveryspace_properties_removal_scope(self) -> None:
-        """Verify properties_field_removal only modifies config, not resource level"""
-
-        # Test data with 'properties' at both levels
-        resource_data = {
-            "kind": "discoveryspace",
-            "identifier": "test-space",
-            "properties": "SHOULD_NOT_BE_REMOVED",  # Resource level
-            "config": {
-                "properties": ["prop1", "prop2"],  # Config level (should be removed)
-                "sampleStoreIdentifier": "store-1",
-            },
-        }
-
-        # Get validator
-        validator = LegacyValidatorRegistry.get_validator(
-            "discoveryspace_properties_field_removal"
-        )
-        assert validator is not None
-
-        # Apply validator
-        result = validator.validator_function(resource_data.copy())
-
-        # Verify: resource-level field unchanged, config-level field removed
-        assert "properties" in result  # Resource level preserved
-        assert result["properties"] == "SHOULD_NOT_BE_REMOVED"
-        assert "properties" not in result["config"]  # Config level removed
-        assert result["config"]["sampleStoreIdentifier"] == "store-1"
-
-    def test_discoveryspace_entitysource_migration_scope(self) -> None:
-        """Verify entitysource_to_samplestore only modifies config, not resource level"""
-
-        # Test data with entitySourceIdentifier at both levels
-        resource_data = {
-            "kind": "discoveryspace",
-            "identifier": "test-space",
-            "entitySourceIdentifier": "SHOULD_NOT_BE_REMOVED",  # Resource level
-            "config": {
-                "entitySourceIdentifier": "old-source",  # Config level (should migrate)
-            },
-        }
-
-        # Get validator
-        validator = LegacyValidatorRegistry.get_validator(
-            "discoveryspace_entitysource_to_samplestore"
-        )
-        assert validator is not None
-
-        # Apply validator
[…]
