### 2026-03-25 c168a99
refactor(test): avoid mocks
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
