+++ tests/core/test_upgrade_transaction_safety.py
@@ -1,15 +1,27 @@
 # Copyright IBM Corporation 2025, 2026
 # SPDX-License-Identifier: MIT
 
-"""Tests for Phase 1 transaction safety in upgrade handler"""
+"""Integration tests for Phase 1 transaction safety in upgrade handler"""
 
-from unittest.mock import MagicMock, patch
+import json
 
 import pytest
+import sqlalchemy
 import typer
 
+from orchestrator.cli.core.config import AdoConfiguration
+from orchestrator.cli.models.parameters import AdoUpgradeCommandParameters
+from orchestrator.cli.utils.generic.wrappers import get_sql_store
+from orchestrator.cli.utils.resources.handlers import handle_ado_upgrade
 from orchestrator.core.legacy.registry import LegacyValidatorRegistry, legacy_validator
 from orchestrator.core.resources import CoreResourceKinds
+from orchestrator.core.samplestore.config import (
+    SampleStoreConfiguration,
+    SampleStoreModuleConf,
+    SampleStoreSpecification,
+)
+from orchestrator.core.samplestore.resource import SampleStoreResource
+from orchestrator.metastore.project import ProjectContext
 
 
 class TestUpgradeTransactionSafety:
@@ -19,214 +31,243 @@ class TestUpgradeTransactionSafety:
         """Clear the registry before each test"""
         LegacyValidatorRegistry._validators = {}
 
-    def test_all_resources_validated_before_any_saved(self) -> None:
+    @pytest.mark.parametrize("valid_ado_project_context", ["sqlite"], indirect=True)
+    def test_all_resources_validated_before_any_saved(
+        self,
+        valid_ado_project_context: ProjectContext,
+    ) -> None:
         """Test that all resources are validated before any are saved"""
 
-        # Register a test validator
+        # Register a test validator that transforms old_field -> new_field
         @legacy_validator(
             identifier="test_transaction_validator",
             resource_type=CoreResourceKinds.SAMPLESTORE,
-            deprecated_field_paths=["config.old_field"],
+            deprecated_field_paths=["config.metadata.old_field"],
             deprecated_from_version="1.0.0",
             removed_from_version="2.0.0",
             description="Test transaction validator",
         )
         def test_validator(data: dict) -> dict:
-            if "config" in data and "old_field" in data["config"]:
-                data["config"]["new_field"] = data["config"].pop("old_field")
+            if "config" in data and "metadata" in data["config"]:
+                metadata = data["config"]["metadata"]
+                if "old_field" in metadata:
+                    metadata["new_field"] = metadata.pop("old_field")
             return data
 
-        # Create mock resources
-        mock_resource1 = MagicMock()
-        mock_resource1.model_dump.return_value = {
-            "kind": "samplestore",
-            "identifier": "res1",
-            "config": {"old_field": "value1"},
-        }
-
-        mock_resource2 = MagicMock()
-        mock_resource2.model_dump.return_value = {
-            "kind": "samplestore",
-            "identifier": "res2",
-            "config": {"old_field": "value2"},
-        }
-
-        # Mock resource class
-        mock_resource_class = MagicMock()
-        validated_resources = []
-
-        def mock_validate(data: dict) -> MagicMock:
-            validated = MagicMock()
-            validated.model_dump.return_value = data
-            validated_resources.append(data["identifier"])
-            return validated
-
-        mock_resource_class.model_validate.side_effect = mock_validate
-
-        # Mock SQL store
-        mock_sql_store = MagicMock()
-        mock_sql_store.getResourcesOfKind.return_value = {
-            "res1": mock_resource1,
-            "res2": mock_resource2,
-        }
[…]
