### 2026-03-25 4919a63
test: add isolated legacy validator registry fixtures
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
