### 2026-03-25 2e81c54
fix(tests): update test
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
