### 2026-03-25 44e2866
fix: use correct paths for deprecated sample store fields
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
