+++ orchestrator/core/legacy/validators/samplestore/entitysource_migrations.py
@@ -15,7 +15,7 @@ from orchestrator.core.resources import CoreResourceKinds
 @legacy_validator(
     identifier="samplestore_module_type_entitysource_to_samplestore",
     resource_type=CoreResourceKinds.SAMPLESTORE,
-    deprecated_field_paths=["config.moduleType"],
+    deprecated_field_paths=["config.specification.module.moduleType"],
     deprecated_from_version="0.9.6",
     removed_from_version="1.0.0",
     description="Converts moduleType value from 'entity_source' to 'sample_store'",
@@ -23,16 +23,20 @@ from orchestrator.core.resources import CoreResourceKinds
 def migrate_module_type(data: dict) -> dict:
     """Convert moduleType from entity_source to sample_store
 
-    This validator checks for moduleType field within the config
+    This validator checks for moduleType field within the config.specification.module
     and converts it from 'entity_source' to 'sample_store'.
 
     Old format:
         config:
-            moduleType: "entity_source"
+            specification:
+                module:
+                    moduleType: "entity_source"
 
     New format:
         config:
-            moduleType: "sample_store"
+            specification:
+                module:
+                    moduleType: "sample_store"
 
     Args:
         data: The resource data dictionary
@@ -44,11 +48,13 @@ def migrate_module_type(data: dict) -> dict:
     if not isinstance(data, dict):
         return data
 
-    # Check and update config.moduleType
-    if has_nested_field(data, "config.moduleType"):
-        module_type = get_nested_value(data, "config.moduleType")
+    # Check and update config.specification.module.moduleType
+    if has_nested_field(data, "config.specification.module.moduleType"):
+        module_type = get_nested_value(data, "config.specification.module.moduleType")
         if module_type == "entity_source":
-            set_nested_value(data, "config.moduleType", "sample_store")
+            set_nested_value(
+                data, "config.specification.module.moduleType", "sample_store"
+            )
 
     return data
 
@@ -56,7 +62,7 @@ def migrate_module_type(data: dict) -> dict:
 @legacy_validator(
     identifier="samplestore_module_class_entitysource_to_samplestore",
     resource_type=CoreResourceKinds.SAMPLESTORE,
-    deprecated_field_paths=["config.moduleClass"],
+    deprecated_field_paths=["config.specification.module.moduleClass"],
     deprecated_from_version="0.9.6",
     removed_from_version="1.0.0",
     description="Converts moduleClass values from EntitySource to SampleStore naming (CSVEntitySource -> CSVSampleStore, SQLEntitySource -> SQLSampleStore)",
@@ -64,16 +70,20 @@ def migrate_module_type(data: dict) -> dict:
 def migrate_module_class(data: dict) -> dict:
     """Convert moduleClass from EntitySource to SampleStore naming
 
-    This validator checks for moduleClass field within the config
+    This validator checks for moduleClass field within the config.specification.module
     and converts it from EntitySource to SampleStore naming.
 
     Old format:
         config:
-            moduleClass: "CSVEntitySource" or "SQLEntitySource"
+            specification:
+                module:
+                    moduleClass: "CSVEntitySource" or "SQLEntitySource"
 
     New format:
         config:
-            moduleClass: "CSVSampleStore" or "SQLSampleStore"
+            specification:
+                module:
+                    moduleClass: "CSVSampleStore" or "SQLSampleStore"
 
     Args:
         data: The resource data dictionary
@@ -90,11 +100,15 @@ def migrate_module_class(data: dict) -> dict:
         "SQLEntitySource": "SQLSampleStore",
     }
 
-    # Check and update config.moduleClass
-    if has_nested_field(data, "config.moduleClass"):
-        module_class = get_nested_value(data, "config.moduleClass")
+    # Check and update config.specification.module.moduleClass
[…]
