### 2026-03-25 #10178 @matt-graham
Hi @Torigh. Thanks for offering to review this submission! We aim for reviews to be completed typically within a 4 - 6 week window and it is likely there will need to be some back and forth discussion between the reviewers and authors over this process. Just wanting to confirm you will be available over this sort of length of period to take part in the review process in case your message means you would only have availability today.
