### 2026-03-26 7d2e4d3
fix: restore migrate_module_name's original behaviour
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
