### 2026-03-26 #10304 @editorialbot
[REVIEW]: ado: a Python framework for computational experimentation and benchmarking
**Submitting author:** @AlessandroPomponio (<a href="http://orcid.org/0000-0003-1655-7500">Alessandro Pomponio</a>)
**Repository:** https://github.com/IBM/ado
**Branch with paper.md** (empty if default branch): maj_paper
**Version:** 1.6.0
