### 2026-03-26 7072d18
fix: add deprecated fields to simulate field_validator behaviour
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
