### 2026-03-26 29de01e
feat(legacy): add support for gt4sd migration
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
