+++ tests/core/legacy/validators/samplestore/test_gt4sd_transformer_migration.py
@@ -1,0 +1,289 @@
+# Copyright IBM Corporation 2025, 2026
+# SPDX-License-Identifier: MIT
+
+"""Tests for GT4SDTransformer to CSVSampleStore migration validator"""
+
+from orchestrator.core.legacy.registry import LegacyValidatorRegistry
+
+
+class TestGT4SDTransformerMigration:
+    """Test migrate_gt4sd_transformer_to_csv validator"""
+
+    def test_migrates_gt4sd_transformer_to_csv_sample_store(
+        self, legacy_validators_loaded: None
+    ) -> None:
+        """Test that GT4SDTransformer is migrated to CSVSampleStore with explicit parameters"""
+        validator = LegacyValidatorRegistry.get_validator(
+            "samplestore_gt4sd_transformer_to_csv"
+        )
+        assert validator is not None
+
+        data = {
+            "config": {
+                "copyFrom": [
+                    {
+                        "module": {
+                            "moduleClass": "GT4SDTransformer",
+                            "moduleName": "orchestrator.plugins.samplestores.gt4sd",
+                        },
+                        "storageLocation": {
+                            "path": "data/GM_Comparison/Transfromer/Sample_0/test_generations.csv"
+                        },
+                        "parameters": {
+                            "generatorIdentifier": "gt4sd-pfas-transformer-model-one"
+                        },
+                    }
+                ]
+            }
+        }
+
+        result = validator.validator_function(data)
+
+        # Check module class and name were updated
+        copy_from = result["config"]["copyFrom"][0]
+        assert copy_from["module"]["moduleClass"] == "CSVSampleStore"
+        assert copy_from["module"]["moduleName"] == "orchestrator.core.samplestore.csv"
+
+        # Check identifierColumn was added
+        assert copy_from["parameters"]["identifierColumn"] == "smiles"
+
+        # Check experiments configuration was added
+        assert "experiments" in copy_from["parameters"]
+        experiments = copy_from["parameters"]["experiments"]
+        assert len(experiments) == 1
+        assert (
+            experiments[0]["experimentIdentifier"]
+            == "transformer-toxicity-inference-experiment"
+        )
+        assert "observedPropertyMap" in experiments[0]
+        assert "constitutivePropertyMap" in experiments[0]
+        assert experiments[0]["constitutivePropertyMap"] == ["smiles"]
+
+        # Check property map was correctly added
+        property_map = experiments[0]["observedPropertyMap"]
+        assert property_map["logws"] == "GenLogws"
+        assert property_map["logd"] == "GenLogd"
+        assert property_map["loghl"] == "GenLoghl"
+        assert property_map["pka"] == "GenPka"
+        assert property_map["biodegradation halflife"] == "GenBiodeg"
+        assert property_map["bcf"] == "GenBcf"
+        assert property_map["ld50"] == "GenLd50"
+        assert property_map["scscore"] == "GenScscore"
+
+        # Check original generatorIdentifier was preserved
+        assert (
+            copy_from["parameters"]["generatorIdentifier"]
+            == "gt4sd-pfas-transformer-model-one"
+        )
+
+    def test_preserves_existing_identifier_column(
+        self, legacy_validators_loaded: None
+    ) -> None:
+        """Test that existing identifierColumn is not overwritten"""
+        validator = LegacyValidatorRegistry.get_validator(
+            "samplestore_gt4sd_transformer_to_csv"
+        )
+        assert validator is not None
+
+        data = {
+            "config": {
+                "copyFrom": [
+                    {
+                        "module": {
+                            "moduleClass": "GT4SDTransformer",
+                            "moduleName": "orchestrator.plugins.samplestores.gt4sd",
+                        },
+                        "parameters": {
[…]
