### 2026-03-26 f5d43b6
feat(legacy): add support for removing storageLocation
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
