### 2026-03-26 9bf2219
refactor(cli): replace error print with info
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
