### 2026-03-26 5a54b5e
test: force mysql on tests due to CI issues
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
