### 2026-03-26 b05ae29
fix: use enum value
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
