### 2026-03-26 d53db7b
fix: reload validators module to ensure we actually have everything loaded
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
