### 2026-03-26 05d084b
fix(cli): replace getResource with containsResourceWithIdentifier in delete operations
Avoids issue where one couldn't delete invalid resources

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
