### 2026-03-26 b3810d4
docs(ordered_pip): apply feedback from review
Signed-off-by: Vassilis Vassiliadis <vassilis.vassiliadis@ibm.com>
