### 2026-03-26 2a669e5
fix: use session-scoped fixture to ensure state is consistent
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
