### 2026-03-26 53372b9
fix(cli): replace getResource with containsResourceWithIdentifier in delete operations (#748)
Avoids issue where one couldn't delete invalid resources

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
