+++ orchestrator/cli/resources/discovery_space/delete.py
@@ -24,15 +24,15 @@ def delete_discovery_space(parameters: AdoDeleteCommandParameters) -> None:
 
     sql = get_sql_store(project_context=parameters.ado_configuration.project_context)
     with Status(ADO_SPINNER_QUERYING_DB) as status:
-        try:
-            sql.getResource(
-                parameters.resource_id,
-                kind=CoreResourceKinds.DISCOVERYSPACE,
-                raise_error_if_no_resource=True,
-            )
-        except ResourceDoesNotExistError:
+        if not sql.containsResourceWithIdentifier(
+            identifier=parameters.resource_id,
+            kind=CoreResourceKinds.DISCOVERYSPACE,
+        ):
             status.stop()
-            raise
+            raise ResourceDoesNotExistError(
+                resource_id=parameters.resource_id,
+                kind=CoreResourceKinds.DISCOVERYSPACE,
+            )
 
         children_resources = sql.getRelatedObjectResourceIdentifiers(
             identifier=parameters.resource_id
