### 2026-03-27 17e8084
docs(website): update ado upgrade documentation
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
