+++ website/docs/getting-started/ado.md
@@ -1,7 +1,8 @@
 <!-- markdownlint-disable code-block-style -->
 <!-- markdownlint-disable no-duplicate-heading -->
 <!-- markdownlint-disable ul-indent -->
-<!-- markdownlint-disable-next-line first-line-h1 -->
+<!-- markdownlint-disable first-line-h1 -->
+
 !!! note
 
     This page provides documentation for the `ado` CLI tool, which needs to be
@@ -97,12 +98,16 @@ Where:
 - `RESOURCE_TYPE` is one of the supported resource types for `ado create`,
   currently:
 
+<!-- prettier-ignore-start -->
+
     - _actuator_
     - _actuatorconfiguration_ (_ac_)
     - _context_ (_ctx_)
     - _operation_ (_op_)
     - _samplestore_ (_store_)
     - _discoveryspace_ (_space_)
+
+<!-- prettier-ignore-end -->
 
 - `--file` or `-f` is a path to the resource configuration file in YAML format.
   It is mandatory in all scenarios, except when running
@@ -217,6 +222,8 @@ Where:
 - `RESOURCE_TYPE` is the type of resource you want to delete. Currently, the
   only supported types are:
 
+<!-- prettier-ignore-start -->
+
     - _actuatorconfiguration_ (_ac_)
     - _context_ (_ctx_)
     - _datacontainer_ (_dcr_)
@@ -224,10 +231,18 @@ Where:
     - _samplestore_ (_store_)
     - _discoveryspace_ (_space_)
 
+<!-- prettier-ignore-end -->
+
 - `RESOURCE_ID` is the unique identifier of the resource to delete.
 - `--force` enables forced deletion of resources in the following cases:
+
+<!-- prettier-ignore-start -->
+
     - When attempting to delete operations while other operations are executing.
     - When attempting to delete sample stores that still contain data.
+
+<!-- prettier-ignore-end -->
+
 - When deleting a local context, users can specify the flags `--delete-local-db`
   or `--no-delete-local-db` to explicitly delete or preserve a local DB when
   deleting its related context. If neither of these flags are specified, the
@@ -270,9 +285,13 @@ Where:
 - `RESOURCE_TYPE` is the type of resource you want to describe. Currently, the
   supported resource types are:
 
+<!-- prettier-ignore-start -->
+
     - _experiment_
     - _datacontainer_ (_dcr_)
     - _discoveryspace_ (_space_)
+
+<!-- prettier-ignore-end -->
 
 - `RESOURCE_ID` is the unique identifier of the resource to describe.
 - The `--file` (or `-f`) flag is **currently only available for spaces** and
@@ -306,19 +325,27 @@ Where:
 
 - `RESOURCE_TYPE` is the type of resource you want to edit. Supported types are:
 
+<!-- prettier-ignore-start -->
+
     - _actuatorconfiguration_ (_ac_)
     - _datacontainer_ (_dcr_)
     - _operation_ (_op_)
     - _samplestore_ (_store_)
     - _discoveryspace_ (_space_)
 
+<!-- prettier-ignore-end -->
+
 - `RESOURCE_ID` is the unique identifier of the resource to edit.
 - `--editor` is the name of the editor you want to use for editing metadata. It
   must be one of the supported ones, which currently are:
 
+<!-- prettier-ignore-start -->
+
     - `vim` (_default_)
     - `vi`
     - `nano`
+
+<!-- prettier-ignore-end -->
 
   Alternatively, you can also set the value for this flag by using the
   environment variable `ADO_EDITOR`.
@@ -376,6 +403,8 @@ Where:
 - `RESOURCE_TYPE` is the type of resource you want to get. Currently, the only
   supported types are:
 
+<!-- prettier-ignore-start -->
+
     - _actuatorconfiguration_ (_ac_)
     - _actuator_
[…]
