### 2026-03-27 51b6c5f
docs(ordered_pip): decouple Kuberay docs from ordered_pip docs
Signed-off-by: Vassilis Vassiliadis <vassilis.vassiliadis@ibm.com>
