+++ backend/kuberay/README.md
@@ -134,6 +134,8 @@ with 4 Nodes each with 8 NVIDIA-A100-SXM4-80GB GPUs, 64 CPU cores, and 1TB memor
         num-gpus: '1'
         resources: '"{\"NVIDIA-A100-SXM4-80GB\": 1}"'
       containerEnv:
+        - name: RAY_RUNTIME_ENV_PLUGINS
+          value: '[{"class":"orchestrator.utilities.ray_env.ordered_pip.OrderedPipPlugin"}]'
         - name: OMP_NUM_THREADS
           value: "1"
         - name: OPENBLAS_NUM_THREADS
@@ -173,6 +175,8 @@ with 4 Nodes each with 8 NVIDIA-A100-SXM4-80GB GPUs, 64 CPU cores, and 1TB memor
         num-gpus: '2'
         resources: '"{\"NVIDIA-A100-SXM4-80GB\": 2}"'
       containerEnv:
+        - name: RAY_RUNTIME_ENV_PLUGINS
+          value: '[{"class":"orchestrator.utilities.ray_env.ordered_pip.OrderedPipPlugin"}]'
         - name: OMP_NUM_THREADS
           value: "1"
         - name: OPENBLAS_NUM_THREADS
@@ -212,6 +216,8 @@ with 4 Nodes each with 8 NVIDIA-A100-SXM4-80GB GPUs, 64 CPU cores, and 1TB memor
         num-gpus: '4'
         resources: '"{\"NVIDIA-A100-SXM4-80GB\": 4}"'
       containerEnv:
+        - name: RAY_RUNTIME_ENV_PLUGINS
+          value: '[{"class":"orchestrator.utilities.ray_env.ordered_pip.OrderedPipPlugin"}]'
         - name: OMP_NUM_THREADS
           value: "1"
         - name: OPENBLAS_NUM_THREADS
@@ -251,6 +257,8 @@ with 4 Nodes each with 8 NVIDIA-A100-SXM4-80GB GPUs, 64 CPU cores, and 1TB memor
         num-gpus: '8'
         resources: '"{\"NVIDIA-A100-SXM4-80GB\": 8, \"full-worker\": 1}"'
       containerEnv:
+        - name: RAY_RUNTIME_ENV_PLUGINS
+          value: '[{"class":"orchestrator.utilities.ray_env.ordered_pip.OrderedPipPlugin"}]'
         - name: OMP_NUM_THREADS
           value: "1"
         - name: OPENBLAS_NUM_THREADS
@@ -305,79 +313,31 @@ with 4 Nodes each with 8 NVIDIA-A100-SXM4-80GB GPUs, 64 CPU cores, and 1TB memor
 
 ## Using the OrderedPip Ray Runtime Environment Plugin
 
-The `OrderedPipPlugin` is a Ray RuntimeEnvPlugin that enables you to control the
-build order of Python packages. This is useful when installing packages packages
-with build-time dependencies. For example, `mamba-ssm` and `torch`.
-
-### Configuration Details
-
-> [!IMPORTANT]
->
-> Each entry in `phases` uses the **identical schema** as Ray's standard `pip`
-> runtime environment field. If you know how to configure `pip`, you already
-> know how to configure each phase in `ordered_pip`.
-
-The `ordered_pip` runtime environment accepts a dictionary with a `phases` key:
-
-- **`phases`**: A list where **each element follows the exact same schema as the
-  standard Ray `pip` field**. This means each phase can be:
-  - A list of package names (e.g., `["torch==2.6.0"]`)
-  - A dictionary with `packages` and optional `pip_install_options` fields
-  - Any other valid `pip` specification format
-
-### Availability
-
-The `OrderedPipPlugin` is pre-installed in ado Docker images.
-However, it is switched off by default.
[…]
