### 2026-03-27 e140101
docs(sft_trainer): recommend using ordered_pip
Signed-off-by: Vassilis Vassiliadis <vassilis.vassiliadis@ibm.com>
