### 2026-03-27 9dd95f0
docs(sft_trainer): fix broken link
Signed-off-by: Vassilis Vassiliadis <vassilis.vassiliadis@ibm.com>
