### 2026-03-27 abd5bcc
chore(deps): update dependency fastapi to v0.135.2
