### 2026-03-27 487f6bd
chore(deps): update dependency uv to v0.11.2
