### 2026-03-30 f6e9ff8
feat(vllm_performance): Guarantee the vLLM deployment name is RFC1123 compliant
Signed-off-by: Christian Pinto <christian.pinto@ibm.com>
