### 2026-03-30 a0934c4
feat(vllm_performance): Updated pydantic Annotated field style to the project template
Signed-off-by: Christian Pinto <christian.pinto@ibm.com>
