### 2026-03-30 0b9464a
build(vllm_performance): Updated pyproject toml
Signed-off-by: Christian Pinto <christian.pinto@ibm.com>
