### 2026-03-30 0991e67
feat(vllm_performance): Upgraded custom arguments in execute_benchmark
Signed-off-by: Christian Pinto <christian.pinto@ibm.com>
