### 2026-03-30 54510e6
refactor: rename legacy validator to legacy migrator
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
