+++ orchestrator/core/legacy/migrators/samplestore/v1_to_v2_csv_migration.py
@@ -1,14 +1,14 @@
 # Copyright IBM Corporation 2025, 2026
 # SPDX-License-Identifier: MIT
 
-"""Legacy validator for migrating CSV sample stores from v1 to v2 format"""
+"""Legacy migrator for migrating CSV sample stores from v1 to v2 format"""
 
-from orchestrator.core.legacy.registry import legacy_validator
+from orchestrator.core.legacy.registry import legacy_migrator
 from orchestrator.core.legacy.utils import get_nested_value, has_nested_field
 from orchestrator.core.resources import CoreResourceKinds
 
 
-@legacy_validator(
+@legacy_migrator(
     identifier="csv_constitutive_columns_migration",
     resource_type=CoreResourceKinds.SAMPLESTORE,
     deprecated_field_paths=[
