+++ orchestrator/core/legacy/migrators/discoveryspace/__init__.py
@@ -1,9 +1,9 @@
 # Copyright IBM Corporation 2025, 2026
 # SPDX-License-Identifier: MIT
 
-"""Legacy validators for discovery space migrations"""
+"""Legacy migrators for discovery space migrations"""
 
-from orchestrator.core.legacy.validators.discoveryspace import (
+from orchestrator.core.legacy.migrators.discoveryspace import (
     entitysource_to_samplestore,
     properties_field_removal,
 )
