+++ orchestrator/cli/utils/legacy/__init__.py
@@ -1,10 +1,10 @@
 # Copyright IBM Corporation 2025, 2026
 # SPDX-License-Identifier: MIT
 
-"""Utilities for working with legacy validators"""
+"""Utilities for working with legacy migrators"""
 
-from orchestrator.cli.utils.legacy.list import list_legacy_validators
+from orchestrator.cli.utils.legacy.list import list_legacy_migrators
 
-__all__ = ["list_legacy_validators"]
+__all__ = ["list_legacy_migrators"]
 
 # Made with Bob

[…]
