### 2026-03-30 a8e1bfb
docs(sfttrainer): update docs to point to ofed image
Signed-off-by: Vassilis Vassiliadis <vassilis.vassiliadis@ibm.com>
