### 2026-03-30 27d0bc3
feat(sql): add function for checking table existence.
