### 2026-03-30 622e275
fix(sql): scope SQLSampleStore table-existence cache by database URL
The _source_tables_verified cache used only the table name as its key,
so two SQLSampleStore instances with the same identifier but pointing to
different databases would incorrectly share a cache entry. The second
store would skip the table-existence check and attempt to query tables
that don't exist in its database.

Fix by keying the cache on (db_url, tablename), consistent with how
_tables_exist_cache is already scoped in SQLStore (sqlstore.py).
