### 2026-03-31 9a9d1b2
refactor(core): remove unused CoreResourceKinds
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
