### 2026-03-31 8f5dc93
fix(core): re-export AdoResource
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
