+++ orchestrator/core/__init__.py
@@ -5,6 +5,7 @@ from .actuatorconfiguration.resource import ActuatorConfigurationResource
 from .datacontainer.resource import DataContainerResource
 from .discoveryspace.resource import DiscoverySpaceResource
 from .operation.resource import OperationResource
+from .resources import ADOResource as ADOResource
 from .resources import CoreResourceKinds
 from .samplestore.resource import SampleStoreResource
 
