### 2026-03-31 920c897
feat(cli): save latest operation id even on failure
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
