### 2026-03-31 56023bc
refactor(cli): remove MeasurementError handling
ref: https://github.com/IBM/ado/issues/290#issuecomment-4161032297
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
