### 2026-03-31 da324fc
feat(core): compress ValidMeasurementResults when serializing
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
