+++ tests/schema/test_measurement_result.py
@@ -280,3 +280,222 @@ def test_invalid_measurement_result_series_representation(
         assert representation["experiment_id"] == str(expected_experiment_identifier)
         assert representation["valid"] == expected_validity
         assert representation["reason"] == expected_reason
+
+
+def test_compressed_serialization_format(
+    entity: Entity,
+    property_values: list[ObservedPropertyValue | ConstitutivePropertyValue],
+) -> None:
+    """Verify new serialization format has experimentReference at top level"""
+    result = ValidMeasurementResult(
+        entityIdentifier=entity.identifier, measurements=property_values
+    )
+    serialized = result.model_dump()
+
+    # Verify new format structure
+    assert "experimentReference" in serialized
+    assert "measurements" in serialized
+    assert "uid" in serialized
+    assert "entityIdentifier" in serialized
+
+    # Verify measurements don't have 'property' field (new format)
+    assert "property" not in serialized["measurements"][0]
+
+    # Verify measurements have simplified structure
+    assert "targetProperty" in serialized["measurements"][0]
+    assert "value" in serialized["measurements"][0]
+    assert "valueType" in serialized["measurements"][0]
+
+
+def test_old_format_deserialization(entity: Entity) -> None:
+    """Verify old format can still be deserialized"""
+    from orchestrator.schema.domain import PropertyDomain, VariableTypeEnum
+
+    # Create old format JSON (with experimentReference in each measurement)
+    old_format_json = {
+        "uid": "12345678-1234-1234-1234-123456789012",
+        "entityIdentifier": entity.identifier,
+        "metadata": {},
+        "measurements": [
+            {
+                "property": {
+                    "experimentReference": {
+                        "actuatorIdentifier": "test_actuator",
+                        "experimentIdentifier": "test_experiment",
+                        "parameterization": None,
+                    },
+                    "targetProperty": {
+                        "identifier": "test_property",
+                        "propertyType": "MEASURED_PROPERTY_TYPE",
+                        "propertyDomain": PropertyDomain(
+                            values=["a", "b"],
+                            variableType=VariableTypeEnum.CATEGORICAL_VARIABLE_TYPE,
+                        ).model_dump(),
+                    },
+                    "metadata": {},
+                },
+                "value": 42.0,
+                "valueType": "NUMERIC_VALUE_TYPE",
+                "uncertainty": None,
+            }
+        ],
+    }
+
+    # Should deserialize successfully
+    result = ValidMeasurementResult.model_validate(old_format_json)
+    assert result.uid == old_format_json["uid"]
+    assert result.entityIdentifier == old_format_json["entityIdentifier"]
+    assert len(result.measurements) == 1
+    assert result.measurements[0].value == 42.0
+
+
+def test_new_format_deserialization(entity: Entity) -> None:
+    """Verify new format deserializes correctly"""
+    from orchestrator.schema.domain import PropertyDomain, VariableTypeEnum
+
+    # Create new format JSON (experimentReference at top level)
+    new_format_json = {
+        "uid": "12345678-1234-1234-1234-123456789012",
+        "entityIdentifier": entity.identifier,
+        "metadata": {},
+        "experimentReference": {
+            "actuatorIdentifier": "test_actuator",
+            "experimentIdentifier": "test_experiment",
+            "parameterization": None,
+        },
+        "measurements": [
+            {
+                "targetProperty": {
+                    "identifier": "test_property",
+                    "propertyType": "MEASURED_PROPERTY_TYPE",
+                    "propertyDomain": PropertyDomain(
+                        values=["a", "b"],
[…]
