### 2026-03-31 29b389a
refactor(core): dump PropertyValue and add required fields
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
