### 2026-03-31 a528baa
refactor(core): use ObservedPropertyValue when deserializing
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
