### 2026-03-31 5378390
docs(core): trim down comment about serialized representation
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
