### 2026-03-31 fa53aaf
feat(sql): create check_table_exists utility function.
Uses fast-path if db is sqlite or mysql or falls back to inspect/has_table if not.
