### 2026-03-31 929ef69
refactor(core): rename functions and drop extended description
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
