### 2026-04-01 158a0c2
fix(vllm_performance): Fixed ruff errors and re-added datasets as dependency
Signed-off-by: Christian Pinto <christian.pinto@ibm.com>
