### 2026-04-01 c0ab519
refactor(core): exclude only experimentReference
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
