### 2026-04-01 f70afb6
feat(core): add support for exclusions in dumps
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
