### 2026-04-01 c414317
refactor(core): use pydantic models directly where possible
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
