### 2026-04-01 b1d9cb8
Apply suggestions from code review
Co-authored-by: Alessandro Pomponio <10339005+AlessandroPomponio@users.noreply.github.com>
Co-authored-by: Michael Johnston <66301584+michael-johnston@users.noreply.github.com>
Signed-off-by: Michael Johnston <66301584+michael-johnston@users.noreply.github.com>
