### 2026-04-01 0b7e165
refactor(core): simplify adding back experiment reference
Co-Authored-By: Michael Johnston <66301584+michael-johnston@users.noreply.github.com>
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
