### 2026-04-01 e97e32d
build(vllm_performance): require minimum datasets version (#775)
fix(vllm_performance): Added datasets to dependencies to avoid pip/uv resolving it to a version that is not supported by vLLM

Signed-off-by: Christian Pinto <christian.pinto@ibm.com>
