### 2026-04-01 57e8ce0
feat(core): compress ValidMeasurementResults when serializing (#769)
* feat(core): compress ValidMeasurementResults when serializing

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* refactor(core): dump PropertyValue and add required fields

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* refactor(core): use ObservedPropertyValue when deserializing

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* docs(core): trim down comment about serialized representation

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* refac
[…]
