### 2026-04-01 84a9702
Update plugins/custom_experiments/cplex_mip/cplex_mip_experiments/solve_mip.py
Co-authored-by: Alessandro Pomponio <10339005+AlessandroPomponio@users.noreply.github.com>
Signed-off-by: Michael Johnston <66301584+michael-johnston@users.noreply.github.com>
