+++ plugins/custom_experiments/cplex_mip/cplex_mip_experiments/solve_mip.py
@@ -156,7 +156,6 @@ Parallel = ConstitutiveProperty(
     },
     propertyDomain=PropertyDomain(
         variableType=VariableTypeEnum.BINARY_VARIABLE_TYPE,
-        values=[False, True],
     ),
 )
 
