### 2026-04-01 19e122e
feat(autoconf): log confusion matrix (#743)
