### 2026-04-02 88def2e
perf(core): use simpler checks for table existence (#670)
* feat(perf): skip redundant DDL checks in SQLSampleStore.__init__

Previously, SQLSampleStore.__init__ called _create_source_table()
unconditionally on every construction.  _create_source_table() calls
meta.create_all(checkfirst=True), which issues a separate table-existence
query for each of the four backing tables even when
the tables already exist — which they always do on read-only CLI commands.
Locally this cost was ~1.45s. The changes reduce this to 0.24s.

The main fix is to probe for table existence instead of blinding issuing 
[…]
