### 2026-04-02 f919aac
feat(tests): Removed property type prefix references from tests
Signed-off-by: Christian Pinto <christian.pinto@ibm.com>
