### 2026-04-02 0cf12e2
fix(cli): use unique entities to calculate entities with no successful measurements
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
