### 2026-04-02 bc7dde1
refactor(cli): improve prints on upgrade error
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
