### 2026-04-02 7749d53
refactor(core): remove property type prefix in string representation (#780)
* feat(ado-core): Remove property type prefix when converting property identifier to string

Signed-off-by: Christian Pinto <christian.pinto@ibm.com>

* feat(tests): Removed property type prefix references from tests

Signed-off-by: Christian Pinto <christian.pinto@ibm.com>

* feat(docs): Updated documentation to remove references to property type prefix

Signed-off-by: Christian Pinto <christian.pinto@ibm.com>

---------

Signed-off-by: Christian Pinto <christian.pinto@ibm.com>
