+++ examples/ml-multi-cloud/EXAMPLE_OBJECTIVE.md
@@ -73,6 +73,7 @@ ado get experiments --details
 will output something similar to:
 
 <!-- markdownlint-disable line-length -->
+
 ```commandline
 ┌────────────────────┬─────────────────────────┐
 │ ACTUATOR ID        │ EXPERIMENT ID           │
@@ -82,6 +83,7 @@ will output something similar to:
 │ mock               │ test-experiment-two     │
 └────────────────────┴─────────────────────────┘
 ```
+
 <!-- markdownlint-enable line-length -->
 
 You can see the custom experiment provided by the package,
@@ -89,39 +91,41 @@ You can see the custom experiment provided by the package,
 `ado describe experiment ml-multicloud-cost-v1.0` outputs:
 
 <!-- markdownlint-disable line-length -->
+
 ```terminaloutput
 Identifier: custom_experiments.ml-multicloud-cost-v1.0
 
 Required Inputs:
-                                                                             
-   Constitutive Properties:                                                  
-    ─────────────────────────────────────────────────────────────────────    
-     Identifier: nodes                                                       
-     Domain:                                                                 
-                                                                             
-        Type: DISCRETE_VARIABLE_TYPE                                         
-        Interval: 1                                                          
-        Range: [0, 1000]                                                     
-                                                                             
-    ─────────────────────────────────────────────────────────────────────    
-    ─────────────────────────────────────────────────────────────────────    
-     Identifier: cpu_family                                                  
-     Domain:                                                                 
-                                                                             
-        Type: DISCRETE_VARIABLE_TYPE                                         
-        Values: [0, 1]                                                       
-                                                                             
-    ─────────────────────────────────────────────────────────────────────    
-   Observed Properties:                                                      
-                                                                             
-      op-benchmark_performance-wallClockRuntime                              
-                                                                             
-                                                                             
+
+   Constitutive Properties:
+    ─────────────────────────────────────────────────────────────────────
+     Identifier: nodes
+     Domain:
+
+        Type: DISCRETE_VARIABLE_TYPE
+        Interval: 1
+        Range: [0, 1000]
+
+    ─────────────────────────────────────────────────────────────────────
+    ─────────────────────────────────────────────────────────────────────
+     Identifier: cpu_family
+     Domain:
+
+        Type: DISCRETE_VARIABLE_TYPE
+        Values: [0, 1]
+
+    ─────────────────────────────────────────────────────────────────────
+   Observed Properties:
+
+      benchmark_performance-wallClockRuntime
+
+
 Outputs:
- ─────────────────────────────────────────────────────────────────────────── 
-   ml-multicloud-cost-v1.0-total_cost                                        
+ ───────────────────────────────────────────────────────────────────────────
+   ml-multicloud-cost-v1.0-total_cost
  ───────────────────────────────────────────────────────────────────────────
 ```
+
 <!-- markdownlint-enable line-length -->
 
 From this, you can see the `ml-multicloud-cost-v1.0` requires an observed
@@ -160,10 +164,10 @@ ado create space -f ml_multicloud_space_with_custom.yaml --set "sampleStoreIdent
 
 > [!IMPORTANT]
 >
-> If an experiment takes the output of another experiment as input
-> both experiments must be in the `discoveryspace`. In the above example if the
-> entry `benchmark_performance` was omitted the `ado create space` command would
-> fail with:
+> If an experiment takes the output of another experiment as input both
+> experiments must be in the `discoveryspace`. In the above example if the entry
+> `benchmark_performance` was omitted the `ado create space` command would fail
+> with:
 >
 > **SpaceInconsistencyError**: MeasurementSpace does not contain an experiment
[…]
