### 2026-04-02 d4b0ac1
feat(cli): support legacy validators in ado upgrade (#629)
* feat(core): implement phase 1 and 2

Made with Bob

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* feat(core): finish phase 2 implementation

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* feat(core): implement phase 3

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* test: add tests

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* feat(cli): add upgrade detection

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* refactor(cli): up
[…]
