+++ tests/core/test_upgrade_transaction_safety.py
@@ -1,0 +1,304 @@
+# Copyright IBM Corporation 2025, 2026
+# SPDX-License-Identifier: MIT
+
+"""Integration tests for Phase 1 transaction safety in upgrade handler"""
+
+import json
+import sqlite3
+
+import pytest
+import sqlalchemy
+import typer
+
+from orchestrator.cli.core.config import AdoConfiguration
+from orchestrator.cli.models.parameters import AdoUpgradeCommandParameters
+from orchestrator.cli.utils.generic.wrappers import get_sql_store
+from orchestrator.cli.utils.resources.handlers import handle_ado_upgrade
+from orchestrator.core.legacy.registry import legacy_migrator
+from orchestrator.core.resources import CoreResourceKinds
+from orchestrator.core.samplestore.config import (
+    SampleStoreConfiguration,
+    SampleStoreModuleConf,
+    SampleStoreSpecification,
+)
+from orchestrator.core.samplestore.resource import SampleStoreResource
+from orchestrator.metastore.project import ProjectContext
+
+sqlite3_version = sqlite3.sqlite_version_info
+
+
+class TestUpgradeTransactionSafety:
+    """Test transaction safety in upgrade handler - validate-all-before-save pattern"""
+
+    @pytest.mark.parametrize("valid_ado_project_context", ["mysql"], indirect=True)
+    def test_all_resources_validated_before_any_saved(
+        self,
+        isolated_legacy_migrator_registry: None,
+        valid_ado_project_context: ProjectContext,
+    ) -> None:
+        """Test that all resources are validated before any are saved"""
+
+        # Register a test migrator that transforms old_field -> new_field
+        @legacy_migrator(
+            identifier="test_transaction_migrator",
+            resource_type=CoreResourceKinds.SAMPLESTORE,
+            deprecated_field_paths=["config.metadata.old_field"],
+            deprecated_from_version="1.0.0",
+            removed_from_version="2.0.0",
+            description="Test transaction migrator",
+        )
+        def test_migrator(data: dict) -> dict:
+            if "config" in data and "metadata" in data["config"]:
+                metadata = data["config"]["metadata"]
+                if "old_field" in metadata:
+                    metadata["new_field"] = metadata.pop("old_field")
+            return data
+
+        # Create two sample store resources with old_field in metadata
+        resource1 = SampleStoreResource(
+            identifier="test_res1",
+            config=SampleStoreConfiguration(
+                specification=SampleStoreSpecification(
+                    module=SampleStoreModuleConf(
+                        moduleClass="SQLSampleStore",
+                        moduleName="orchestrator.core.samplestore.sql",
+                    ),
+                    storageLocation=valid_ado_project_context.metadataStore,
+                ),
+                metadata={"old_field": "value1"},
+            ),
+        )
+
+        resource2 = SampleStoreResource(
+            identifier="test_res2",
+            config=SampleStoreConfiguration(
+                specification=SampleStoreSpecification(
+                    module=SampleStoreModuleConf(
+                        moduleClass="SQLSampleStore",
+                        moduleName="orchestrator.core.samplestore.sql",
+                    ),
+                    storageLocation=valid_ado_project_context.metadataStore,
+                ),
+                metadata={"old_field": "value2"},
+            ),
+        )
+
+        # Save resources to database
+        sql_store = get_sql_store(project_context=valid_ado_project_context)
+        sql_store.updateResource(resource=resource1)
+        sql_store.updateResource(resource=resource2)
+
+        # Now manually add the deprecated field to the raw data in the database
+        # We need to update the JSON directly in the database
+        with sql_store.engine.begin() as conn:
+            # Get current data
[…]
