+++ orchestrator/core/legacy/migrators/discoveryspace/__init__.py
@@ -1,0 +1,13 @@
+# Copyright IBM Corporation 2025, 2026
+# SPDX-License-Identifier: MIT
+
+"""Legacy migrators for discovery space migrations"""
+
+from orchestrator.core.legacy.migrators.discoveryspace import (
+    entitysource_to_samplestore,
+    properties_field_removal,
+)
+
+__all__ = ["entitysource_to_samplestore", "properties_field_removal"]
+
+# Made with Bob
