### 2026-04-02 2d10242
fix(cli): prevent switching to invalid context and provide failsafe
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
