### 2026-04-02 18bf020
fix: update test
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
