### 2026-04-02 d474828
test: add tests for new context behaviours
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
