### 2026-04-02 9b587fe
feat(cli): add version info to ado get operators
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
