### 2026-04-02 7a4ce9e
refactor: use dynamic versions for operators
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
