### 2026-04-02 0ec2eb8
fix(cli): prevent switching to invalid context and provide failsafe (#783)
* fix(cli): prevent switching to invalid context and provide failsafe

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* fix: update test

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* test: add tests for new context behaviours

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

---------

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
