### 2026-04-04 #10304 @mzrghorbani
I have completed my review of **ado**. The software successfully fulfills the JOSS criteria.

Here is a summary of my findings:
- **Repository & License:** The codebase is well-organized and uses an OSI-approved MIT license.
- **Documentation & Open Source Practice:** The repository demonstrates excellent open-source practices, including comprehensive `CONTRIBUTING.md`, `DEVELOPING.md`, and a clear `CODE_OF_CONDUCT.md`. The Git history confirms a sustained development timeline involving multiple authors.
- **Installation & Functionality:** I was able to cleanly install the framework via `uv` as detailed in the documentation. I ran the internal `pytest` suite which passed successfully (excluding local docker requirements). Furthermore, I ran the `optimization_test_functions` example, validating the core `Discovery Space` abstraction and its ability to seamlessly scale out operations using `RayTune`.

The paper clearly articulates the research gap in orchestrating generic experiment campaigns (vs context-free DAGs) and the software solves it effectively.

Everything looks great. I recommend this submission for acceptance!

*(Disclosure: I used the GitHub Copilot AI agent to assist with navigating the codebase, running the installation/test commands in a local virtual environment, and grammar-checking this review summary.)*
