### 2026-04-07 #10304 @gmarupilla
The paper is well written, and I am happy to support acceptance. I found it clear, readable, and well structured, and it does not need major editorial work. The software design discussion, especially the `Discovery Space` and `TRACE` framing, is the strongest part of the manuscript.

My main suggestion is to strengthen the `State of the field` section. The paper specifies the relevant neighboring tools, but at present, the comparison feels a bit more assertive than comparative. I think the manuscript would benefit from a clearer build vs. contribute justification –– what the existing tools already do well, where they fall short in this use case, and what ado contributes that is genuinely distinct.

Also, the `Research Impact Statement` would be stronger with more concrete evidence or more careful framing. The current examples of internal use and public artifacts are useful and suggest credible significance, but the paper would benefit from being clearer about what is already realized impact and what is expected near term impact. These are suggestions to strengthen the paper further, no concerns on preventing from recommending the acceptance.
