### 2026-04-07 150bac1
refactor(cli): swap order of type and version columns
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
