+++ orchestrator/cli/resources/operator/get.py
@@ -43,10 +43,10 @@ def get_operator(parameters: AdoGetCommandParameters) -> None:
         for function_name in collection.function_operations:
             entry = {
                 "OPERATOR": function_name,
-                "TYPE": collection.type.value,
                 "VERSION": collection.function_operation_versions.get(
                     function_name, ""
                 ),
+                "TYPE": collection.type.value,
             }
             if parameters.show_details:
                 entry["DESCRIPTION"] = normalize_and_truncate_at_period(
