### 2026-04-07 75d6b17
feat(cli): add support for -o name in ado get
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
