### 2026-04-07 a26f7c8
build: exclude uv.lock from tombi formatting
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
