+++ tombi.toml
@@ -1,3 +1,3 @@
 [files]
 include = ["**/*.toml"]
-exclude = [".venv/**", ".tox/**"]
+exclude = [".venv/**", ".tox/**", "uv.lock"]
