### 2026-04-07 e6b9a2f
feat(orchestration): Explore class registration
Explore operator function decorator now allows registering the class that embodies the operator.

This means orchestrate_explore_operation can be passed an OperatorFunctionConf instance and retrieve the class rather than needed an OperatorModuleConf instance.
