### 2026-04-07 5d44927
refactor(operators): rename OperationCollections to OperatorCollection
