### 2026-04-07 d7fbd68
feat(cli): add options to avoid truncating rich table content
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
