### 2026-04-07 0fa7656
feat(cli): add --no-trunc flag for commands that output rich tables
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
