### 2026-04-07 bb5eecd
docs: format comment
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
