### 2026-04-07 e7bc38b
feat(operators): Add a type for operator functions
Need to use a protocol due to using kwargs
