### 2026-04-07 5acfc41
refactor(operators): Update to use new protocol
