### 2026-04-07 c951e8d
refactor(operators): Use pydantic model for operators
Also compress the decoration/registration to single functions
