+++ orchestrator/modules/operators/collections.py
@@ -4,6 +4,7 @@
 import functools
 import logging
 import typing
+import warnings
 from typing import Annotated
 
 import pydantic
@@ -15,169 +16,200 @@ from orchestrator.core.operation.config import (
     DiscoveryOperationEnum,
     FunctionOperationInfo,
 )
-from orchestrator.modules.operators.base import DiscoveryOperationBase, OperationOutput
+from orchestrator.modules.operators.base import (
+    DiscoveryOperationBase,
+    OperationOutput,
+    OperatorFunction,
+)
 from orchestrator.modules.operators.orchestrate import orchestrate_general_operation
 
 moduleLog = logging.getLogger("operation_collections")
 
 
-class OperatorCollection(pydantic.BaseModel):
+class Operator(pydantic.BaseModel):
+    """Metadata and implementation for a registered operator.
+
+    Attributes:
+        name: Canonical name the operator is registered under.
+        function: The callable implementing the operator.
+        version: Version string for the operator (e.g. "v0.1").
+        description: Human-readable description of the operator.
+        configuration_model: Pydantic model class used to validate parameters.
+        example_configuration: Default instance of the configuration model.
+        cls: Ray actor class backing the operator (explore operators only).
+        type: The discovery operation type this operator belongs to.
+    """
+
+    name: str
+    function: OperatorFunction
+    version: Annotated[str, pydantic.Field(default="v0.1")]
+    description: Annotated[str | None, pydantic.Field(default=None)]
+    configuration_model: Annotated[
+        type[pydantic.BaseModel] | None, pydantic.Field(default=None)
+    ]
+    example_configuration: Annotated[
+        pydantic.BaseModel | None, pydantic.Field(default=None)
+    ]
+    cls: Annotated[typing.Any, pydantic.Field(default=None)]
     type: DiscoveryOperationEnum
-    function_operations: Annotated[
-        dict[typing.AnyStr, typing.Callable], pydantic.Field(default_factory=dict)
-    ]
-    object_operations: Annotated[
-        dict[typing.AnyStr, DiscoveryOperationBase],
-        pydantic.Field(default_factory=dict),
-    ]
-    function_operation_models: Annotated[
-        dict[typing.AnyStr, type[pydantic.BaseModel]],
-        pydantic.Field(default_factory=dict),
-    ]
-    function_operation_model_defaults: Annotated[
-        dict[typing.AnyStr, pydantic.BaseModel], pydantic.Field(default_factory=dict)
-    ]
-    function_operation_versions: Annotated[
-        dict[typing.AnyStr, str], pydantic.Field(default_factory=dict)
-    ]
-    function_operation_descriptions: Annotated[
-        dict[typing.AnyStr, str], pydantic.Field(default_factory=dict)
-    ]
-    operator_classes: Annotated[
-        dict[typing.AnyStr, type],
-        pydantic.Field(default_factory=dict),
-    ]
 
     model_config = ConfigDict(arbitrary_types_allowed=True)
 
-    def add_operation_function(self, name: str, fn: typing.Callable) -> None:
-        """Registers a callable under the given name."""
-        self.function_operations[name] = fn
 
-    def add_operation_version(self, name: str, version: str) -> None:
-        """Stores the version string for the named operation."""
-        self.function_operation_versions[name] = version
+class OperatorCollection(pydantic.BaseModel):
+    """A registry of operators of a single discovery operation type.
 
-    def add_operation_description(self, name: str, version: str) -> None:
-        """Stores the description string for the named operation."""
-        self.function_operation_descriptions[name] = version
+    Operators are added via the decorator functions (e.g. ``characterize_operation``,
+    ``explore_operation``).  Each registered name maps to an :class:`Operator`
[…]
