### 2026-04-07 0de3555
refactor(operators): Update tests
