### 2026-04-07 6bebd5a
feat(cli): add support for -o name in ado get  (#794)
* feat(cli): add support for -o name in ado get

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* feat(cli): add support for -o name in ado get actuators

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* feat(cli): add support for -o name in ado get operators

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* feat(cli): add support for -o name in ado get context

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

---------

Signed-off-by: Alessandro Pomponio <alessan
[…]
