### 2026-04-07 d761d2d
refactor(actuators): use new util function
