### 2026-04-07 6becc0c
refactor(operators): use undecorated operator classes
remove ray decorator if present. Will be re-added if necessary.
