### 2026-04-07 751876c
feat(operator): Validate function call signature
