### 2026-04-07 b5ee69b
feat(cli): do not truncate IDENTIFIER column by default
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
