+++ orchestrator/cli/resources/operator/get.py
@@ -115,6 +115,6 @@ def get_operator(parameters: AdoGetCommandParameters) -> None:
             show_edge=True,
             show_index=True,
             box=rich.box.SQUARE,
-            do_not_truncate_column_content=parameters.no_trunc,
+            do_not_truncate_columns=parameters.no_trunc,
         )
     )
