+++ orchestrator/cli/resources/context/get.py
@@ -86,7 +86,7 @@ def get_context(
                 show_edge=True,
                 show_index=True,
                 box=rich.box.SQUARE,
-                do_not_truncate_column_content=parameters.no_trunc,
+                do_not_truncate_columns=parameters.no_trunc,
             )
         )
         return
