+++ orchestrator/cli/resources/actuator/get.py
@@ -116,6 +116,6 @@ def get_actuator(parameters: AdoGetCommandParameters) -> None:
             output_df,
             box=rich.box.SQUARE,
             show_edge=True,
-            do_not_truncate_column_content=parameters.no_trunc,
+            do_not_truncate_columns=parameters.no_trunc,
         )
     )
