### 2026-04-07 2d5ec00
feat(cli): add --no-trunc flag for commands that output rich tables (#797)
* feat(cli): add options to avoid truncating rich table content

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* feat(cli): add --no-trunc flag for commands that output rich tables

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* docs: format comment

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* fix(cli): set no_trunc parameter

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

---------

Signed-off-by: Alessandro Pomponio <alessandro.pom
[…]
