+++ orchestrator/cli/utils/resources/handlers.py
@@ -163,7 +163,11 @@ def handle_ado_get_default_format(
 
             console_print(
                 dataframe_to_rich_table(
-                    output_df, box=rich.box.SQUARE, show_index=True, show_edge=True
+                    output_df,
+                    box=rich.box.SQUARE,
+                    show_index=True,
+                    show_edge=True,
+                    do_not_truncate_column_content=parameters.no_trunc,
                 )
             )
             return
@@ -183,7 +187,12 @@ def handle_ado_get_default_format(
         )
 
         console_print(
-            dataframe_to_rich_table(output_df, box=rich.box.SQUARE, show_edge=True)
+            dataframe_to_rich_table(
+                output_df,
+                box=rich.box.SQUARE,
+                show_edge=True,
+                do_not_truncate_column_content=parameters.no_trunc,
+            )
         )
 
 
