### 2026-04-08 919b353
feat(cli): add version info to ado get operators  (#785)
* feat(cli): add version info to ado get operators

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* refactor: use dynamic versions for operators

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* refactor(cli): swap order of type and version columns

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* refactor(ray_tune): use ado-ray-tune version in operatorIdentifier

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

---------

Signed-off-by: Alessandro Pomponio <a
[…]
