+++ plugins/operators/trim/src/trim/operator.py
@@ -3,6 +3,7 @@
 
 
 import logging
+from importlib.metadata import version
 
 from no_priors_characterization.utils import get_source_and_target
 
@@ -31,6 +32,7 @@ logger_trim = logging.getLogger(__name__)
                 Retrieves all measured entities from the entity source and samples the others following a certain order.
                 If the number of measured entity is too small, Trim instantiates a no-priors characterization operation.
                 """,
+    version=version("ado-trim"),
 )
 def trim(
     discoverySpace: DiscoverySpace = None,  # type: ignore[name-defined]
