### 2026-04-08 cd1cf8c
feat(cli): increase table width if terminal is too small
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
