### 2026-04-08 c4678de
feat(cli): prevent truncation of operator column in ado get operators
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
