+++ orchestrator/cli/resources/actuator/get.py
@@ -21,6 +21,9 @@ from orchestrator.utilities.rich import dataframe_to_rich_table
 
 def get_actuator(parameters: AdoGetCommandParameters) -> None:
 
+    if not parameters.no_trunc:
+        parameters.no_trunc = ["ACTUATOR ID"]
+
     console_print(
         f"{INFO}This is a local command. It will not reflect the actuators on a remote cluster.",
         stderr=True,
