### 2026-04-08 fb58b32
feat(cli): prevent truncation of context column in ado get contexts
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
