### 2026-04-08 a3315bf
feat(cli): prevent truncation of experiment id column in ado get experiments
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
