+++ orchestrator/cli/resources/experiment/get.py
@@ -25,6 +25,9 @@ def get_experiment(parameters: AdoGetCommandParameters) -> None:
     Details mode: Adds DESCRIPTION and DEPRECATED columns
     """
 
+    if not parameters.no_trunc:
+        parameters.no_trunc = ["EXPERIMENT ID"]
+
     console_print(
         f"{WARN}This is a local command. It will not reflect the experiments on a remote cluster.",
         stderr=True,
