### 2026-04-08 07b2b49
fix(test): add support for do_not_truncate_columns in render_ado_resources_to_cli_output
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
