### 2026-04-08 fdaf9e9
refactor(operator): remove decorator
