### 2026-04-08 f16adda
refactor(cli): accept list in no_trunc
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
