### 2026-04-08 6596f14
refactor(test): remove dead code
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
