### 2026-04-08 be04fd4
refactor(operator): Renaming and enable decoration of explore classes
OperatorFunctionConf -> OperatorReference
Operator -> OperatorMetadata and moved to config.py
explore decorator -> now can be used to decorate an explore operator class
