+++ orchestrator/modules/operators/collections.py
@@ -10,168 +10,66 @@ from typing import Annotated
 import pydantic
 from pydantic import ConfigDict
 
-import orchestrator.core.metadata
+import orchestrator.core.operation.config
 from orchestrator.core.discoveryspace.space import DiscoverySpace
 from orchestrator.core.operation.config import (
     DiscoveryOperationEnum,
     FunctionOperationInfo,
+    OperatorMetadata,
+    OperatorReference,
 )
 from orchestrator.modules.operators.base import (
     DiscoveryOperationBase,
     OperationOutput,
     OperatorFunction,
 )
-from orchestrator.modules.operators.orchestrate import orchestrate_general_operation
+from orchestrator.modules.operators.orchestrate import (
+    orchestrate_explore_operation,
+    orchestrate_general_operation,
+)
 
 moduleLog = logging.getLogger("operation_collections")
 
 
-class Operator(pydantic.BaseModel):
-    """Metadata and implementation for a registered operator.
+def _validate_operator_cls(cls: type, context: str) -> None:
+    """Assert that ``cls`` is a subclass of :class:`DiscoveryOperationBase`.
 
-    Attributes:
-        name: Canonical name the operator is registered under.
-        function: The callable implementing the operator.
-        version: Version string for the operator (e.g. "v0.1").
-        description: Human-readable description of the operator.
-        configuration_model: Pydantic model class used to validate parameters.
-        example_configuration: Default instance of the configuration model.
-        cls: Undecorated base class backing the operator (explore operators only).
-            Always the unwrapped Python class, never a Ray ``ActorClass``.
-        type: The discovery operation type this operator belongs to.
+    ``OperatorMetadata.cls`` is typed as ``type | None`` in the core layer to
+    keep ``config.py`` decoupled from ``base.py``.  This function is the single
+    place in ``modules/operators`` where the concrete constraint is enforced.
+
+    Args:
+        cls: The class to validate.
+        context: A short description of the call site, included in the error
+            message.
+
+    Raises:
+        TypeError: If ``cls`` is not a subclass of ``DiscoveryOperationBase``.
     """
-
-    name: str
-    # bare Callable; parameterised generics break pydantic's isinstance validator
-    function: typing.Callable
-    version: Annotated[str, pydantic.Field(default="v0.1")]
-    description: Annotated[str | None, pydantic.Field(default=None)]
-    configuration_model: Annotated[
-        type[pydantic.BaseModel] | None, pydantic.Field(default=None)
-    ]
-    example_configuration: Annotated[
-        pydantic.BaseModel | None, pydantic.Field(default=None)
-    ]
-    cls: Annotated[type[DiscoveryOperationBase] | None, pydantic.Field(default=None)]
-    type: DiscoveryOperationEnum
-
-    model_config = ConfigDict(arbitrary_types_allowed=True)
-
-    @pydantic.field_validator("function", mode="after")
-    @classmethod
-    def _validate_function_signature(cls, fn: typing.Callable) -> typing.Callable:
-        """Validate that ``function`` conforms to :class:`~orchestrator.modules.operators.base.OperatorFunction`.
-
-        :class:`~orchestrator.modules.operators.base.OperatorFunction` is the
-        Protocol that defines the required call signature and is the single
-        source of truth.  Its ``__call__`` method is introspected at validation
-        time so that any future change to the Protocol is automatically
-        reflected here without updating this validator.
-
-        For each positional parameter defined in the Protocol (paired by
-        position with the actual function's positional parameters) and for the
-        return type, the check is skipped when the annotation is absent on
-        either side — unannotated or partially-annotated callables are
-        accepted.
-
-        ``typing.get_type_hints`` is used rather than ``__annotations__``
[…]
