+++ orchestrator/core/operation/config.py
@@ -193,8 +193,178 @@ class OperatorModuleConf(ModuleConf):
         return c.operatorIdentifier()
 
 
-class OperatorFunctionConf(pydantic.BaseModel):
-    """Describes an operator vended as a function"""
+class OperatorMetadata(pydantic.BaseModel):
+    """Registry metadata for a registered operator."""
+
+    name: Annotated[
+        str,
+        pydantic.Field(description="Canonical name the operator is registered under."),
+    ]
+    function: Annotated[
+        typing.Callable | None,
+        pydantic.Field(
+            default=None,
+            description=(
+                "The callable implementing the operator. None when returned by "
+                "operator_metadata() before the decorator injects it."
+            ),
+        ),
+    ]
+    version: Annotated[
+        str,
+        pydantic.Field(
+            default="v0.1",
+            description="Semantic version string for the operator (e.g. 'v0.1').",
+        ),
+    ]
+    description: Annotated[
+        str | None,
+        pydantic.Field(
+            default=None,
+            description="Human-readable description of the operator.",
+        ),
+    ]
+    configuration_model: Annotated[
+        type[pydantic.BaseModel] | None,
+        pydantic.Field(
+            default=None,
+            description="Pydantic model class used to validate operation parameters.",
+        ),
+    ]
+    example_configuration: Annotated[
+        pydantic.BaseModel | None,
+        pydantic.Field(
+            default=None,
+            description="Default instance of the configuration model.",
+        ),
+    ]
+    cls: Annotated[
+        type | None,
+        pydantic.Field(
+            default=None,
+            description=(
+                "For explore operators: the unwrapped Python class implementing the "
+                "operator. None for function-only operators. The concrete "
+                "modules/operators layer enforces that this is a DiscoveryOperationBase "
+                "subclass; config.py treats it as an opaque type to stay decoupled."
+            ),
+        ),
+    ]
+    type: Annotated[
+        DiscoveryOperationEnum,
+        pydantic.Field(
+            description="The discovery operation type this operator belongs to."
+        ),
+    ]
+
+    model_config = ConfigDict(arbitrary_types_allowed=True)
+
+    @property
+    def operatorIdentifier(self) -> str:
+        """Canonical identifier for this operator: ``{name}-{version}``."""
+        return f"{self.name}-{self.version}"
+
+    @pydantic.field_validator("function", mode="after")
+    @classmethod
+    def _validate_function_signature(
+        cls, fn: typing.Callable | None
+    ) -> typing.Callable | None:
+        """Validate that ``function`` conforms to the ``OperatorFunction`` Protocol.
+
+        Skips validation when ``fn`` is ``None`` (returned by
+        ``operator_metadata()`` before the decorator injects the function).
+        Uses a lazy import of ``OperatorFunction`` to avoid a circular import
+        (``base.py`` imports ``config.py`` at module level).
+
+        Args:
+            fn: The callable to validate, or ``None``.
+
+        Returns:
+            The callable unchanged if validation passes, or ``None``.
+
+        Raises:
+            ValueError: If any present annotation does not match the
+                corresponding annotation in the Protocol.
+        """
+        import inspect
+
+        if fn is None:
+            return fn
+
+        # Lazy import to avoid circular dependency: base.py imports config.py
+        try:
+            from orchestrator.modules.operators.base import OperatorFunction
+
+            proto_hints = typing.get_type_hints(OperatorFunction.__call__)
+            proto_sig = inspect.signature(OperatorFunction.__call__)
+        except Exception:  # noqa: BLE001
+            return fn
+
+        try:
+            hints = typing.get_type_hints(fn)
[…]

[…]
