### 2026-04-08 1600183
build: remove py-markdown-table
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
