### 2026-04-08 0cbe251
refactor(operator): Renaming operator_module parameter operator_reference
