### 2026-04-08 579d6b6
refactor(operator): Remove support for OperatorModule path.
