### 2026-04-08 f739daa
refactor(operator): Rename param to operator_reference
