### 2026-04-08 2f7a255
refactor(cli): simplify ado show summary space (#803)
* refactor(cli): simplify ado show summary space

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* build: remove py-markdown-table

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

---------

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
