+++ orchestrator/cli/resources/discovery_space/show_summary.py
@@ -1,6 +1,5 @@
 # Copyright IBM Corporation 2025, 2026
 # SPDX-License-Identifier: MIT
-
 import datetime
 import pathlib
 
@@ -58,21 +57,26 @@ def show_discovery_space_summary(parameters: AdoShowSummaryCommandParameters) ->
 
         # Time-consuming imports
         status.update("Preparing to create your summary")
-        from py_markdown_table.markdown_table import markdown_table
 
         from orchestrator.cli.models.space import SpaceSummary
 
-        data = []
-        result = ""
-        if len(parameters.resource_ids) == 1:
-            autogenerated_name = pathlib.Path(
-                f"space_summary_{parameters.resource_ids[0]}.csv"
+        summaries: list[SpaceSummary] = []
+        for space_id in space_resources:
+            status.update(
+                f"Preparing summary for space {magenta(space_id)} ({len(summaries) + 1}/{len(space_resources)})"
             )
-        else:
-            autogenerated_name = pathlib.Path(
-                f"space_summary_{datetime.datetime.now(datetime.timezone.utc).timestamp()}.csv"
+            summaries.append(
+                SpaceSummary(space_id, parameters.ado_configuration.project_context)
             )
 
+        if not summaries:
+            console_print(
+                f"{ERROR}No data was retrieved for any of the spaces: {parameters.resource_ids}",
+                stderr=True,
+            )
+            raise typer.Exit(1)
+
+        result = ""
         if parameters.output_format == AdoShowSummarySupportedOutputFormats.MARKDOWN:
 
             if parameters.include_properties:
@@ -82,64 +86,47 @@ def show_discovery_space_summary(parameters: AdoShowSummaryCommandParameters) ->
                     stderr=True,
                 )
 
-            for space_id in space_resources:
-                status.update(
-                    f"Preparing summary for space {magenta(space_id)} ({len(data) + 1}/{len(space_resources)})"
-                )
-                data.append(
-                    SpaceSummary(
-                        space_id, parameters.ado_configuration.project_context
-                    ).to_markdown_text()
-                )
+            result = "\n".join([summary.to_markdown_text() for summary in summaries])
 
-            result = "\n".join(data)
-
-            if parameters.render_output:
-                from rich.markdown import Markdown
-
-                result = Markdown(result)
         else:
 
-            for space_id in space_resources:
-                status.update(
-                    f"Preparing summary for space {magenta(space_id)} ({len(data) + 1}/{len(space_resources)})"
-                )
-                data.append(
-                    SpaceSummary(
-                        space_id, parameters.ado_configuration.project_context
-                    ).to_dataframe(
+            df = pd.concat(
+                [
+                    summary.to_dataframe(
                         include_properties=parameters.include_properties,
                         columns_to_hide=parameters.columns_to_hide,
                     )
-                )
+                    for summary in summaries
+                ]
+            ).fillna("")
 
-            if not data:
-                console_print(
-                    f"{ERROR}No data was retrieved for any of the spaces: {parameters.resource_ids}",
-                    stderr=True,
-                )
-                raise typer.Exit(1)
+            if parameters.output_format == AdoShowSummarySupportedOutputFormats.CSV:
+                result = df.to_csv()
+            elif parameters.output_format == AdoShowSummarySupportedOutputFormats.TABLE:
+                result = df.to_markdown()
 
-            df = pd.concat(data).fillna("")
+        # Legacy code before using output_file in the parameters
+        if len(parameters.resource_ids) == 1:
+            autogenerated_name = pathlib.Path(
+                f"space_summary_{parameters.resource_ids[0]}.csv"
+            )
+        else:
+            autogenerated_name = pathlib.Path(
[…]
