### 2026-04-08 43a469a
refactor(cli): simplify ado show summary space (#803)
* refactor(cli): simplify ado show summary space

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* build: remove py-markdown-table

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

---------

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
