+++ pyproject.toml
@@ -25,7 +25,6 @@ dependencies = [
   "jsonpath-ng>=1.7.0",
   "numpy>=1.26.4",
   "pandas>=2.2.0",
-  "py-markdown-table>=1.3.0",
   "pydantic>=2.10",
   "pymysql[rsa]>=1.1.1",
   "pyyaml>=6.0.2",
