+++ orchestrator/cli/utils/output/dataframes.py
@@ -48,7 +48,7 @@ def df_to_output(
                 show_edge=True,
                 show_index=True,
                 box=rich.box.SQUARE,
-                do_not_truncate_column_content=no_trunc,
+                do_not_truncate_columns=no_trunc,
             )
         )
         if (
