### 2026-04-08 23af9d6
feat(cli): do not truncate id columns by default in ado get commands (#800)
* feat(cli): do not truncate IDENTIFIER column by default

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* feat(cli): increase table width if terminal is too small

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* feat(cli): prevent truncation of operator column in ado get operators

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* feat(cli): prevent truncation of actuator id column in ado get actuators

Signed-off-by: Alessandro Pomponio <alessandro.pomponio
[…]
