+++ orchestrator/utilities/rich.py
@@ -40,7 +40,7 @@ def dataframe_to_rich_table(
     show_index: bool = False,
     overflow: "OverflowMethod" = "ellipsis",
     no_wrap: bool = False,
-    do_not_truncate_column_content: bool = False,
+    do_not_truncate_columns: bool | list[str] = False,
 ) -> Table:
     """Convert a pandas DataFrame to a rich Table.
 
@@ -56,13 +56,16 @@ def dataframe_to_rich_table(
             DataFrame's index name if available, or "INDEX" as a default.
             Default is False for backward compatibility.
         overflow: How to handle content that exceeds column width. Options include
-            "ellipsis" (default), "ignore", "fold", "crop". When do_not_truncate_column_content
-            is True, this is automatically set to "ignore".
-        no_wrap: Whether to disable text wrapping in cells. When do_not_truncate_column_content
-            is True, this is automatically set to True.
-        do_not_truncate_column_content: If True, automatically calculates column widths
-            to fit all content without truncation. This sets overflow="ignore" and no_wrap=True,
-            and calculates the minimum table width needed to display all content.
+            "ellipsis" (default), "ignore", "fold", "crop". When do_not_truncate_columns
+            is active, this is automatically set to "ignore" for affected columns.
+        no_wrap: Whether to disable text wrapping in cells. When do_not_truncate_columns
+            is active, this is automatically set to True for affected columns.
+        do_not_truncate_columns: Controls which columns should not be truncated:
+            - False (default): All columns use default truncation behavior
+            - True: No columns are truncated (all widths calculated, table width set)
+            - list[str]: Only specified column names are not truncated (table width NOT set)
+            When active, automatically sets overflow="ignore" and no_wrap=True
+            for affected columns.
 
     Returns:
         A rich Table object ready for rendering
@@ -72,33 +75,72 @@ def dataframe_to_rich_table(
     # Initialize variables for column width control
     table_width = None
     column_width = {}
+    console_width = Console().width
 
-    if do_not_truncate_column_content:
-        # Override overflow and no_wrap settings to prevent truncation
-        overflow = "ignore"
-        no_wrap = True
+    # Determine which columns need width calculation based on input type
+    columns_with_no_truncation = []
+    disable_truncation_for_all_columns = (
+        False  # Track if truncation is disabled for ALL columns
+    )
 
-        # Calculate column name lengths
-        column_names_length = {col: len(col) for col in df.columns}
-        column_names_length[index_name] = len(index_name)
+    if do_not_truncate_columns is True:
+        # Disable truncation for ALL columns - we'll set table width
+        disable_truncation_for_all_columns = True
+        columns_with_no_truncation = list(df.columns)
+        if show_index:
+            columns_with_no_truncation.insert(0, index_name)
 
-        # Calculate longest string in each column
-        longest_string_in_column = df.apply(
-            lambda col: col.astype(str).str.len().max()
-        ).to_dict()
-        longest_string_in_column[index_name] = len(str(df.index.max()))
+    elif isinstance(do_not_truncate_columns, list):
+        # Disable truncation for ONLY specified columns - do NOT set table width
+        disable_truncation_for_all_columns = False
+        columns_with_no_truncation = [
+            col for col in do_not_truncate_columns if col in df.columns
+        ]
+        if show_index and index_name in do_not_truncate_columns:
+            columns_with_no_truncation.insert(0, index_name)
 
-        # Determine column widths (max of column name length and longest content)
-        column_width = {
-            col: max(column_names_length[col], longest_string_in_column[col])
-            for col in column_names_length
-        }
+    # Calculate widths for columns that should not be truncated
+    if columns_with_no_truncation:
+        for col in columns_with_no_truncation:
+            if col == index_name:
+                # Handle index column
[…]
