### 2026-04-08 f59fbe4
refactor(operator): Use operation_metadata only
Provide fallback implementation for old explore classes.
