+++ orchestrator/modules/operators/base.py
@@ -60,60 +60,170 @@ class OperatorFunction(Protocol):
 # Some operations are not RayActors: They don't have to use Actuators and StateUpdateQueue or Ray. They can use ray-workers
 
 
-class DiscoveryOperationBase(metaclass=abc.ABCMeta):
+class DiscoveryOperationBase:
 
-    @abc.abstractmethod
     def operationIdentifier(self) -> str:
-        """A unique id for the operation instance being run by the operator
+        """A unique id for the operation instance being run by the operator.
 
-        should have form $operatorIdentifier-$version-$uid"""
+        Returns ``"<classname>-<8-hex-uuid>"``, stable for the lifetime of
+        the instance.  Override when a deterministic or richer identifier is
+        needed (e.g. embedding the operator version or a caller-supplied run
+        id).
+        """
+        import uuid
+
+        attr = "_ado_operation_id"
+        if not hasattr(self, attr):
+            object.__setattr__(
+                self,
+                attr,
+                f"{type(self).__name__.lower()}-{uuid.uuid4().hex[:8]}",
+            )
+        return getattr(self, attr)
 
     @classmethod
-    @abc.abstractmethod
     def operatorIdentifier(cls) -> str:
-        """The identifier of this operator
+        """The identifier of this operator (``<name>-<version>``).
 
-        should have form method-version"""
+        Only required for operators loaded via the deprecated
+        ``OperatorModuleConf`` (``moduleName``/``moduleClass``) path.
+        Operators registered via ``@explore_operation`` or
+        ``operator_metadata()`` do not need to implement this.
+
+        Raises:
+            NotImplementedError: If not overridden.
+        """
+        raise NotImplementedError(
+            f"{cls.__name__}.operatorIdentifier() is not implemented. "
+            "New operators should implement operator_metadata() instead."
+        )
 
     @classmethod
-    @abc.abstractmethod
     def operationType(cls) -> DiscoveryOperationEnum:
-        """The type of operation this operator applies"""
+        """The type of operation this operator applies.
+
+        Only required for operators loaded via the deprecated
+        ``OperatorModuleConf`` path.  New operators declare the type inside
+        ``operator_metadata()``.
+
+        Raises:
+            NotImplementedError: If not overridden.
+        """
+        raise NotImplementedError(
+            f"{cls.__name__}.operationType() is not implemented. "
+            "New operators should implement operator_metadata() instead."
+        )
 
     @classmethod
-    @abc.abstractmethod
     def defaultOperationParameters(
         cls,
     ) -> pydantic.BaseModel:
-        """A default pytdantic model for this operations parameters with this operator"""
+        """A default pydantic model for this operator's parameters.
+
+        Not called by the framework.  New operators declare defaults via
+        ``OperatorMetadata.example_configuration`` inside
+        ``operator_metadata()``.
+
+        Raises:
+            NotImplementedError: If not overridden.
+        """
+        raise NotImplementedError(
+            f"{cls.__name__}.defaultOperationParameters() is not implemented. "
+            "New operators should implement operator_metadata() instead."
+        )
 
     @classmethod
-    @abc.abstractmethod
     def validateOperationParameters(
         cls,
         parameters: dict | pydantic.BaseModel,
     ) -> pydantic.BaseModel:
-        """If the parameters are valid returns a model for them.
+        """Validate and coerce ``parameters`` into the operator's model type.
 
-        Otherwise, will raise ValidationErrors"""
+        Used as a fallback by ``orchestrate_explore_operation`` when no
+        ``configuration_model`` is registered in ``OperatorMetadata``.  New
+        operators should declare ``configuration_model`` inside
+        ``operator_metadata()`` instead of overriding this method.
+
+        Raises:
[…]
