### 2026-04-08 ce200b1
refactor(operator): Use operation_metadata only
Use default operation_metadata in the decorator
