### 2026-04-08 1876eeb
refactor(cli)!: use --output flag everywhere and output to stdout
Adds --output-file for maintaining capabilities where output was to file by default

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
