+++ orchestrator/cli/models/parameters.py
@@ -90,6 +90,7 @@ class AdoShowEntitiesCommandParameters(pydantic.BaseModel):
     entities_property_format: AdoShowEntitiesSupportedPropertyFormats
     entities_type: AdoShowEntitiesSupportedEntityTypes
     no_trunc: bool
+    output_file: Path | None
     properties: list[str] | None
     resource_configuration: Path | None
     resource_id: str | None
@@ -104,6 +105,7 @@ class AdoShowRequestsCommandParameters(pydantic.BaseModel):
     ado_configuration: AdoConfiguration
     hide_fields: list[str] | None
     no_trunc: bool
+    output_file: Path | None
     output_format: AdoShowRequestsSupportedOutputFormats
     resource_id: str
 
@@ -112,6 +114,7 @@ class AdoShowResultsCommandParameters(pydantic.BaseModel):
     ado_configuration: AdoConfiguration
     hide_fields: list[str] | None
     no_trunc: bool
+    output_file: Path | None
     output_format: AdoShowResultsSupportedOutputFormats
     resource_id: str
 
@@ -120,6 +123,7 @@ class AdoShowSummaryCommandParameters(pydantic.BaseModel):
     ado_configuration: AdoConfiguration
     columns_to_hide: list[str] | None
     include_properties: list[str] | None
+    output_file: Path | None
     output_format: AdoShowSummarySupportedOutputFormats
     query: list[dict[str, str | None]] | None
     render_output: bool
