### 2026-04-08 9c526d3
refactor(operator): simplification
