+++ orchestrator/modules/operators/collections.py
@@ -20,6 +20,7 @@ from orchestrator.core.operation.config import (
 )
 from orchestrator.modules.operators.base import (
     DiscoveryOperationBase,
+    DiscoverySpaceSubscribingDiscoveryOperation,
     OperationOutput,
     OperatorFunction,
 )
@@ -29,27 +30,6 @@ from orchestrator.modules.operators.orchestrate import (
 )
 
 moduleLog = logging.getLogger("operation_collections")
-
-
-def _validate_operator_cls(cls: type, context: str) -> None:
-    """Assert that ``cls`` is a subclass of :class:`DiscoveryOperationBase`.
-
-    ``OperatorMetadata.cls`` is typed as ``type | None`` in the core layer to
-    keep ``config.py`` decoupled from ``base.py``.  This function is the single
-    place in ``modules/operators`` where the concrete constraint is enforced.
-
-    Args:
-        cls: The class to validate.
-        context: A short description of the call site, included in the error
-            message.
-
-    Raises:
-        TypeError: If ``cls`` is not a subclass of ``DiscoveryOperationBase``.
-    """
-    if not (isinstance(cls, type) and issubclass(cls, DiscoveryOperationBase)):
-        raise TypeError(
-            f"{context}: expected a DiscoveryOperationBase subclass, got {cls!r}."
-        )
 
 
 class OperatorCollection(pydantic.BaseModel):
@@ -301,11 +281,43 @@ def characterize_operation(
     return _register
 
 
+def _validate_explore_cls(t: type, metadata: OperatorMetadata) -> None:
+    """Validate a class-decorated explore operator and its metadata.
+
+    Args:
+        t: The decorated class.
+        metadata: The :class:`~orchestrator.core.operation.config.OperatorMetadata`
+            returned by ``t.operator_metadata()``.
+
+    Raises:
+        TypeError: If ``t`` is not a
+            :class:`~orchestrator.modules.operators.base.DiscoverySpaceSubscribingDiscoveryOperation`
+            subclass, if ``metadata.configuration_model`` is not set, or if
+            ``metadata.cls`` is set to a class other than ``t``.
+    """
+    if not issubclass(t, DiscoverySpaceSubscribingDiscoveryOperation):
+        raise TypeError(
+            f"@explore_operation: {t.__name__} must be a subclass of "
+            "DiscoverySpaceSubscribingDiscoveryOperation (i.e. inherit from "
+            "Search or Characterize)."
+        )
+    if metadata.configuration_model is None:
+        raise TypeError(
+            f"@explore_operation on {t.__name__}: operator_metadata() must set "
+            "configuration_model."
+        )
+    if metadata.cls is not None and metadata.cls is not t:
+        raise TypeError(
+            f"@explore_operation on {t.__name__}: operator_metadata().cls is "
+            f"{metadata.cls!r} but the decorated class is {t!r}. "
+            "Leave cls as None in operator_metadata() — the decorator sets it."
+        )
+
+
 def explore_operation(
     target: "type[DiscoveryOperationBase] | None" = None,
     *,
     name: str | None = None,
-    operator_class: "type[DiscoveryOperationBase] | None" = None,
     description: str | None = None,
     configuration_model: type[pydantic.BaseModel] | None = None,
     version: str | None = "v0.1",
@@ -315,13 +327,11 @@ def explore_operation(
 
     Supports two usage patterns:
 
-    **Class decoration** (no arguments) — all metadata is taken from the
-    class's :meth:`~orchestrator.modules.operators.base.DiscoveryOperationBase.operator_metadata`
-    classmethod.  The decorator generates the :data:`~orchestrator.modules.operators.base.OperatorFunction`
-    body, fills in ``function`` and ``cls`` on the returned
[…]
