### 2026-04-08 a31da83
refactor(operator): use class decorator
