### 2026-04-08 976288e
refactor(operator): use class decorator
