### 2026-04-08 4c2a086
docs: update skills and plugins documentation
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
