### 2026-04-08 15a1c88
refactor(operators): move validation code
to avoid circular imports
