+++ orchestrator/core/operation/config.py
@@ -264,99 +264,6 @@ class OperatorMetadata(pydantic.BaseModel):
         """Canonical identifier for this operator: ``{name}-{version}``."""
         return f"{self.name}-{self.version}"
 
-    @pydantic.field_validator("function", mode="after")
-    @classmethod
-    def _validate_function_signature(
-        cls, fn: typing.Callable | None
-    ) -> typing.Callable | None:
-        """Validate that ``function`` conforms to the ``OperatorFunction`` Protocol.
-
-        Skips validation when ``fn`` is ``None`` (returned by
-        ``operator_metadata()`` before the decorator injects the function).
-        Uses a lazy import of ``OperatorFunction`` to avoid a circular import
-        (``base.py`` imports ``config.py`` at module level).
-
-        Args:
-            fn: The callable to validate, or ``None``.
-
-        Returns:
-            The callable unchanged if validation passes, or ``None``.
-
-        Raises:
-            ValueError: If any present annotation does not match the
-                corresponding annotation in the Protocol.
-        """
-        import inspect
-
-        if fn is None:
-            return fn
-
-        # Lazy import to avoid circular dependency: base.py imports config.py
-        try:
-            from orchestrator.modules.operators.base import OperatorFunction
-
-            proto_hints = typing.get_type_hints(OperatorFunction.__call__)
-            proto_sig = inspect.signature(OperatorFunction.__call__)
-        except Exception:  # noqa: BLE001
-            return fn
-
-        try:
-            hints = typing.get_type_hints(fn)
-            sig = inspect.signature(fn)
-        except Exception:  # noqa: BLE001
-            return fn
-
-        # --- return type ---
-        expected_return = proto_hints.get("return")
-        actual_return = hints.get("return")
-        if (
-            expected_return is not None
-            and actual_return is not None
-            and actual_return != expected_return
-        ):
-            raise ValueError(
-                f"Operator function return type must be {expected_return!r}, "
-                f"got {actual_return!r}."
-            )
-
-        # --- positional parameters (matched by position, not name) ---
-        proto_positional = [
-            p
-            for p in proto_sig.parameters.values()
-            if p.name != "self"
-            and p.kind
-            in (
-                inspect.Parameter.POSITIONAL_OR_KEYWORD,
-                inspect.Parameter.POSITIONAL_ONLY,
-            )
-        ]
-        actual_positional = [
-            p
-            for p in sig.parameters.values()
-            if p.kind
-            in (
-                inspect.Parameter.POSITIONAL_OR_KEYWORD,
-                inspect.Parameter.POSITIONAL_ONLY,
-            )
-        ]
-        for idx, (proto_param, actual_param) in enumerate(
-            zip(proto_positional, actual_positional, strict=False), start=1
-        ):
-            expected_type = proto_hints.get(proto_param.name)
-            actual_type = hints.get(actual_param.name)
-            if (
-                expected_type is not None
-                and actual_type is not None
-                and actual_type != expected_type
-            ):
-                raise ValueError(
-                    f"Operator function parameter {actual_param.name!r} "
-                    f"(position {idx}) must be {expected_type!r}, "
-                    f"got {actual_type!r}."
-                )
-
-        return fn
-
 
 class OperatorReference(pydantic.BaseModel):
     """Identifies a registered operator by name and operation type.
