+++ orchestrator/modules/operators/collections.py
@@ -4,7 +4,6 @@
 import functools
 import logging
 import typing
-import warnings
 from typing import Annotated
 
 import pydantic
@@ -23,6 +22,7 @@ from orchestrator.modules.operators.base import (
     DiscoverySpaceSubscribingDiscoveryOperation,
     OperationOutput,
     OperatorFunction,
+    validate_operator_function_signature,
 )
 from orchestrator.modules.operators.orchestrate import (
     orchestrate_explore_operation,
@@ -57,131 +57,15 @@ class OperatorCollection(pydantic.BaseModel):
         """Returns all registered operator names."""
         return list(self.operators.keys())
 
-    def list_operations(self) -> list[str]:
-        """Returns all registered operator names.
-
-        Deprecated:
-            Use :meth:`list_operators` instead.
-        """
-        warnings.warn(
-            "list_operations() is deprecated; use list_operators() instead.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        return self.list_operators()
-
-    def configuration_model_for_operation(
-        self, name: str
-    ) -> type[pydantic.BaseModel] | None:
-        """Returns the configuration model for the named operator.
-
-        Deprecated:
-            Access ``collection.operators[name].configuration_model`` directly.
-
-        Args:
-            name: The registered operator name.
-
-        Returns:
-            The pydantic model class used to validate parameters, or None.
-
-        Raises:
-            ValueError: If the operator name is not registered.
-        """
-        warnings.warn(
-            "configuration_model_for_operation() is deprecated; "
-            "access collection.operators[name].configuration_model directly.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        if name not in self.operators:
-            raise ValueError(f"Unknown operator {name}")
-        return self.operators[name].configuration_model
-
-    def default_configuration_model_for_operation(
-        self, name: str
-    ) -> pydantic.BaseModel | None:
-        """Returns the example configuration instance for the named operator.
-
-        Deprecated:
-            Access ``collection.operators[name].example_configuration`` directly.
-
-        Args:
-            name: The registered operator name.
-
-        Returns:
-            The default pydantic model instance, or None.
-
-        Raises:
-            ValueError: If the operator name is not registered.
-        """
-        warnings.warn(
-            "default_configuration_model_for_operation() is deprecated; "
-            "access collection.operators[name].example_configuration directly.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        if name not in self.operators:
-            raise ValueError(f"Unknown operator {name}")
-        return self.operators[name].example_configuration
-
-    def description_for_operation(self, name: str) -> str | None:
-        """Returns the description for the named operator.
-
-        Deprecated:
-            Access ``collection.operators[name].description`` directly.
-
-        Args:
-            name: The registered operator name.
-
-        Returns:
-            The description string, or None.
-
-        Raises:
-            ValueError: If the operator name is not registered.
-        """
-        warnings.warn(
-            "description_for_operation() is deprecated; "
-            "access collection.operators[name].description directly.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        if name not in self.operators:
-            raise ValueError(f"Unknown operator {name}")
-        return self.operators[name].description
-
-    def operator_class_for_operation(self, name: str) -> type[DiscoveryOperationBase]:
-        """Returns the actor class registered for the named explore operator.
-
-        Deprecated:
-            Access ``collection.operators[name].cls`` directly.
-
-        Args:
-            name: The registered operator name.
-
-        Returns:
-            The DiscoveryOperationBase subclass for this operator.
-
-        Raises:
[…]

[…]
