### 2026-04-08 96cc132
refactor(operators): simplification
