### 2026-04-08 ed25145
fix(sql): Literal is too restrictive
