### 2026-04-09 e7f38a6
refactor(operators): Remove unused classes
We only need the class based operators for explore type operators.
