### 2026-04-09 60653b9
fix(vllm-perf): Addressed second review comments
Signed-off-by: Christian Pinto <christian.pinto@ibm.com>
