+++ plugins/actuators/vllm_performance/ado_actuators/vllm_performance/experiment_executor.py
@@ -430,7 +430,6 @@ def run_resource_and_workload_experiment(
                 result = execute_geospatial_benchmark(
                     base_url=benchmark_parameters.endpoint,
                     model=benchmark_parameters.model,
-                    interpreter=actuator_parameters.interpreter,
                     num_prompts=benchmark_parameters.num_prompts,
                     request_rate=benchmark_parameters.request_rate,
                     max_concurrency=benchmark_parameters.max_concurrency,
@@ -478,7 +477,6 @@ def run_resource_and_workload_experiment(
                 result = execute_random_benchmark(
                     base_url=benchmark_parameters.endpoint,
                     model=benchmark_parameters.model,
-                    interpreter=actuator_parameters.interpreter,
                     num_prompts=benchmark_parameters.num_prompts,
                     request_rate=benchmark_parameters.request_rate,
                     max_concurrency=benchmark_parameters.max_concurrency,
@@ -598,7 +596,6 @@ def run_workload_experiment(
                 result = execute_geospatial_benchmark(
                     base_url=benchmark_parameters.endpoint,
                     model=benchmark_parameters.model,
-                    interpreter=actuator_parameters.interpreter,
                     num_prompts=benchmark_parameters.num_prompts,
                     request_rate=benchmark_parameters.request_rate,
                     max_concurrency=benchmark_parameters.max_concurrency,
@@ -646,7 +643,6 @@ def run_workload_experiment(
                 result = execute_random_benchmark(
                     base_url=benchmark_parameters.endpoint,
                     model=benchmark_parameters.model,
-                    interpreter=actuator_parameters.interpreter,
                     num_prompts=benchmark_parameters.num_prompts,
                     request_rate=benchmark_parameters.request_rate,
                     max_concurrency=benchmark_parameters.max_concurrency,
