+++ website/docs/examples/vllm-performance-geospatial.md
@@ -1,0 +1,282 @@
+# Benchmarking Geospatial Models with vLLM
+
+<!-- markdownlint-disable no-blanks-blockquote -->
+
+> [!NOTE] The scenario
+>
+> **In this example, the
+> [_vllm_performance_ actuator](../actuators/vllm_performance.md) is used to
+> benchmark geospatial models (IBM-NASA Prithvi) for Earth observation tasks.**
+>
+> Geospatial models process satellite imagery for tasks like flood detection,
+> land use classification, and environmental monitoring. Unlike text-based LLMs,
+> these models:
+>
+> - Accept base64-encoded satellite images as input
+> - Output classification results rather than text tokens
+> - Have different performance characteristics and optimization requirements
+>
+> In this example:
+>
+> - We will define a space of geospatial model deployment configurations to test
+> - Use the `test-geospatial-deployment-v1` experiment to create and benchmark
+>   vLLM deployments serving Prithvi models
+> - Explore how deployment parameters affect inference latency for flood
+>   detection tasks
+
+> [!IMPORTANT] Prerequisites
+>
+> - Be logged-in to your Kubernetes/OpenShift cluster
+> - Have access to a namespace where you can create vLLM deployments
+> - Install the following Python packages locally:
+>
+> ```bash
+> pip install ado-vllm-performance
+> ```
+
+> [!TIP] TL;DR
+>
+> Create the following files and execute:
+>
+> ```bash
+> # Create resources and run operation
+> ado create op -f geospatial_operation.yaml \
+>    --with space=geospatial_space.yaml --with ac=vllm_actuator_configuration.yaml
+> ```
+>
+> See
+> [configuring the `vllm_performance` actuator](../actuators/vllm_performance.md#configuring-the-vllm_performance-actuator)
+> for configuration options.
+
+<!-- markdownlint-enable no-blanks-blockquote -->
+
+## Verify the installation
+
+Verify the installation with:
+
+```commandline
+ado get actuators --details
+```
+
+The actuator `vllm_performance` should appear in the list. To see the geospatial
+experiments:
+
+```commandline
+ado get experiments --details
+```
+
+You should see experiments including `test-geospatial-deployment-v1`,
+`test-geospatial-endpoint-v1`, `test-geospatial-deployment-custom-dataset-v1`,
+and `test-geospatial-endpoint-custom-dataset-v1`.
+
+## Create an actuator configuration
+
+The vllm-performance actuator needs information about the target cluster. This
+is provided via an `actuatorconfiguration`.
+
+First execute:
+
+```commandline
+ado template actuatorconfiguration --actuator-identifier vllm_performance -o vllm_actuator_configuration.yaml
+```
+
+Edit the file and set correct values for at least the `namespace` field. In this
+example we are assuming the namespace the user has access to is named
+`vllm-testing`.
+
+```yaml
+# you MUST set this to a namespace where you can create vLLM deployments
+namespace: vllm-testing
+# Required to access Prithvi models
+hf_token: <your HuggingFace access token>
+```
+
+Then save this configuration:
+
+```bash
+ado create actuatorconfiguration -f vllm_actuator_configuration.yaml
+```
+
+## Define the geospatial configurations to test
+
+For geospatial models, we focus on deployment parameters that affect inference
+latency since these models output classification results rather than generating
+tokens. Key parameters include:
+
+- **GPU configuration**: Type and number of GPUs
+- **Memory allocation**: CPU and GPU memory
+- **Batch processing**: `max_num_seq` for concurrent requests
+- **Workload pattern**: Request rate and concurrency
[…]
