### 2026-04-09 73c9d31
feat(vllm_performance): promote geospatial experiments to stable and add endpoint experiment with custom dataset (#744)
* feat(vllm_performance): Added performance_testing-geospatial-endpoint-custom-dataset experiment

Signed-off-by: Christian Pinto <christian.pinto@ibm.com>

* fix(vllm_performance): set correct experiment identifiers and property configuration to test existing vllm geospatial endpoint

Signed-off-by: Michele Gazzetti <michele.gazzetti1@ibm.com>

* fix(vllm_performance): update default values to benchmark geospatial models

Signed-off-by: Michele Gazzetti <michele.gazzetti1@ib
[…]
