+++ examples/example_yamls/geospatial_flood_detection_space.yaml
@@ -1,0 +1,36 @@
+# Copyright IBM Corporation 2025, 2026
+# SPDX-License-Identifier: MIT
+metadata:
+  name: geospatial-flood-detection-space
+  description: "Explore Prithvi geospatial model deployment configurations for flood detection"
+entitySpace:
+  - identifier: model
+    propertyDomain:
+      values:
+        - "ibm-nasa-geospatial/Prithvi-EO-2.0-300M-TL-Sen1Floods11"
+  - identifier: n_gpus
+    propertyDomain:
+      values: [1]
+  - identifier: gpu_type
+    propertyDomain:
+      values:
+        - "NVIDIA-A100-80GB-PCIe"
+  - identifier: memory
+    propertyDomain:
+      values:
+        - "64Gi"
+        - "128Gi"
+  - identifier: max_num_seq
+    propertyDomain:
+      values: [32, 64, 128]
+  - identifier: request_rate
+    propertyDomain:
+      values: [10, 50, 100]
+  - identifier: dataset
+    propertyDomain:
+      values:
+        - "india_url_in_b64_out"
+        - "valencia_url_in_b64_out"
+measurementSpace:
+  - actuatorIdentifier: vllm_performance
+    experimentIdentifier: test-geospatial-deployment-v1
