+++ website/docs/getting-started/ado.md
@@ -654,7 +654,8 @@ The complete syntax of the `ado show entities` command is as follows:
 ```shell
 ado show entities RESOURCE_TYPE [RESOURCE_ID] [--use-latest] [--file | -f <file.yaml>]\
                   [--property-format {observed | target}] \
-                  [--output-format {console | csv | json}] \
+                  [--output | -o {console | csv | json}] \
+                  [--output-file <path>] \
                   [--property <property-name>] \
                   [--include {sampled | matching | missing | unsampled}] \
                   [--aggregate {mean | median | variance | std | min | max}]
@@ -691,15 +692,20 @@ Where:
 
 <!-- prettier-ignore-end -->
 
-- `--output-format` is the format in which to display the entity data. One of:
+- `--output` (or `-o`) is the format in which to display the entity data. One
+  of:
 
 <!-- prettier-ignore-start -->
 
     - `console` (print to stdout)
-    - `csv` (output as CSV file)
-    - `json` (output as JSON file)
+    - `csv` (write CSV to stdout, or to file if `--output-file` is specified)
+    - `json` (write JSON to stdout, or to file if `--output-file` is specified)
 
 <!-- prettier-ignore-end -->
+
+- `--output-file` specifies a file path to write the output to. If not provided,
+  output is written to stdout (except for console format which always prints to
+  stdout).
 
 - `--property` (can be specified multiple times) is used to filter what measured
   properties need to be output.
@@ -740,7 +746,15 @@ Where:
 ```shell
  ado show entities space space-abc123-456def --include matching \
                                              --property-format target \
-                                             --output-format csv
+                                             -o csv --output-file entities.csv
+```
+
+Or to write CSV to stdout for piping:
+
+```shell
+ ado show entities space space-abc123-456def --include matching \
+                                             --property-format target \
+                                             -o csv > entities.csv
 ```
 
 <!-- markdownlint-disable line-length -->
@@ -750,7 +764,7 @@ Where:
 <!-- markdownlint-enable line-length -->
 
 ```shell
-ado show entities operation randomwalk-0.5.0-123abc --output-format json \
+ado show entities operation randomwalk-0.5.0-123abc -o json \
                                                     --property my-property-1 \
                                                     --property my-property-2
 ```
@@ -766,7 +780,8 @@ The complete syntax of the `ado show requests` command is as follows:
 
 ```shell
 ado show requests operation [RESOURCE_ID] [--use-latest] \
-                            [--output-format | -o <console | csv | json>] \
+                            [--output | -o <console | csv | json>] \
+                            [--output-file <path>] \
                             [--hide <field>]
 ```
 
@@ -775,8 +790,10 @@ ado show requests operation [RESOURCE_ID] [--use-latest] \
 - `--use-latest` will use the identifier of the latest (i.e. most recent)
   operation created locally. It is not context aware. It is ignored if a
   RESOURCE_ID is provided.
-- `--output-format` determines whether the output will be printed to console or
-  saved to a file.
+- `--output` (or `-o`) determines the output format. Output is written to stdout
+  by default, or to a file if `--output-file` is specified.
+- `--output-file` specifies a file path to write the output to. If not provided,
+  output is written to stdout.
 - `--hide` can be specified multiple times and allows hiding fields from the
   output.
 
@@ -785,7 +802,13 @@ ado show requests operation [RESOURCE_ID] [--use-latest] \
 ###### Show measurement requests for an operation and save them as csv
 
 ```shell
-ado show requests operation randomwalk-0.5.0-123abc -o csv
[…]
