### 2026-04-09 04a67a0
refactor(cli)!: use --output flag everywhere and output to stdout (#804)
* refactor(cli)!: use --output flag everywhere and output to stdout

Adds --output-file for maintaining capabilities where output was to file by default

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* docs(website): update documentation for -o/--output/--output-file

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* docs: update skills and plugins documentation

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* docs(website): update show entities call

Signed-off
[…]
