+++ orchestrator/cli/resources/discovery_space/show_summary.py
@@ -1,7 +1,5 @@
 # Copyright IBM Corporation 2025, 2026
 # SPDX-License-Identifier: MIT
-import datetime
-import pathlib
 
 import typer
 from rich.status import Status
@@ -105,25 +103,14 @@ def show_discovery_space_summary(parameters: AdoShowSummaryCommandParameters) ->
             elif parameters.output_format == AdoShowSummarySupportedOutputFormats.TABLE:
                 result = df.to_markdown()
 
-        # Legacy code before using output_file in the parameters
-        if len(parameters.resource_ids) == 1:
-            autogenerated_name = pathlib.Path(
-                f"space_summary_{parameters.resource_ids[0]}.csv"
-            )
-        else:
-            autogenerated_name = pathlib.Path(
-                f"space_summary_{datetime.datetime.now(datetime.timezone.utc).timestamp()}.csv"
-            )
-        if (
-            autogenerated_name
-            and parameters.output_format == AdoShowSummarySupportedOutputFormats.CSV
-        ):
-            autogenerated_name.write_text(result)
+        if parameters.output_file:
+            parameters.output_file.write_text(result)
             console_print(
-                f"{SUCCESS} Output saved as {magenta(str(autogenerated_name))}",
+                f"{SUCCESS} Output saved as {magenta(str(parameters.output_file))}",
                 stderr=True,
             )
             return
+
         if parameters.render_output:
             from rich.markdown import Markdown
 
