### 2026-04-09 dcac75a
test(sql): update version string in test
