### 2026-04-09 03ad248
feat(core): sort constitutive properties before generating entity id
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
