### 2026-04-09 c741efa
fix(tests): version string
