### 2026-04-09 e99e4c9
feat(core): sort constitutive properties before generating entity id (#815)
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
