### 2026-04-09 17d4b5c
chore(typing): update type
