### 2026-04-09 0d6f315
test(operator): test function sig validation
