### 2026-04-10 90d98de
refactor(operator): remove deprecated methods
From DiscoveryOperationBase
