+++ tests/operators/test_operators.py
@@ -38,27 +38,6 @@ from orchestrator.modules.operators.randomwalk import (
 )
 
 
-def test_randomwalk_class_methods() -> None:
-
-    import orchestrator.metastore.project
-
-    assert (
-        RandomWalk.operationType()
-        == orchestrator.core.operation.config.DiscoveryOperationEnum.SEARCH
-    )
-    assert RandomWalk.operatorIdentifier().split("-")[0] == "randomwalk"
-
-
-def test_raytune_class_methods() -> None:
-    import orchestrator.metastore.project
-
-    assert (
-        RayTune.operationType()
-        == orchestrator.core.operation.config.DiscoveryOperationEnum.SEARCH
-    )
-    assert RayTune.operatorIdentifier().split("-")[0] == "raytune"
-
-
 def test_operator_function_conf() -> None:
 
     function = orchestrator.core.operation.config.OperatorReference(
@@ -80,30 +59,15 @@ def test_operator_module_conf(
     operator_module_conf: orchestrator.core.operation.config.OperatorModuleConf,
 ) -> None:
 
+    from orchestrator.modules.module import load_module_class_or_function
+
     assert (
         operator_module_conf.operationType
         == orchestrator.core.operation.config.DiscoveryOperationEnum.SEARCH
     )
-    assert (
-        operator_module_conf.operatorIdentifier.split("-")[0]
-        == operator_module_conf.moduleClass.lower()
-    )
-
-
-def test_operator_module_conf_random_walk() -> None:
-
-    module = orchestrator.core.operation.config.OperatorModuleConf(
-        moduleName="orchestrator.modules.operators.randomwalk",
-        moduleClass="RandomWalk",
-    )
-
-    assert module.operatorIdentifier
-    assert isinstance(module.operatorIdentifier, str)
-    assert (
-        module.operationType
-        == orchestrator.core.operation.config.DiscoveryOperationEnum.SEARCH
-    )
-    assert module.operatorIdentifier.split("-")[0] == "randomwalk"
+    cls = load_module_class_or_function(operator_module_conf)
+    expected_name = cls.operator_metadata().name
+    assert operator_module_conf.operatorIdentifier.startswith(f"{expected_name}-")
 
 
 def test_characterize(expected_characterize_operators: list[str]) -> None:
@@ -171,11 +135,7 @@ def test_explore_operator_class_registration(
     for name in expected_explore_operators:
         operator = orchestrator.modules.operators.collections.explore.operators[name]
         cls = operator.cls
-        # Ray-decorated classes become ActorClass objects so issubclass is not
-        # usable; verify the registered object is non-None and exposes the
-        # operatorIdentifier classmethod expected of explore operators.
         assert cls is not None
-        assert hasattr(cls, "operatorIdentifier")
 
 
 def test_explore_operator_function_conf_identifier_matches_registered_name() -> None:
@@ -279,7 +239,7 @@ def test_random_walk_operation_configuration() -> None:
         orchestrator.modules.operators.collections.explore.operators[
             "random_walk"
         ].example_configuration
-        == RandomWalk.defaultOperationParameters()
+        == RandomWalk.operator_metadata().example_configuration
     )
 
 
@@ -308,7 +268,7 @@ def test_raytune_operation_configuration(
         orchestrator.modules.operators.collections.explore.operators[
             "ray_tune"
         ].example_configuration
-        == RayTune.defaultOperationParameters()
+        == RayTune.operator_metadata().example_configuration
     )
 
 
@@ -329,12 +289,10 @@ def test_random_walk_config(
     import pydantic
 
     assert randomWalkConf is not None
-    assert RandomWalk.validateOperationParameters(
[…]
