### 2026-04-10 41e9b4a
refactor(operator): explore decorator only applies to classes
