### 2026-04-10 2c13836
refactor(operator): make version, parameters model and example parameters required
