### 2026-04-10 70989b4
refactor(operator): update to use OperatorMetadata
simplify function calls by passing OperatorMetadata instance everywhere instead of OperatorReference + additional data
