### 2026-04-10 a28abfa
feat(cli): add support for -o in ado contexts
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
