### 2026-04-10 e38a5fc
fix(tests): update tests where behaviour changed
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
