### 2026-04-10 95425dc
build(hooks): add default stage for hooks (#822)
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
