+++ .pre-commit-config.yaml
@@ -1,6 +1,7 @@
 default_install_hook_types:
   - pre-commit
   - commit-msg
+default_stages: [pre-commit]
 repos:
   - repo: https://github.com/ibm/detect-secrets
     rev: 0.13.1+ibm.64.dss
