### 2026-04-10 ca58630
feat(cli): add support for -o in ado contexts (#821)
* feat(cli): add support for -o in ado contexts

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* fix(tests): update tests where behaviour changed

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

---------

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
