+++ orchestrator/cli/resources/context/get.py
@@ -9,10 +9,9 @@ import yaml
 from orchestrator.cli.models.parameters import AdoGetCommandParameters
 from orchestrator.cli.models.types import AdoGetSupportedOutputFormats
 from orchestrator.cli.utils.output.prints import (
-    ADO_INFO_EMPTY_DATAFRAME,
+    ADO_NO_CONTEXT_AVAILABLE_ERROR,
     ERROR,
     HINT,
-    WARN,
     console_print,
     context_not_in_available_contexts_error_str,
     cyan,
@@ -26,10 +25,9 @@ if typing.TYPE_CHECKING:
 
 def get_context(
     parameters: AdoGetCommandParameters,
-    simplify_output: bool = False,
 ) -> None:
-    import rich.box
 
+    # Never truncate the CONTEXT (name) column
     if not parameters.no_trunc:
         parameters.no_trunc = ["CONTEXT"]
 
@@ -39,11 +37,7 @@ def get_context(
     # The only possible way for this should be when the user is
     # providing a context with -c and the ado context dir is empty
     if len(available_contexts) == 0:
-        console_print(
-            f"{WARN}There are no contexts available.\n"
-            f"{HINT}You can create a context with {cyan('ado create context')}",
-            stderr=True,
-        )
+        console_print(ADO_NO_CONTEXT_AVAILABLE_ERROR, stderr=True)
         raise typer.Exit(1)
 
     if parameters.resource_id:
@@ -65,23 +59,18 @@ def get_context(
     parameters.exclude_default = False
 
     if parameters.output_format == AdoGetSupportedOutputFormats.NAME:
-        # NAME format: output only context identifiers, one per line
+        # NAME format: output only context names, one per line
         for context in sorted(available_contexts):
             console_print(context)
         return
 
     if parameters.output_format == AdoGetSupportedOutputFormats.DEFAULT:
-        if simplify_output:
-            _simple_contexts_formatting(contexts=available_contexts)
-            return
+        import rich.box
 
         contexts_df = _prepare_context_dataframe(
             contexts=available_contexts,
-            default_context=parameters.ado_configuration.active_context,
+            active_context=parameters.ado_configuration.active_context,
         )
-        if contexts_df.empty:
-            console_print(ADO_INFO_EMPTY_DATAFRAME, stderr=True)
-            return
 
         console_print(
             dataframe_to_rich_table(
@@ -98,19 +87,25 @@ def get_context(
         format_resource_for_ado_get_custom_format,
     )
 
+    to_print = []
     try:
-        to_print = [
-            ProjectContext.model_validate(
-                yaml.safe_load(
-                    parameters.ado_configuration.project_context_path_for_context(
-                        ctx
-                    ).read_text()
+        for ctx in available_contexts:
+            to_print.append(
+                ProjectContext.model_validate(
+                    yaml.safe_load(
+                        parameters.ado_configuration.project_context_path_for_context(
+                            ctx
+                        ).read_text()
+                    )
                 )
             )
-            for ctx in available_contexts
-        ]
     except pydantic.ValidationError as e:
-        console_print(f"{ERROR}One of the contexts was not valid:\n{e}", stderr=True)
+        console_print(
+            f"{ERROR}Context {cyan(ctx)} was not valid:\n\n{e}\n\n"
+            f"{HINT}You can manually update the context file at: '{parameters.ado_configuration.project_context_path_for_context(ctx)}'\n"
+            f"\tAlternatively, delete the context with {cyan(f'ado delete context {ctx}')}",
+            stderr=True,
+        )
         raise typer.Exit(1) from e
 
[…]
