### 2026-04-10 4c5ce62
refactor(cli): change default output type for ado get to table
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
