+++ orchestrator/cli/commands/get.py
@@ -119,7 +119,6 @@ def get_resource(
             "--output",
             "-o",
             rich_help_panel=OUTPUT_CONFIGURATION_OPTIONS,
-            show_default=False,
             help="Output information in a different format. The 'json', 'raw', and 'yaml' formats will output the entire resource. The 'name' format outputs only resource identifiers (similar to kubectl get -o name). Not all formats may be supported by all resources.",
         ),
     ] = AdoGetSupportedOutputFormats.DEFAULT.value,
