### 2026-04-10 06131ae
docs(cli): improve docstring for ado get
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
