### 2026-04-10 5378880
refactor(cli): rename DEFAULT enum value
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
