+++ orchestrator/cli/models/types.py
@@ -93,10 +93,10 @@ class AdoEditSupportedResourceTypes(Enum):
 #################### ado get ####################
 class AdoGetSupportedOutputFormats(Enum):
     CONFIG = _CONFIG
-    DEFAULT = _TABLE
     JSON = _JSON
     NAME = _NAME
     RAW = _RAW
+    TABLE = _TABLE
     YAML = _YAML
 
 
