+++ orchestrator/cli/commands/get.py
@@ -121,7 +121,7 @@ def get_resource(
             rich_help_panel=OUTPUT_CONFIGURATION_OPTIONS,
             help="Output information in a different format. The 'json', 'raw', and 'yaml' formats will output the entire resource. The 'name' format outputs only resource identifiers (similar to kubectl get -o name). Not all formats may be supported by all resources.",
         ),
-    ] = AdoGetSupportedOutputFormats.DEFAULT.value,
+    ] = AdoGetSupportedOutputFormats.TABLE.value,
     exclude_default: Annotated[
         bool,
         typer.Option(
@@ -166,7 +166,7 @@ def get_resource(
             Make an attempt to minimize the output produced.
             This might entail applying transformations on the model, changing it from the original.
 
-            Ignored when the output type is default or raw.
+            Ignored when the output type is table or raw.
             If set, implies --exclude-default --exclude-unset --exclude-none.
             """,
             rich_help_panel=OUTPUT_CONFIGURATION_OPTIONS,
@@ -236,7 +236,7 @@ def get_resource(
             help="""
             Provide a space configuration to match other spaces. Only for spaces.
 
-            If set, disregards --query and --label, and uses the default output format.
+            If set, disregards --query and --label, and uses the table output format.
             """,
             file_okay=True,
             dir_okay=False,
@@ -252,7 +252,7 @@ def get_resource(
             Provide a space id to match other spaces. Only for spaces.
             Takes precedence over --matching-space.
 
-            If set, disregards --query and --label, and uses the default output format.
+            If set, disregards --query and --label, and uses the table output format.
             """,
             show_default=False,
             rich_help_panel=DISCOVERY_SPACE_ONLY_OPTIONS,
@@ -304,13 +304,13 @@ def get_resource(
 
     if (
         matching_space or matching_space_id
-    ) and output_format != AdoGetSupportedOutputFormats.DEFAULT:
+    ) and output_format != AdoGetSupportedOutputFormats.TABLE:
         console_print(
             f"{WARN}--matching-space and --matching-space-id only support "
-            f"the {AdoGetSupportedOutputFormats.DEFAULT.value} output format.",
+            f"the {AdoGetSupportedOutputFormats.TABLE.value} output format.",
             stderr=True,
         )
-        output_format = AdoGetSupportedOutputFormats.DEFAULT
+        output_format = AdoGetSupportedOutputFormats.TABLE
 
     if exclude_fields and output_format not in {
         AdoGetSupportedOutputFormats.JSON,
