+++ orchestrator/cli/resources/data_container/get.py
@@ -16,7 +16,7 @@ def get_data_container(parameters: AdoGetCommandParameters) -> None:
         handle_ado_get_default_format,
     )
 
-    if parameters.output_format == AdoGetSupportedOutputFormats.DEFAULT:
+    if parameters.output_format == AdoGetSupportedOutputFormats.TABLE:
         handle_ado_get_default_format(
             parameters=parameters,
             resource_type=CoreResourceKinds.DATACONTAINER,
