### 2026-04-10 5461f59
refactor(cli): remove _DEFAULT type
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
