+++ orchestrator/cli/models/types.py
@@ -22,7 +22,6 @@ _SAMPLE_STORE_SINGULAR = "samplestore"
 _CONFIG = "config"
 _CONSOLE = "console"
 _CSV = "csv"
-_DEFAULT = "default"
 _JSON = "json"
 _MARKDOWN_SHORT = "md"
 _NAME = "name"
