### 2026-04-10 89a13d2
refactor(cli): rename console output in ado show entities
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
