### 2026-04-10 2eb3fe2
refactor(cli): rename console output in ado show requests
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
