### 2026-04-10 07493a3
refactor(cli): rename console output in ado show results
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
