### 2026-04-10 410ccf0
refactor(cli): remove _CONSOLE var
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
