### 2026-04-10 a64f8b3
fix(test): update assertions
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
