### 2026-04-10 222d084
feat(cli): add request alias to measurementrequest (#830)
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
