### 2026-04-10 0eb6893
feat(cli): add request alias to measurementrequest (#830)
Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>
