+++ orchestrator/cli/utils/resources/mappings.py
@@ -10,6 +10,7 @@ cli_shorthands_to_cli_names: dict[str, str] = {
     "dcr": CoreResourceKinds.DATACONTAINER.value,
     "exp": "experiment",
     "op": CoreResourceKinds.OPERATION.value,
+    "request": "measurementrequest",
     "space": CoreResourceKinds.DISCOVERYSPACE.value,
     "store": CoreResourceKinds.SAMPLESTORE.value,
 }
