### 2026-04-10 bea0a7a
chore(deps): update dependency google-api-core to v2.30.3
