### 2026-04-10 3834516
chore(deps): update opentelemetry-python monorepo to v1.41.0
