### 2026-04-13 ed92d0f
chore(style): remove unnecessary pydantic option
