### 2026-04-13 2d95c44
refactor(cli): use table consistently for default output type in ado commands (#829)
* refactor(cli): change default output type for ado get to table

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* docs(cli): improve docstring for ado get

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* refactor(cli): rename DEFAULT enum value

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* refactor(cli): remove _DEFAULT type

Signed-off-by: Alessandro Pomponio <alessandro.pomponio1@ibm.com>

* refactor(cli): rename console output in ado show ent
[…]
