+++ website/docs/getting-started/ado.md
@@ -393,7 +393,7 @@ The complete syntax of the `ado get` command is as follows:
 <!-- markdownlint-disable line-length -->
 
 ```shell
-ado get RESOURCE_TYPE [RESOURCE_ID] [--output | -o <default | yaml | json | config | raw>] \
+ado get RESOURCE_TYPE [RESOURCE_ID] [--output | -o <table | yaml | json | config | raw>] \
                                     [--exclude-default | --no-exclude-default] \
                                     [--exclude-unset | --no-exclude-unset ] \
                                     [--exclude-none | --no-exclude-none ] \
@@ -434,7 +434,7 @@ Where:
 
 <!-- prettier-ignore-start -->
 
-    - The `default` format shows the _identifier_, the _name_, and the _age_ of
+    - The `table` format shows the _identifier_, the _name_, and the _age_ of
       the matching resources.
     - The `name` format outputs only the resource identifiers, one per line
       (similar to `kubectl get -o name`).
@@ -464,11 +464,11 @@ Where:
 - `--minimize` minimizes the output. This might entail applying transformations
   on the model, changing it from the original. If set, it implies
   `--exclude-default`, `--exclude-unset`, and `--exclude-none`. This option is
-  ignored when the output type is `default` or `raw`.
+  ignored when the output type is `table` or `raw`.
 - The `--from-sample-store`, `--from-space`, `--from-operation` flags are
   available **only for `ado get requests`** and allow specifying what
   samplestore/space/operation the measurement request belongs to.
-- When using the `--details` flag with the `default` output format, additional
+- When using the `--details` flag with the `table` output format, additional
   columns with the _description_ and the _labels_ of the matching resources are
   printed.
 - The `--show-deprecated` flag is available **only for `ado get experiments`**
@@ -666,7 +666,7 @@ The complete syntax of the `ado show entities` command is as follows:
 ```shell
 ado show entities RESOURCE_TYPE [RESOURCE_ID] [--use-latest] [--file | -f <file.yaml>]\
                   [--property-format {observed | target}] \
-                  [--output | -o {console | csv | json}] \
+                  [--output | -o {table | csv | json}] \
                   [--output-file <path>] \
                   [--property <property-name>] \
                   [--include {sampled | matching | missing | unsampled}] \
@@ -709,15 +709,14 @@ Where:
 
 <!-- prettier-ignore-start -->
 
-    - `console` (print to stdout)
+    - `table` (print to stdout)
     - `csv` (write CSV to stdout, or to file if `--output-file` is specified)
     - `json` (write JSON to stdout, or to file if `--output-file` is specified)
 
 <!-- prettier-ignore-end -->
 
-- `--output-file` specifies a file path to write the output to. If not provided,
-  output is written to stdout (except for console format which always prints to
-  stdout).
+- `--output-file` specifies a file path to write the output to (except for table
+  format which always prints to stdout).
 
 - `--property` (can be specified multiple times) is used to filter what measured
   properties need to be output.
@@ -792,7 +791,7 @@ The complete syntax of the `ado show requests` command is as follows:
 
 ```shell
 ado show requests operation [RESOURCE_ID] [--use-latest] \
-                            [--output | -o <console | csv | json>] \
+                            [--output | -o <table | csv | json>] \
                             [--output-file <path>] \
                             [--hide <field>]
 ```
@@ -844,7 +843,7 @@ The complete syntax of the `ado show results` command is as follows:
 
 ```shell
 ado show results operation [RESOURCE_ID] [--use-latest] \
-                           [--output | -o <console | csv | json>] \
+                           [--output | -o <table | csv | json>] \
                            [--output-file <path>] \
                            [--hide <field>]
 ```
