### 2026-05-01 #10304 @matt-graham
Hi @Torigh. Do you think you will be able to complete your review by the end of next week (8th May)? If you don't think you will be able to that is not a problem as we already have two reviews which meets the JOSS requirement, but it would be helpful for us to confirm if this will be the case now so we can proceed accordingly. Thanks!
